generate_null_distribution_T3

This function generate the null distribution of T_eta, should be run before calculating TWT. Please see Setp 2 in our manuscript for more information.

generate_null_distribution_T3(m, n, cov_mat, cutoff_value)

Arguments

m

Dimension.

n

Number of the generated sample size.

cov_mat

Covariance matrix for generating random samples.

cutoff_value

Truncated value, should be in 0 to 1 and must contain 1.

Value

Null distribution matrix with number of rows equal n and number of columns equal length of cutoff value minus 1.

Examples

ThreeWayTest::generate_null_distribution_T3(m=6,n=1000,
cov_mat=diag(nrow = 6, ncol = 6), cutoff_value=c(0.2,0.4,0.6,0.8,1))
#>               [,1]       [,2]       [,3]       [,4]
#>    [1,]  1.1847407  1.3583924  1.4957900  1.6308019
#>    [2,]  6.7786989  8.2797016  9.7287442 10.4772053
#>    [3,]  3.3346802  3.9791371  4.5774513  4.6496873
#>    [4,] 10.7561178 10.9004102 10.9761836 10.9871585
#>    [5,]  1.5754893  1.7894014  1.9305090  1.9465371
#>    [6,]  1.9132954  2.1313065  2.2867775  2.3519649
#>    [7,]  3.7978856  5.3536035  6.2930424  7.2276993
#>    [8,]  4.1426451  4.7399309  4.9360348  5.0036687
#>    [9,]  3.0915091  3.4341429  3.6931692  3.7581431
#>   [10,]  9.0586064  9.8438690 10.3821273 10.7687091
#>   [11,]  1.8011062  2.0643805  2.3151544  2.5017040
#>   [12,]  9.3801632  9.6377526  9.8204027  9.9597606
#>   [13,]  3.1677519  3.9608036  4.1864523  4.3832197
#>   [14,] 10.1723205 11.9647659 12.0729339 12.1338175
#>   [15,]  5.4106573  6.3462196  6.7201217  6.7374400
#>   [16,]  2.2782189  2.3967109  2.4330192  2.4368886
#>   [17,]  1.9750012  2.3307726  2.6004962  2.7825681
#>   [18,]  5.7352170  6.4570940  6.8270447  7.0929778
#>   [19,]  2.2025420  2.5196259  2.7088335  2.7738253
#>   [20,]  3.9198113  4.4951531  4.7874189  4.7976112
#>   [21,]  0.4524328  0.5285548  0.5668653  0.5699394
#>   [22,]  9.1963772 10.1318911 10.2396436 10.3259177
#>   [23,]  2.9861699  3.7067318  4.1211688  4.4130500
#>   [24,]  1.7449190  2.3650380  2.6219048  2.7699647
#>   [25,]  1.2789124  1.7315478  2.0335861  2.1551227
#>   [26,]  2.2199232  2.7051932  2.9895914  3.0879673
#>   [27,]  9.3997672 10.7915686 10.8412438 10.8600129
#>   [28,]  4.9825133  5.2612558  5.4729673  5.6801078
#>   [29,]  1.6866246  1.9729721  2.2363443  2.2674013
#>   [30,]  5.2397333  6.0773910  6.1443666  6.1447188
#>   [31,]  2.7067873  3.3769799  3.6031860  3.8195410
#>   [32,]  1.6528335  2.0599990  2.4046591  2.5367328
#>   [33,]  6.3439048  7.2256653  7.8584913  7.9749028
#>   [34,]  0.9892106  1.0712355  1.1006208  1.1238904
#>   [35,]  1.3361651  1.8655647  2.3897110  2.8308704
#>   [36,]  5.9993223  6.6015504  6.7302893  6.7377108
#>   [37,]  2.3542091  2.4536691  2.5509296  2.6008266
#>   [38,]  5.7033906  6.2478332  6.3336286  6.3893119
#>   [39,]  4.0920571  4.4208884  4.6342948  4.6434654
#>   [40,]  3.6856697  3.8952649  3.9494692  3.9699053
#>   [41,]  3.4741424  3.7617451  3.8406493  3.8706390
#>   [42,]  5.8519913  6.4035699  6.8926209  7.3551443
#>   [43,]  3.5670544  4.4293243  4.4784322  4.5226583
#>   [44,]  3.9483665  4.4711590  4.4737003  4.4760872
#>   [45,]  3.0804812  4.4420650  4.9252960  5.3520171
#>   [46,]  1.5861833  1.9013198  2.1822117  2.2585055
#>   [47,]  2.9725313  3.3585039  3.7048112  3.7333604
#>   [48,]  4.8900925  5.7310854  5.8865376  6.0009899
#>   [49,]  3.7442547  4.2297210  4.6717602  4.8962156
#>   [50,]  3.0973922  3.7902616  4.3355586  4.7969413
#>   [51,]  3.1512758  3.4828331  3.7943211  3.8504935
#>   [52,]  3.6935633  4.4008769  4.8929049  5.3252142
#>   [53,]  3.9506610  4.3686592  4.6164246  4.7739795
#>   [54,]  7.2861678  7.9405576  8.4569114  8.4846433
#>   [55,]  3.1425379  3.3474853  3.4749063  3.5083596
#>   [56,]  6.0052903  6.7211432  7.4163580  7.9041627
#>   [57,]  3.3218865  4.4685279  5.2694653  5.8007316
#>   [58,]  6.3715278  6.7082708  7.0347721  7.1790747
#>   [59,]  1.7907982  2.4198871  2.8741229  2.9761876
#>   [60,]  6.9167705  7.9792119  8.3360779  8.4626152
#>   [61,]  5.6588888  7.9701071 10.0275410 10.3591074
#>   [62,]  1.1254819  1.4418713  1.7490398  1.8897162
#>   [63,]  5.4486375  5.8328249  6.0058586  6.1568795
#>   [64,]  3.4293798  4.0932247  4.4843604  4.7181983
#>   [65,]  1.9739035  2.3715631  2.6044214  2.7983415
#>   [66,] 10.6936514 12.7698187 14.4319235 14.4807087
#>   [67,]  4.4819653  5.1882474  5.7089127  6.0242676
#>   [68,]  7.0698398  9.1924235 10.0593994 10.3129540
#>   [69,]  4.4356387  5.3125100  5.7663535  5.8061289
#>   [70,]  1.1264152  1.5769901  1.7986138  1.9648743
#>   [71,]  3.3218941  3.6625154  3.9483458  4.2128541
#>   [72,]  4.6502911  5.8418194  6.3162971  6.4030199
#>   [73,]  5.6924185  6.0539126  6.1048436  6.1289961
#>   [74,]  3.7140707  4.0473016  4.1595963  4.2398809
#>   [75,] 11.1332449 13.5699114 13.9124152 13.9131730
#>   [76,]  7.0475421  7.9636712  8.1453924  8.1483116
#>   [77,]  5.4225242  5.7964458  6.1024994  6.3622303
#>   [78,]  9.5274400 10.9530824 12.0144515 12.9853218
#>   [79,]  5.7933216  7.2738482  8.0000144  8.1539015
#>   [80,]  1.4947410  1.9674997  2.4112379  2.7435886
#>   [81,]  2.8445219  3.9234380  4.9208124  5.0897388
#>   [82,]  5.2517311  6.4260464  6.8756071  6.9473876
#>   [83,]  3.2602910  4.2807600  4.6824784  4.6865351
#>   [84,]  5.9926129  6.9692200  7.2168215  7.4453784
#>   [85,]  1.4196656  1.8066663  2.1750505  2.3654344
#>   [86,]  2.8268316  3.3995722  3.8565992  3.9582651
#>   [87,]  6.2717232  7.5721516  8.0155704  8.3184222
#>   [88,]  9.0436658 11.7234972 12.1412891 12.3259019
#>   [89,]  7.3324132 10.0964114 11.5192343 11.6580963
#>   [90,]  1.4806551  1.9646730  2.3563300  2.4611575
#>   [91,]  4.2243551  5.1062635  5.9567224  6.6033508
#>   [92,] 11.2945922 12.7248282 13.2236851 13.2619666
#>   [93,]  7.2522452  9.3397824 10.4690787 10.9599361
#>   [94,]  2.5675991  2.9977532  3.4153647  3.7177402
#>   [95,]  7.2006849  9.7605142 10.8911482 11.0187915
#>   [96,]  4.7007744  6.0368948  6.8290305  7.4252973
#>   [97,]  5.9370510  6.9356705  7.0834678  7.2219961
#>   [98,]  3.9215584  4.8274038  5.5535510  6.0364419
#>   [99,]  4.5898167  5.4994745  5.9846739  6.3356386
#>  [100,]  8.1049064  9.2516638  9.8219376 10.2893017
#>  [101,]  5.8425101  6.8032301  7.7270017  7.8113613
#>  [102,]  4.3790423  5.0706713  5.7216334  5.9912914
#>  [103,]  5.9976845  8.4108936  8.5047372  8.5929971
#>  [104,]  3.6582921  3.8682292  4.0442379  4.0873644
#>  [105,]  3.1101859  3.5104849  3.7746950  3.8448772
#>  [106,]  5.2486140  6.7064270  7.6754244  7.9206349
#>  [107,]  3.0053193  3.2135537  3.3993538  3.5124812
#>  [108,]  3.1218766  3.8624488  4.2803682  4.4232126
#>  [109,]  3.3916988  4.4039616  5.3410019  5.4795299
#>  [110,]  2.9557245  3.2317048  3.4323943  3.5236755
#>  [111,]  5.3284950  5.4195564  5.4712033  5.5180608
#>  [112,]  4.9614475  6.5084009  7.8550725  8.5266525
#>  [113,]  1.1195129  1.4512439  1.5566945  1.5760597
#>  [114,]  6.4500179  7.5541218  7.7765336  7.9896071
#>  [115,]  3.8863170  4.3111532  4.6468909  4.6531321
#>  [116,]  1.2848655  1.7824701  1.9850293  2.0015018
#>  [117,]  3.1664721  4.4334240  5.2315461  5.2416371
#>  [118,]  5.8586390  6.3351186  6.5021672  6.5334538
#>  [119,]  2.7654915  3.2260010  3.6204737  3.7245217
#>  [120,]  4.5089128  5.3135806  6.1009181  6.4868401
#>  [121,]  6.5209401  7.7822078  8.7498654  8.7618120
#>  [122,]  4.8726818  5.8645918  6.2532596  6.4343514
#>  [123,]  2.0909201  2.5336483  2.6999269  2.7236372
#>  [124,]  4.1577597  5.1923278  5.7632709  5.8528631
#>  [125,]  9.3921183 11.3403083 11.9240499 12.3780008
#>  [126,]  2.5705764  3.5020414  4.2204959  4.7777934
#>  [127,]  2.6740333  3.1621965  3.6027013  3.7722105
#>  [128,]  6.2356095  6.8221990  7.1653360  7.1672844
#>  [129,]  7.3474340  8.9849909 10.3836948 10.6220248
#>  [130,]  2.6680183  3.1284082  3.5285815  3.5288129
#>  [131,]  5.4986046  6.8319756  7.9294735  8.7520802
#>  [132,]  5.4261176  5.9313856  6.2001475  6.2279490
#>  [133,]  1.9608887  2.0777729  2.1585952  2.2182029
#>  [134,]  1.8195195  2.5736180  2.7801437  2.8014257
#>  [135,] 12.9998542 14.8681771 16.2677658 17.4942686
#>  [136,]  2.2091756  2.9696002  3.4936413  3.5585610
#>  [137,]  1.8741931  2.5464813  2.8204753  2.8294650
#>  [138,]  4.0105115  5.2972385  5.5365805  5.6273600
#>  [139,]  6.1255920  6.8104378  6.8327985  6.8493049
#>  [140,]  8.6521403  8.9358645  9.2085893  9.4798860
#>  [141,]  5.2047955  6.1025961  6.9833008  7.8575202
#>  [142,]  4.2282774  5.0901488  5.5922594  5.7569249
#>  [143,]  0.7767254  1.0164320  1.2266790  1.3994706
#>  [144,]  2.7980119  3.2013873  3.3657063  3.4640917
#>  [145,]  4.8086065  6.5441014  7.0974084  7.5455991
#>  [146,]  1.6220491  1.9647853  2.2539152  2.2897248
#>  [147,]  4.1579731  4.7530870  5.2824801  5.6329791
#>  [148,]  9.0887024  9.2248926  9.3474215  9.4407070
#>  [149,]  0.8845941  1.1778487  1.2814157  1.3142163
#>  [150,]  5.1192326  5.3873927  5.5334071  5.5957186
#>  [151,]  7.3846842  7.8723614  8.1173524  8.1498258
#>  [152,]  4.0628905  4.6203463  4.7096261  4.7377991
#>  [153,]  1.8659880  2.3384118  2.7762369  3.0700969
#>  [154,]  6.8844846  8.2449408  9.0088809  9.5029767
#>  [155,]  1.9417761  2.1938950  2.4257598  2.6240619
#>  [156,]  1.4789327  1.7963530  1.9734523  2.0569781
#>  [157,]  2.0739338  2.6331943  3.1615251  3.3158082
#>  [158,]  6.1070213  7.1037425  8.0940760  8.8162175
#>  [159,]  2.3415598  2.7568627  2.8314742  2.8602188
#>  [160,]  3.1709677  4.1277155  5.0021646  5.0515074
#>  [161,]  1.3380252  1.5868823  1.8044999  1.8919991
#>  [162,]  3.6314991  4.7939513  5.5428514  6.2742477
#>  [163,]  5.3176939  6.1006103  6.6749680  7.1008112
#>  [164,]  4.0585929  4.8885121  5.2580712  5.3351646
#>  [165,]  4.1855678  5.6441035  6.3345399  6.6570672
#>  [166,]  1.6488671  1.9146748  2.1255492  2.2162754
#>  [167,]  4.5790586  5.4156457  6.0185980  6.5198392
#>  [168,]  3.4333044  3.9584074  4.1410016  4.1943786
#>  [169,]  0.3641027  0.3964439  0.4238496  0.4293742
#>  [170,]  4.7856028  4.9433028  5.0555618  5.1658751
#>  [171,]  0.4728770  0.5792951  0.6825073  0.7828095
#>  [172,]  1.5495636  2.1851045  2.5347975  2.8591774
#>  [173,]  4.7719011  5.7094071  6.3208935  6.6327810
#>  [174,]  5.8364216  7.2789356  7.4187327  7.4828516
#>  [175,]  2.9584911  3.3260991  3.5302150  3.6150641
#>  [176,]  4.9332049  5.6074545  6.0515332  6.3634564
#>  [177,]  2.4215731  2.7338593  2.9842853  3.1819044
#>  [178,]  6.4556041  6.6531034  6.8248767  6.9179637
#>  [179,]  4.4516613  4.9087168  5.1236956  5.1826677
#>  [180,]  1.6530095  2.1520022  2.3031844  2.3920488
#>  [181,]  0.9930094  1.3930187  1.5572806  1.6714455
#>  [182,]  2.6964584  3.2471601  3.5864539  3.9056149
#>  [183,]  7.8265996  8.9750138  9.9822905 10.3382465
#>  [184,]  6.0259360  6.9092910  7.3538838  7.7449844
#>  [185,]  5.0173485  5.8050739  5.9802789  6.0727830
#>  [186,]  2.4333930  3.3629668  3.8738715  3.9966945
#>  [187,]  9.5656128 12.0637431 12.8269464 13.3229749
#>  [188,]  4.0265214  5.3627469  5.5874344  5.6380600
#>  [189,]  2.7848127  3.3988188  3.6931252  3.8479085
#>  [190,]  3.6066251  4.4813418  5.2489033  5.3211573
#>  [191,]  3.7484250  4.4268719  4.5306910  4.5383598
#>  [192,]  6.9064860  8.8596290  9.6010970 10.2216083
#>  [193,]  0.6311384  0.7037506  0.7093644  0.7119530
#>  [194,]  7.2607596  8.0368044  8.4225272  8.6461601
#>  [195,]  5.3544943  5.9197898  6.0519729  6.1710515
#>  [196,]  4.7753761  5.8103302  6.2434713  6.3832118
#>  [197,]  6.2066311  7.5170424  8.6790868  9.6911767
#>  [198,]  2.8892114  3.5376207  4.0128886  4.4592565
#>  [199,]  1.8127726  2.3707674  2.6455197  2.7041805
#>  [200,]  3.5283251  5.1745211  5.9023103  6.4506536
#>  [201,]  3.6723918  4.6621610  5.4499812  5.6329069
#>  [202,]  1.5593513  1.9557042  2.2775780  2.5667658
#>  [203,]  2.8781685  3.6482719  3.7221368  3.7474479
#>  [204,]  8.3410996 10.3391152 11.4626931 12.0865609
#>  [205,]  3.4771781  3.6050472  3.6649272  3.7194228
#>  [206,]  5.6327580  6.3934073  6.9430272  7.4373185
#>  [207,]  4.0172991  4.1493056  4.2181631  4.2356585
#>  [208,] 10.4417146 11.9342716 13.1296579 13.6383881
#>  [209,]  3.4057097  4.5685701  5.6027456  6.2837858
#>  [210,]  3.1921234  4.0150231  4.8102336  5.5611053
#>  [211,]  1.0987305  1.3316732  1.3694463  1.4052214
#>  [212,]  8.1850197  8.4062540  8.4355012  8.4443638
#>  [213,]  1.8248749  1.9275788  1.9433524  1.9475059
#>  [214,]  2.5905390  2.9994768  3.1565469  3.2657489
#>  [215,]  3.5984202  4.2296322  4.3460376  4.3749277
#>  [216,]  4.5329073  6.2204884  7.3568355  7.5116253
#>  [217,]  3.9813467  4.0873274  4.1671038  4.2423632
#>  [218,]  2.7958041  3.3478501  3.4371943  3.5004932
#>  [219,]  5.2062659  5.7897559  6.0849814  6.2500954
#>  [220,]  7.3923022  8.8354132  9.4723859  9.8969008
#>  [221,]  4.0195702  4.3464786  4.5620196  4.5835726
#>  [222,]  7.5278118  8.3309137  8.9629882  9.2830585
#>  [223,]  1.3540836  1.8754332  2.0191000  2.1077727
#>  [224,]  2.4588634  2.8061560  3.0942366  3.3111378
#>  [225,]  6.2892169  7.0639740  7.6165382  7.7644822
#>  [226,]  0.9795663  1.3963712  1.7889567  2.1523591
#>  [227,]  3.1426212  3.5316925  3.7827584  4.0034328
#>  [228,]  5.3662613  6.7409520  7.6115707  7.6200268
#>  [229,]  5.2016855  5.8839638  6.4730172  6.5940363
#>  [230,]  4.7458851  6.3701547  6.5717302  6.6987933
#>  [231,]  0.4137556  0.5236088  0.5767519  0.6212017
#>  [232,]  3.0887069  3.8344794  4.1633740  4.3405707
#>  [233,]  5.7193843  6.5008019  6.8411514  7.0843814
#>  [234,]  9.3262957 10.2044814 10.5666032 10.6402141
#>  [235,]  4.4318100  4.8791896  5.1797220  5.2632629
#>  [236,]  3.3593232  3.9798015  4.1971314  4.2109906
#>  [237,]  3.7440792  4.1165517  4.1606996  4.1930125
#>  [238,]  4.0383584  5.1620449  5.4221421  5.4342902
#>  [239,]  6.6330489  8.0712698  9.2392842  9.3804313
#>  [240,]  4.7382051  5.6265986  5.8579364  6.0323840
#>  [241,]  3.2864830  3.8294442  4.2703194  4.5947988
#>  [242,]  2.2185723  2.9377833  3.5902840  3.7235047
#>  [243,]  1.2575642  1.4723907  1.5678046  1.6267610
#>  [244,]  5.9665681  7.2171414  7.3942123  7.4114334
#>  [245,]  0.6918649  0.7993736  0.8392500  0.8712908
#>  [246,]  2.7192130  3.4529455  3.5991591  3.6563259
#>  [247,]  8.1693992  9.4105298 10.0068809 10.3119116
#>  [248,]  3.7318070  4.7676911  5.4872801  5.9977726
#>  [249,]  7.6029012  9.6500457 10.5416773 10.6667757
#>  [250,]  4.3051477  5.4821432  6.0773942  6.4006001
#>  [251,]  1.5769502  1.8739528  1.9571270  1.9977797
#>  [252,]  7.4623698  9.4403738 10.3102542 10.7841102
#>  [253,]  6.9118923  7.5475949  7.6503894  7.6568354
#>  [254,]  3.5448221  4.4078585  4.4713014  4.5297563
#>  [255,]  6.0665348  6.4166598  6.5980445  6.6497789
#>  [256,]  6.3715924  6.9937118  7.2865605  7.5234268
#>  [257,]  4.0643455  4.3036335  4.3685726  4.4164231
#>  [258,]  5.8130766  6.0637115  6.2301799  6.2531427
#>  [259,]  3.5001918  4.9518809  5.0799606  5.1074229
#>  [260,]  3.2628153  3.9296647  4.4381707  4.5988122
#>  [261,]  0.7386049  0.9388134  1.1353095  1.1985694
#>  [262,]  6.0542128  6.5036265  6.8362065  6.9849395
#>  [263,]  6.9913101  7.4728472  7.6666426  7.8238531
#>  [264,]  3.1957029  4.2360580  4.9462662  5.2998478
#>  [265,]  4.3218499  4.9127058  5.2634923  5.5628133
#>  [266,]  2.5361255  2.8788356  3.0863311  3.2831070
#>  [267,] 14.4064531 18.7712840 20.7849420 20.8499328
#>  [268,]  4.4379476  4.8110413  5.1601582  5.1776783
#>  [269,]  1.7778076  2.2184326  2.6306360  2.6542276
#>  [270,]  2.8668713  3.7839204  4.0086313  4.1290198
#>  [271,]  3.1352535  3.7292320  3.8078079  3.8283731
#>  [272,]  7.0911335  7.5494676  7.6064304  7.6363090
#>  [273,]  1.0085045  1.1450894  1.1539347  1.1580015
#>  [274,]  4.3955920  5.8328512  6.6071318  6.9493156
#>  [275,]  5.3034766  5.3910474  5.4477179  5.4793983
#>  [276,]  2.2388511  3.0067370  3.5603716  3.8591191
#>  [277,]  2.6999875  3.6480001  4.0546043  4.0722201
#>  [278,]  0.9718929  1.1889898  1.3060547  1.3897091
#>  [279,]  2.6597616  3.0739449  3.4713832  3.7720180
#>  [280,]  3.7676869  4.2383521  4.5456865  4.6369519
#>  [281,]  6.7485695  7.8179717  8.1431361  8.3859167
#>  [282,]  9.0074222  9.2887207  9.3413439  9.3464697
#>  [283,]  3.1290871  4.1803381  4.4140020  4.4581727
#>  [284,]  2.1500134  2.6168343  2.8393304  2.9058972
#>  [285,]  2.4305674  3.0322853  3.5568455  3.8445072
#>  [286,]  4.9459315  7.0286199  7.7728082  8.2869897
#>  [287,]  4.5763066  4.7374653  4.8332110  4.9210941
#>  [288,]  5.6147611  6.0866320  6.2022873  6.2320251
#>  [289,]  2.9285097  3.3413684  3.6914016  3.9197386
#>  [290,]  2.7116425  3.4086997  3.4181519  3.4194984
#>  [291,]  2.9426194  3.4102063  3.8378320  3.8960293
#>  [292,]  1.9491530  2.2018993  2.4047071  2.4270505
#>  [293,]  3.2119307  3.9620922  4.3147312  4.5997819
#>  [294,] 14.1915903 16.6140642 17.9408947 18.7647713
#>  [295,]  8.2084561 10.9128174 12.9604316 13.6616417
#>  [296,]  5.5173260  5.9061724  6.2248587  6.3561949
#>  [297,]  2.8179718  3.3504391  3.7668450  3.9717501
#>  [298,]  0.7754705  0.8266028  0.8642048  0.8741767
#>  [299,]  3.7159654  5.1103436  6.2982478  6.3316414
#>  [300,]  3.8373330  4.8814934  5.4067441  5.8149167
#>  [301,]  5.2680949  6.0046476  6.1590086  6.2279321
#>  [302,]  3.2777337  3.5554805  3.6493788  3.7168737
#>  [303,]  2.1559053  2.5750028  2.8432518  3.0756678
#>  [304,]  2.2171534  2.4769541  2.6606862  2.8350152
#>  [305,]  3.5948831  4.3618540  4.5337100  4.5900441
#>  [306,]  6.1341742  6.5610265  6.6601834  6.7004963
#>  [307,]  5.8828909  8.1421980  8.2016009  8.2558031
#>  [308,]  2.1960984  3.0623909  3.8404029  3.8827028
#>  [309,]  4.3159420  5.4425447  6.5063571  6.5797904
#>  [310,]  1.5476443  1.7513731  1.8166153  1.8811586
#>  [311,]  4.8457604  5.9334249  6.6807880  7.0153251
#>  [312,]  6.1857389  7.3399550  7.8023814  8.2586348
#>  [313,]  2.5491127  3.1214510  3.4752244  3.7695273
#>  [314,]  4.9175502  5.6754242  6.3448495  6.7435496
#>  [315,] 15.2206038 15.6405763 15.7578468 15.8350498
#>  [316,]  3.5498734  4.0384473  4.3333385  4.6206147
#>  [317,]  4.4572355  6.2850491  7.2413363  7.8391416
#>  [318,] 12.5989519 13.5603078 14.1743884 14.4247236
#>  [319,]  6.1051828  6.5367336  6.8589211  6.9784903
#>  [320,]  3.4613017  4.4503674  5.2033494  5.4881387
#>  [321,]  1.3973293  1.6120943  1.7828861  1.9376779
#>  [322,]  8.8127192 10.1451890 11.4759498 11.8197458
#>  [323,]  1.9180616  2.0635407  2.2045184  2.3437823
#>  [324,]  1.9342422  2.0342202  2.1049163  2.1124703
#>  [325,]  1.4749783  1.6929653  1.8904401  2.0834119
#>  [326,]  2.2572336  2.6958304  3.0977989  3.4771454
#>  [327,]  1.2009870  1.2716504  1.3255595  1.3442088
#>  [328,]  1.8719633  2.3424104  2.5014877  2.5040124
#>  [329,]  5.6670720  6.5997423  7.4791451  8.3021352
#>  [330,]  2.6168569  2.9821728  3.1791333  3.2269017
#>  [331,]  4.5358071  4.9271675  5.0714583  5.1856548
#>  [332,]  1.4873595  1.6529723  1.8130030  1.9382257
#>  [333,]  1.3577801  1.4520893  1.4604874  1.4628104
#>  [334,]  3.8550154  4.5438745  4.8611449  4.8809935
#>  [335,]  6.0859404  6.7946268  6.9858582  7.1274653
#>  [336,]  2.0650400  2.4900747  2.5592562  2.5902796
#>  [337,]  1.2261694  1.5429137  1.6008026  1.6412620
#>  [338,]  3.0188732  3.3240769  3.6094549  3.6336714
#>  [339,]  1.1136577  1.3336445  1.3790327  1.4103901
#>  [340,]  3.0880888  3.1604249  3.1964051  3.2250139
#>  [341,]  4.8818242  5.4634906  5.7555020  5.9282574
#>  [342,]  5.0102444  6.6852654  7.0224571  7.1890923
#>  [343,]  6.7165396  7.7922249  8.7937358  9.4647712
#>  [344,]  5.5059450  6.9463345  7.8569011  8.3957658
#>  [345,] 11.0343189 14.2220638 15.3575841 15.9620001
#>  [346,]  7.7138048  8.3434179  8.8701627  9.2744713
#>  [347,]  8.8162916 12.5937279 13.4323478 13.5045748
#>  [348,]  5.4914881  6.1022017  6.6607201  7.2174583
#>  [349,]  5.4568368  6.5133922  6.7200368  6.8671807
#>  [350,]  1.6414478  2.0137295  2.1290015  2.1595506
#>  [351,]  6.4704367  9.2572360 10.1483744 10.3314715
#>  [352,]  5.3844560  7.6030539  9.6568470 10.5565066
#>  [353,]  1.8688413  2.3595559  2.8017873  2.8867117
#>  [354,]  0.7995030  1.0626936  1.2100176  1.3175460
#>  [355,]  5.8717351  7.7564612  9.6342053  9.7481630
#>  [356,] 10.5142077 13.2900516 14.4401609 14.7975520
#>  [357,]  1.1242050  1.4453037  1.6957516  1.9407355
#>  [358,]  8.0470643  9.7347750 10.8025139 10.9896081
#>  [359,]  2.9575361  3.9723903  4.8216024  5.1646039
#>  [360,]  9.7187241 11.5949439 12.4385883 12.6118803
#>  [361,]  3.1916845  4.2549613  4.3861704  4.4576190
#>  [362,]  6.9518865  7.9559885  8.7462986  9.1372698
#>  [363,]  7.7604064  9.4361241 10.8352356 11.5530395
#>  [364,]  7.7692160  9.4905853 11.0716027 12.5263944
#>  [365,]  4.1332693  4.4280371  4.6649952  4.7507009
#>  [366,]  2.4443609  2.5721392  2.6949820  2.7866271
#>  [367,]  3.5533397  4.7794334  5.9134602  6.7292130
#>  [368,]  3.7456782  4.8410799  5.8114398  5.9113024
#>  [369,]  2.7375900  3.3844194  3.8121276  3.8871400
#>  [370,]  6.5202880  6.9470971  7.3067415  7.4523468
#>  [371,]  9.9241351 14.1025524 15.1631712 15.6895288
#>  [372,]  3.7976098  4.5597033  4.6824442  4.7126861
#>  [373,]  3.0686115  3.8868481  4.2273924  4.5244103
#>  [374,]  2.1935943  2.5888060  2.9015468  2.9723471
#>  [375,]  6.0393015  6.7886331  7.1220856  7.2656790
#>  [376,]  2.6255269  3.1553850  3.5649170  3.6338051
#>  [377,]  5.3496874  6.5256437  7.1996127  7.7280313
#>  [378,]  4.3541768  6.3341545  7.1488002  7.5042298
#>  [379,]  3.0468711  4.0497359  4.8163456  4.8388417
#>  [380,]  8.0118923  8.1883020  8.2611818  8.3103335
#>  [381,]  1.4519841  1.6938624  1.8489104  1.9410242
#>  [382,]  1.8866673  2.2809979  2.5674567  2.5987362
#>  [383,] 10.2359441 11.1426566 11.6009931 11.6456146
#>  [384,]  8.2421466 10.3822541 10.9325152 11.2144653
#>  [385,]  2.1107499  2.4694449  2.7956912  2.8371667
#>  [386,]  4.0115059  5.1903840  6.2815051  7.3222790
#>  [387,]  3.0302821  3.1676160  3.2631241  3.2727704
#>  [388,]  2.1071500  2.7462045  3.0515313  3.1367033
#>  [389,]  4.3991997  5.8934477  7.1677322  7.7374671
#>  [390,]  6.6069332  7.5141623  7.9097281  8.2622649
#>  [391,]  5.5244768  5.8912610  5.9619680  5.9802143
#>  [392,]  3.6152894  4.0915780  4.5421659  4.8720641
#>  [393,]  5.1932485  6.4008894  6.6425766  6.6809772
#>  [394,]  6.7168942  8.0372798  8.1776253  8.2849416
#>  [395,]  2.9329705  3.3412448  3.6901375  3.8939209
#>  [396,]  2.6984844  3.2662689  3.7766638  3.7846355
#>  [397,]  3.5122474  4.4477579  4.6860824  4.7912467
#>  [398,]  1.9808660  2.3432404  2.5737493  2.6751763
#>  [399,]  3.4632982  4.6351060  5.2198005  5.3159691
#>  [400,]  2.0925659  2.4136513  2.7021139  2.9612416
#>  [401,]  1.0928752  1.4018787  1.4799987  1.5196048
#>  [402,] 12.1562023 14.7737759 15.0713419 15.1010440
#>  [403,]  6.7571803  7.7241286  8.0065348  8.2506475
#>  [404,]  2.4011636  3.0207392  3.3944067  3.6660187
#>  [405,]  2.1629260  2.6237651  2.6761279  2.7114290
#>  [406,]  2.8865554  3.1488719  3.3370687  3.3980493
#>  [407,]  5.4643020  6.9688346  7.6978010  8.0994962
#>  [408,]  3.1250235  3.7289121  4.0975264  4.1148422
#>  [409,]  7.1876401  9.0768832 10.7500327 11.3212944
#>  [410,]  2.1793958  2.7093475  2.8428050  2.9197949
#>  [411,]  4.0346165  4.7851509  5.3289834  5.7847392
#>  [412,]  3.0347408  3.9977780  4.7250986  5.1215766
#>  [413,]  2.5024533  3.0728372  3.2716967  3.2926910
#>  [414,]  5.7965005  6.0157437  6.1504203  6.1865026
#>  [415,] 11.5832494 12.9024705 13.0787273 13.2208245
#>  [416,]  2.5632557  3.2784714  3.7571152  3.9158182
#>  [417,]  5.8759260  6.9985202  7.7018253  8.1997070
#>  [418,]  5.2745162  5.8439262  6.2420478  6.5551766
#>  [419,]  1.0661546  1.4297056  1.7313939  1.9535955
#>  [420,]  1.9257877  2.7632078  3.5378183  4.1456050
#>  [421,]  2.2699993  2.7662960  2.9033860  2.9411518
#>  [422,]  2.6296329  2.9272779  3.2196211  3.2880717
#>  [423,]  1.8129101  1.8468534  1.8675025  1.8779416
#>  [424,]  1.8647347  2.2083883  2.4526477  2.4943686
#>  [425,]  4.4981587  5.4924935  6.1480659  6.4537453
#>  [426,]  6.6156948  7.3219261  7.7204607  7.7418353
#>  [427,]  3.8478927  4.5032687  4.9987125  5.3732565
#>  [428,]  2.9541512  4.2144095  5.0930777  5.5994867
#>  [429,]  3.7288531  4.6127296  4.7718337  4.8478812
#>  [430,]  2.0358078  2.1736003  2.2007527  2.2103547
#>  [431,]  2.0698539  2.3535701  2.3631329  2.3725473
#>  [432,]  4.0406643  4.7014512  5.2778476  5.8281150
#>  [433,]  1.5246270  2.0498854  2.4694363  2.6038679
#>  [434,]  5.7891131  7.9370240  9.3349615  9.3721563
#>  [435,]  4.8828563  5.7891554  6.4416562  6.4868439
#>  [436,]  2.3092213  2.8567056  2.8717764  2.8786547
#>  [437,]  3.3429250  3.5494237  3.7100686  3.7351984
#>  [438,]  1.6711201  1.9853878  2.0443622  2.0453365
#>  [439,]  7.4998847 10.8962823 12.7398347 13.5561815
#>  [440,]  3.2016193  3.2826363  3.3219972  3.3589168
#>  [441,]  1.5844180  2.1721519  2.6865224  3.1744463
#>  [442,]  5.0329875  5.4908263  5.5777211  5.6437746
#>  [443,]  4.0590080  5.3821755  6.6386161  7.5112386
#>  [444,]  6.4137778  7.5235003  8.2923765  8.3700346
#>  [445,]  2.4040591  2.8726541  3.3101870  3.6834800
#>  [446,]  3.2260627  4.1131621  4.5258566  4.6470384
#>  [447,]  7.3907866  7.8297344  7.8857018  7.9297184
#>  [448,]  2.6591297  3.1188504  3.4594088  3.6773387
#>  [449,]  0.9208242  1.0376763  1.1477919  1.1970691
#>  [450,]  2.0657531  2.7197152  3.0467900  3.1606091
#>  [451,]  1.2312200  1.4330670  1.5670107  1.6402401
#>  [452,]  5.5471914  5.9715118  6.2973728  6.5303745
#>  [453,]  2.3947210  2.6466861  2.7989423  2.8511443
#>  [454,]  3.7390299  4.6774043  5.4250519  5.4858908
#>  [455,]  2.7661301  3.8202186  3.9381420  4.0507962
#>  [456,]  3.6801328  4.7085946  5.4009146  5.5816584
#>  [457,]  6.1724466  7.4945284  8.0428167  8.2562014
#>  [458,]  5.3051151  5.6017658  5.8114842  5.9101690
#>  [459,]  3.9950133  4.1355635  4.1747008  4.1882324
#>  [460,]  1.4840647  1.9690802  2.2349076  2.3500587
#>  [461,]  2.0997879  2.8734618  3.5328731  3.6508954
#>  [462,]  2.8304771  2.9605107  3.0451837  3.0703194
#>  [463,]  5.1753741  6.4428701  7.3318737  7.3582370
#>  [464,] 10.2131604 12.3468227 12.7900085 12.9041257
#>  [465,]  2.9553190  3.5336480  4.0161410  4.1901078
#>  [466,]  3.7143527  4.8321638  5.3156669  5.5223375
#>  [467,]  5.3951013  5.9220031  6.4104505  6.7195544
#>  [468,]  2.7842979  3.1508741  3.3709160  3.5868017
#>  [469,]  1.4064121  1.7373510  1.9454090  1.9810941
#>  [470,]  5.1968585  6.0928324  6.5525112  6.6590965
#>  [471,]  4.2550466  5.4118293  6.0864337  6.2873116
#>  [472,]  2.1464543  2.4698038  2.6940273  2.7095142
#>  [473,]  6.5162813  6.6703578  6.8041890  6.8996177
#>  [474,]  7.1562513  8.5412384  8.7368964  8.8339784
#>  [475,]  2.2481578  2.4897117  2.7151521  2.8311064
#>  [476,] 10.9515835 13.1128310 13.5361623 13.8887125
#>  [477,]  9.6366384 12.3137380 13.4332310 13.8950301
#>  [478,] 11.3703507 12.3129515 12.7217354 12.8353122
#>  [479,]  5.5795297  7.1142553  8.3135980  8.9307430
#>  [480,]  3.9231702  5.0031196  5.5935579  6.1034951
#>  [481,]  7.0457470  8.3178376  8.6248020  8.6656682
#>  [482,]  1.5367701  1.8696862  1.9755713  2.0759268
#>  [483,]  4.0947596  4.3254925  4.4599824  4.4849977
#>  [484,]  2.6067104  2.8771369  3.1018760  3.2271761
#>  [485,]  5.0892410  5.1975843  5.2590669  5.2853693
#>  [486,]  6.0915719  8.1203134  9.3525502  9.5137412
#>  [487,]  4.5128485  5.2448317  5.5217004  5.5780438
#>  [488,]  2.8713933  3.2573660  3.5790957  3.8239013
#>  [489,]  2.3523990  2.7771257  3.1242812  3.2636659
#>  [490,]  2.8903229  3.2790699  3.5806436  3.8346921
#>  [491,]  3.8303228  4.7820580  4.9825157  5.1524667
#>  [492,]  4.1962506  4.9373302  5.3550763  5.4291793
#>  [493,]  2.6049345  3.0078220  3.2991170  3.3807323
#>  [494,]  2.3581072  3.2154596  3.7472050  4.2597187
#>  [495,]  2.1856446  2.6651916  2.8513182  3.0367077
#>  [496,]  1.9836283  2.8399734  3.3772603  3.5334851
#>  [497,]  6.2963229  7.4162008  8.2519862  8.7932553
#>  [498,]  2.3805256  2.6155486  2.6730061  2.7245703
#>  [499,] 10.4023854 14.0289975 15.2986774 15.6514537
#>  [500,]  4.3475699  5.9218970  7.2713951  8.0568355
#>  [501,]  1.9234733  2.7123879  3.2283507  3.2796323
#>  [502,]  1.8202926  2.1877760  2.2403311  2.2915916
#>  [503,]  1.7739588  2.1363152  2.4585328  2.5769885
#>  [504,]  7.5443666  8.9944686  9.4124816  9.4762356
#>  [505,] 13.9522300 15.6692771 16.7654787 17.4252701
#>  [506,]  2.4544467  3.5165900  4.0872026  4.4682998
#>  [507,]  5.4310123  7.7317164  8.4168353  8.8103396
#>  [508,]  2.0379088  2.5776370  2.7052044  2.7716783
#>  [509,]  1.5608504  1.7201106  1.7915596  1.8315759
#>  [510,]  6.6153147  8.5416203 10.4455255 10.5976466
#>  [511,]  4.3151795  5.3700910  5.6429521  5.7334291
#>  [512,]  4.0671517  4.2527328  4.3814128  4.4018773
#>  [513,]  8.0302050  9.5889109 10.3809516 10.8641204
#>  [514,] 12.6507070 15.1107724 15.3409048 15.4749241
#>  [515,]  8.7679732  9.3014667  9.4228878  9.4432261
#>  [516,]  7.3467652  8.2748664  8.5916240  8.5982834
#>  [517,]  1.8232916  2.5530487  2.9066497  3.0506892
#>  [518,]  4.1334358  4.3641912  4.5627236  4.7140672
#>  [519,]  3.1354926  3.6725032  4.0935831  4.3214437
#>  [520,]  2.2801107  2.4193784  2.5398986  2.5871162
#>  [521,]  1.5717219  1.8910366  2.2056737  2.2244620
#>  [522,]  3.1405887  4.0347449  4.0634963  4.0901088
#>  [523,] 12.7329490 13.8311111 14.4008691 14.9427110
#>  [524,]  8.7944791  9.4272064  9.9038829 10.1913693
#>  [525,]  3.0301487  3.4732234  3.7392828  3.8403675
#>  [526,]  1.4213850  1.5460872  1.6683995  1.7801515
#>  [527,]  3.0573573  3.5931514  3.7802735  3.8110984
#>  [528,]  3.0384202  3.0682204  3.0883036  3.0945721
#>  [529,]  8.1723416 10.2870663 11.7173257 13.0243055
#>  [530,]  2.2763732  3.0237410  3.1195130  3.1472953
#>  [531,]  1.5043234  1.9485726  2.2414891  2.3566293
#>  [532,]  5.3603197  5.4791904  5.5445644  5.5597357
#>  [533,]  5.1990231  5.8061548  6.1587754  6.2027776
#>  [534,]  2.7648344  3.4774708  3.9527105  4.1965274
#>  [535,]  2.6442318  3.1202351  3.5134176  3.6226699
#>  [536,]  3.1867331  3.2580386  3.3257315  3.3602366
#>  [537,]  1.4827654  1.7617297  1.9954665  2.0212646
#>  [538,]  2.9713104  3.2426924  3.3903300  3.4344251
#>  [539,]  3.6595241  4.5712383  4.7203006  4.8687706
#>  [540,]  3.6550414  5.1271575  5.9771227  6.2878411
#>  [541,]  2.3916338  3.3828403  3.9308268  4.4111625
#>  [542,]  8.0814004 10.6127437 13.0527080 13.3226775
#>  [543,]  6.0558010  7.3375080  8.5971683  9.4756776
#>  [544,] 11.5364323 15.4029642 18.1994227 18.7163080
#>  [545,]  1.9828757  2.2174499  2.3641381  2.3802118
#>  [546,]  4.7519158  5.5784798  6.2026212  6.3178959
#>  [547,]  4.8983455  5.5246959  6.0772660  6.5989755
#>  [548,]  2.9067962  3.0719012  3.1901443  3.2991780
#>  [549,]  5.8493248  5.9219432  5.9813060  6.0316909
#>  [550,]  3.6333766  4.0732794  4.1536557  4.2148894
#>  [551,]  1.4051487  1.6518942  1.8788673  1.9774304
#>  [552,]  3.8043386  5.1924431  6.0770266  6.4545387
#>  [553,]  3.9828462  4.4907697  4.6831085  4.7068223
#>  [554,]  3.3494870  4.9829821  6.1251303  6.5551278
#>  [555,]  5.6975591  6.2548415  6.5536285  6.6530060
#>  [556,]  1.7576119  1.9802621  2.0481730  2.1058339
#>  [557,]  6.2244137  7.4836316  7.7895228  7.8186098
#>  [558,]  1.8275498  2.3977031  2.6628971  2.7687848
#>  [559,]  3.7739030  4.7377454  5.1240517  5.2965515
#>  [560,]  2.1300887  2.5571542  2.9629838  3.2887235
#>  [561,]  5.6157490  6.2491591  6.3215215  6.3936358
#>  [562,]  7.6599855  8.2765789  8.7951149  9.1026334
#>  [563,]  2.0784338  2.2133283  2.3306637  2.3359660
#>  [564,]  5.9752457  6.6265160  7.0404055  7.1092582
#>  [565,]  7.6465471 10.6036707 12.2257656 12.4095491
#>  [566,]  5.7513451  7.6981754  9.5380364 11.0212321
#>  [567,]  2.8784219  3.2191443  3.3063202  3.3677449
#>  [568,]  4.9115466  5.2799868  5.5982833  5.6401193
#>  [569,]  2.2869127  2.5956208  2.7543300  2.8906675
#>  [570,]  1.9954207  2.6256616  2.7116846  2.7881555
#>  [571,]  4.5053553  4.8480898  5.1315037  5.2437907
#>  [572,]  6.5528310  9.2684108 10.5858659 10.9902721
#>  [573,]  3.4374612  4.1236405  4.5166050  4.5688366
#>  [574,]  2.1601271  2.3382813  2.4454531  2.5416273
#>  [575,]  4.3225103  5.3151405  5.8576834  6.0488509
#>  [576,]  2.5455804  2.7869930  2.8746067  2.8937016
#>  [577,]  2.8170133  3.2912780  3.3679077  3.4423617
#>  [578,]  5.6802094  7.6255116  8.3905348  8.4913135
#>  [579,]  8.2170084  8.5292213  8.7847154  8.9916472
#>  [580,]  1.2801006  1.6806956  2.0537275  2.1989263
#>  [581,]  4.6840679  6.3590008  7.4134099  8.1873728
#>  [582,]  2.8203339  3.1297511  3.2888255  3.3034583
#>  [583,]  2.9758810  3.3309855  3.5029603  3.5805063
#>  [584,]  5.9694370  7.2654521  7.2943339  7.3182327
#>  [585,]  1.4486080  1.5064894  1.5373487  1.5441343
#>  [586,]  6.9294911  7.3230143  7.6486627  7.8395390
#>  [587,]  2.2816063  2.5800817  2.8573468  2.9214496
#>  [588,]  1.0036363  1.2429400  1.3447661  1.3829091
#>  [589,]  8.7438647  9.3505657  9.6714459  9.7501470
#>  [590,]  4.6364226  5.5696036  6.2640734  6.6706369
#>  [591,]  2.4964286  2.8410344  3.0420517  3.1638533
#>  [592,]  1.1201400  1.4626055  1.7277116  1.9271675
#>  [593,]  2.2426644  3.1020512  3.2136864  3.2202707
#>  [594,]  3.2902336  4.1652537  4.7159386  5.2561528
#>  [595,]  0.8764013  1.1200553  1.2666215  1.3685571
#>  [596,]  5.9184415  6.7895975  7.3526517  7.3659524
#>  [597,]  2.1022220  2.5448049  2.9165180  2.9790379
#>  [598,]  6.2668261  8.2627440  9.3149521  9.6320005
#>  [599,]  4.4261930  5.8574468  6.8601531  7.7589751
#>  [600,]  7.9612169  9.0740709  9.7787116 10.2811530
#>  [601,]  2.5318207  3.0896342  3.5454789  3.6434442
#>  [602,]  6.5998919  7.5528362  7.8072816  8.0116911
#>  [603,]  2.9624290  3.4787388  3.7039931  3.9031037
#>  [604,]  4.3827538  5.1811795  5.8168045  5.9040145
#>  [605,]  2.8995173  3.9945965  4.0071043  4.0082850
#>  [606,]  5.9065839  6.1513038  6.3319707  6.3414303
#>  [607,]  3.5185140  4.6301965  5.2482614  5.4004393
#>  [608,]  8.8261973  8.9814396  9.0749086  9.0918503
#>  [609,] 11.0279717 11.4527744 11.8063536 11.8612028
#>  [610,]  1.3141485  1.6485630  1.9691695  2.0147416
#>  [611,]  2.0392395  2.6486470  2.7901010  2.8734982
#>  [612,]  4.0307924  4.4246829  4.6594660  4.7229134
#>  [613,]  3.1449746  4.2119423  4.8418186  5.2098493
#>  [614,]  2.8392269  3.2870356  3.6772975  4.0406442
#>  [615,]  5.1101544  6.3889399  6.9062432  6.9143734
#>  [616,]  2.8704692  3.8144881  4.6735713  4.7881967
#>  [617,]  2.1499933  2.7895946  3.1357980  3.3601917
#>  [618,]  5.3071222  5.9309449  6.4334564  6.5390712
#>  [619,]  5.2881865  7.0489344  7.5124436  7.8949415
#>  [620,]  4.0446761  5.2939560  5.7870651  6.0160784
#>  [621,]  1.1825782  1.6488386  2.0966445  2.4943799
#>  [622,]  2.5601669  2.7017482  2.7705150  2.8189990
#>  [623,]  4.4912030  5.3152276  6.1206047  6.8778498
#>  [624,]  4.4682781  5.0213616  5.4532449  5.5678502
#>  [625,]  5.0756238  5.3793916  5.4814782  5.5816900
#>  [626,]  2.1608555  2.2138612  2.2656898  2.2974922
#>  [627,]  9.0850153 10.3932118 11.0924499 11.1123968
#>  [628,]  1.6155845  2.1966707  2.7368629  2.9991230
#>  [629,]  2.4784224  3.3526023  4.0378194  4.5624563
#>  [630,]  1.9349622  2.2218983  2.3489521  2.4422564
#>  [631,]  9.8891305 10.6344687 10.9503281 11.0569044
#>  [632,]  2.5258156  3.4118116  4.0207939  4.2885445
#>  [633,]  1.5864978  2.0102516  2.2881420  2.4416040
#>  [634,]  4.4671351  4.7668340  4.9570533  5.0798500
#>  [635,]  3.2845380  4.4886383  5.5102103  5.7761939
#>  [636,]  2.6162129  3.1531271  3.1637157  3.1686943
#>  [637,]  2.6947699  2.9157903  3.0937427  3.1024664
#>  [638,]  6.5698392  7.6641920  8.1937867  8.5300743
#>  [639,]  4.4074383  5.2784563  5.3776134  5.4183227
#>  [640,]  9.7542636 11.4665200 12.3656527 12.4818565
#>  [641,]  4.6207479  5.3142649  5.3745921  5.4189960
#>  [642,]  2.6938365  3.7095229  4.4609338  5.0164644
#>  [643,]  3.8162696  4.0947751  4.3485139  4.4357229
#>  [644,]  3.0958017  3.1782493  3.2326310  3.2605127
#>  [645,]  5.0936803  6.5112457  7.6224927  7.7454961
#>  [646,] 11.4951590 12.9893289 13.4676131 13.6910138
#>  [647,]  4.2717895  4.9712164  5.4693217  5.8890112
#>  [648,]  3.0583111  4.3410026  5.0285498  5.2596129
#>  [649,]  4.5986620  4.8691353  4.9176615  4.9594679
#>  [650,]  4.1941379  5.2768615  5.9453627  6.4513138
#>  [651,]  2.3336788  2.7911388  3.0824782  3.3577898
#>  [652,]  3.3982592  4.3421222  4.8588730  4.8848808
#>  [653,]  1.8479762  2.1950219  2.4197734  2.4704794
#>  [654,]  2.8045484  3.1446868  3.3379750  3.3827297
#>  [655,] 12.8003718 13.5240364 13.6929686 13.8109260
#>  [656,]  7.3389254  7.5898393  7.7090638  7.8133222
#>  [657,]  3.4968548  4.0664916  4.4758083  4.5360561
#>  [658,]  6.3618380  6.8727918  6.9671118  7.0280990
#>  [659,]  3.7010557  3.7884592  3.8711801  3.8859196
#>  [660,]  1.7937897  1.8358347  1.8668166  1.8784958
#>  [661,]  2.0887573  2.1609702  2.2277095  2.2756160
#>  [662,] 12.8525871 16.6685286 17.8586386 18.2935811
#>  [663,]  7.0917646  8.2096913  9.0020365  9.2989339
#>  [664,]  4.3602448  5.5033590  6.5333929  6.6259714
#>  [665,]  2.1434408  2.4022482  2.4354744  2.4517073
#>  [666,]  2.2916229  2.8947225  3.2046299  3.4620158
#>  [667,]  2.4905959  3.4993349  4.4676233  4.5033571
#>  [668,]  4.6923324  5.6768452  5.9617511  6.1045773
#>  [669,]  2.5311857  2.6499923  2.7424584  2.7704570
#>  [670,]  4.1562113  4.4398051  4.6700802  4.6748875
#>  [671,]  5.6005672  5.7566448  5.7709943  5.7746615
#>  [672,]  3.5094510  3.6428799  3.7644986  3.7773782
#>  [673,]  8.5001199  8.9455133  9.3630293  9.3892087
#>  [674,]  1.9551128  2.4657608  2.8250652  2.9656619
#>  [675,]  6.9072241  8.4478021  9.1484831  9.1556428
#>  [676,]  4.8866203  5.5487035  5.8555586  6.0529035
#>  [677,]  2.1881456  2.3325390  2.4031844  2.4428914
#>  [678,]  6.5822231  7.2873501  7.8763977  8.0089178
#>  [679,]  2.5738929  3.4478462  4.0381292  4.5624955
#>  [680,]  3.0547426  4.0540524  4.4405113  4.6179814
#>  [681,]  2.9351559  3.0930856  3.1285257  3.1297046
#>  [682,]  4.9495827  6.2637948  6.6482367  6.8360439
#>  [683,]  3.7276074  4.0126606  4.1491439  4.2286087
#>  [684,]  6.0761289  6.4435064  6.4956546  6.5371334
#>  [685,]  5.2984302  7.0970276  8.0658501  8.4865695
#>  [686,]  2.3868360  2.5433805  2.6039411  2.6213395
#>  [687,]  7.1733109  8.4402163  9.2956792  9.3631816
#>  [688,] 11.9113979 14.4586208 15.2140474 15.5687910
#>  [689,]  5.3395808  6.6144145  7.6240036  8.3824436
#>  [690,]  4.8537279  6.1718395  6.3721426  6.4813492
#>  [691,]  3.7318322  3.9169261  4.0359128  4.0460472
#>  [692,]  2.3146037  2.5017668  2.5314721  2.5593878
#>  [693,]  3.4092193  3.5704436  3.6850862  3.7295869
#>  [694,]  8.3301834  8.9851681  9.2208549  9.2606841
#>  [695,]  2.0921104  2.8036777  3.1177157  3.3392399
#>  [696,]  8.1205517  8.3099047  8.3624904  8.3860531
#>  [697,]  5.0558757  6.2449049  6.3522612  6.3999316
#>  [698,]  3.9919672  4.3982761  4.4355674  4.4358861
#>  [699,]  4.0005426  5.4470264  6.7891620  8.1094198
#>  [700,]  5.9906713  6.5507313  6.8320541  7.1084087
#>  [701,]  9.8169752 10.4423062 10.9366233 11.1724383
#>  [702,]  1.5925480  2.3124642  2.7721847  3.1169671
#>  [703,] 15.0865831 15.8054207 16.4459484 16.8901389
#>  [704,]  6.5834308  7.0478853  7.1169285  7.1416135
#>  [705,]  6.1628176  8.4598424  8.7280533  8.7568955
#>  [706,]  1.9212860  2.1733947  2.3205277  2.3489344
#>  [707,]  5.1762174  6.1454160  6.6442883  7.0002403
#>  [708,]  3.2723784  3.8685313  4.4154305  4.8310113
#>  [709,]  0.6598416  0.8585531  0.9868166  1.0133173
#>  [710,]  3.1105848  3.4255711  3.5736043  3.6903754
#>  [711,]  4.2309082  5.0836374  5.5284134  5.8688830
#>  [712,]  0.8201457  1.0134669  1.1377905  1.2189361
#>  [713,]  6.3512518  8.0926096  8.5075306  8.8813485
#>  [714,]  3.4784447  4.1343868  4.2274608  4.2746612
#>  [715,]  3.6421872  3.8944166  4.0382419  4.1342271
#>  [716,]  4.6492482  5.4033674  5.4788926  5.4922959
#>  [717,]  5.0209573  7.2788239  7.6282525  7.9233700
#>  [718,]  3.9731073  4.4245878  4.7126484  4.9679496
#>  [719,]  4.0930015  5.0936182  5.9867880  6.7749193
#>  [720,]  4.2056312  4.4224816  4.6090230  4.7667907
#>  [721,]  4.3000638  4.8546393  5.2070771  5.5352504
#>  [722,]  2.7011003  3.4596333  4.1816881  4.7296402
#>  [723,] 16.5941836 18.6669954 20.7007000 21.5874651
#>  [724,]  2.4157094  2.7180792  2.7732516  2.8082822
#>  [725,]  1.9254453  2.4186029  2.4761225  2.5201573
#>  [726,]  1.3284659  1.4667507  1.4853637  1.4924276
#>  [727,]  2.4080731  3.0586243  3.5265031  3.8009750
#>  [728,]  5.6213342  6.2718688  6.7407303  6.7846119
#>  [729,]  6.4765144  6.8327107  7.1481802  7.3603007
#>  [730,] 10.7008525 11.3700124 11.5228052 11.6428501
#>  [731,]  4.3273399  4.3682471  4.3795962  4.3891655
#>  [732,]  3.5021086  4.7145072  5.4829286  5.8255583
#>  [733,]  8.6914942 11.6623531 11.8590493 11.8746143
#>  [734,]  4.0559302  4.4728776  4.6914848  4.7934268
#>  [735,]  2.2831118  3.3218415  3.3849875  3.3932324
#>  [736,]  4.0099654  4.7657500  5.4771421  5.6721790
#>  [737,]  6.7811345  7.9782230  8.3759812  8.3925111
#>  [738,]  2.3601449  2.4611767  2.5477712  2.6035208
#>  [739,]  3.0844566  3.9285182  4.5093509  4.5490877
#>  [740,]  7.4279817  8.7591600  9.2640271  9.4102588
#>  [741,]  6.8535773  8.2357014  8.8041331  9.0883204
#>  [742,]  6.1819659  8.1999326  8.6768724  8.9535032
#>  [743,]  3.7516953  4.0816754  4.3620024  4.3836751
#>  [744,]  3.3551672  4.2671904  4.9085155  4.9670547
#>  [745,]  4.7339872  5.4357687  5.8353685  6.0269793
#>  [746,]  2.7614062  3.2474765  3.5391155  3.6498912
#>  [747,]  4.6456085  4.8627858  5.0602670  5.1557594
#>  [748,]  8.0580038  8.3817173  8.6670037  8.8077171
#>  [749,]  3.0133624  3.1663820  3.2852118  3.3888617
#>  [750,]  5.1763886  5.2403402  5.2417756  5.2422946
#>  [751,]  7.8049490  8.6773010  9.4450454  9.6935903
#>  [752,]  1.9884766  2.6195860  2.9540629  3.0297560
#>  [753,] 14.8653190 15.3345662 15.7247487 16.0586532
#>  [754,]  3.3800721  3.5322645  3.5982077  3.6006501
#>  [755,]  6.3677655  7.6922272  7.9113744  7.9136582
#>  [756,]  4.3364877  4.9921591  5.1387046  5.2622553
#>  [757,]  5.0340829  6.3591773  6.7898122  6.8291739
#>  [758,]  1.2065744  1.5512927  1.7938713  1.9279707
#>  [759,]  6.2102521  8.3483589 10.1190208 10.6811600
#>  [760,]  5.0355726  5.4713913  5.8450171  6.0140384
#>  [761,]  6.9930031  7.8317824  8.3285891  8.4496487
#>  [762,]  9.1642930 11.3348251 12.8546271 13.4700189
#>  [763,]  1.5673321  1.6053606  1.6309789  1.6464124
#>  [764,]  4.8410047  6.0352402  6.7007544  6.7902541
#>  [765,]  1.8151669  1.9427443  1.9638799  1.9767109
#>  [766,]  1.8865239  1.9738537  1.9932027  2.0028688
#>  [767,]  2.8319781  4.2137942  4.5809995  4.6181363
#>  [768,]  5.2602210  6.0133677  6.4839068  6.8309301
#>  [769,]  6.6561013  7.8043749  8.2783384  8.7454942
#>  [770,]  2.5871612  2.8518341  2.8955309  2.9342345
#>  [771,]  8.2508591  9.5727975 10.6735723 10.8062179
#>  [772,]  6.5514971  6.8978301  6.9626145  6.9938709
#>  [773,]  4.6973751  5.1047748  5.5089249  5.9118369
#>  [774,]  4.4376159  5.1568425  5.6998584  5.7447501
#>  [775,]  1.2603751  1.8364802  2.0673320  2.2758778
#>  [776,]  2.7510505  3.0129374  3.2163909  3.4182744
#>  [777,]  5.2547636  6.0412890  6.4624005  6.6645463
#>  [778,]  1.0140259  1.3315071  1.5945353  1.7640186
#>  [779,]  5.9817677  6.8846512  7.2204224  7.2630768
#>  [780,]  3.6836241  4.8186846  5.1053850  5.1253570
#>  [781,]  2.2945652  2.6548130  2.9003777  3.1100922
#>  [782,]  2.6580801  2.8825829  3.0582669  3.1795549
#>  [783,]  7.2503495  8.6828056  9.2724093  9.6390190
#>  [784,]  3.2316014  4.1570626  4.6887402  4.7341939
#>  [785,]  1.8172644  2.2826912  2.7003164  2.9934059
#>  [786,]  5.3578647  6.5362919  7.6018451  8.5492746
#>  [787,] 12.4736551 17.4489300 20.3546916 21.3202667
#>  [788,]  3.5556939  3.7431498  3.7713114  3.7726293
#>  [789,]  1.0294588  1.1564579  1.2264502  1.2822502
#>  [790,]  2.5082705  3.2986923  3.9643866  4.4329264
#>  [791,]  5.1083766  6.0909693  6.9781042  7.4886941
#>  [792,]  3.3874998  3.9065750  4.0800694  4.1397639
#>  [793,]  5.5726989  5.6855186  5.7565110  5.7899892
#>  [794,]  3.7110825  4.2316627  4.2714794  4.2723404
#>  [795,]  4.9800939  6.6938505  7.5543619  7.7687919
#>  [796,]  7.0465199  7.4725011  7.8968429  7.9122663
#>  [797,]  7.9237018 10.1813921 10.4386588 10.6509744
#>  [798,]  5.7413683  5.9606543  6.1005869  6.1403438
#>  [799,]  6.6899017  7.2729272  7.4282820  7.5087399
#>  [800,]  4.7491782  5.3681986  5.5045045  5.5211708
#>  [801,]  1.1503822  1.3160707  1.4646572  1.5927647
#>  [802,]  6.9965940  9.8334057 10.8025525 10.9782044
#>  [803,]  2.9118228  2.9922550  3.0060640  3.0138368
#>  [804,]  0.9402944  1.2384339  1.5356024  1.7430173
#>  [805,]  5.2829374  5.6677737  5.9807516  6.2779960
#>  [806,]  2.5078047  3.6045609  4.4283638  5.2118432
#>  [807,]  4.0224654  4.4841377  4.9328543  4.9837481
#>  [808,]  4.2116867  4.9619351  5.3649696  5.6751248
#>  [809,]  7.0226881  7.3894506  7.7001006  7.7664941
#>  [810,]  1.1384161  1.4500978  1.6080237  1.6092850
#>  [811,]  1.2982499  1.4639667  1.5907065  1.6201945
#>  [812,]  8.9417967 10.9446785 12.0498784 12.3535340
#>  [813,]  2.3036484  2.4721105  2.5175251  2.5346341
#>  [814,]  6.7325432  6.9717698  7.2086187  7.2387429
#>  [815,]  7.5923984 10.5288214 12.2449172 12.7283554
#>  [816,]  3.1235469  3.2576130  3.2995671  3.3216558
#>  [817,]  5.0656404  5.9180552  6.2572653  6.3827976
#>  [818,] 14.4344993 14.4614161 14.4752176 14.4837082
#>  [819,]  4.3857833  5.5112020  5.5504744  5.5652812
#>  [820,]  4.8995561  5.4741834  5.6943166  5.8122575
#>  [821,]  2.4045093  2.5279615  2.6150173  2.6705906
#>  [822,]  2.5744322  3.7131911  4.1689481  4.5614629
#>  [823,]  1.5533395  2.1609099  2.6997487  2.9547566
#>  [824,]  4.4565610  4.9218984  5.0625380  5.1624669
#>  [825,]  6.4643868  7.6943863  8.6927831  9.0165787
#>  [826,]  4.1917078  4.9572675  5.0915111  5.1178299
#>  [827,]  5.5274055  6.6962179  7.0117403  7.1215661
#>  [828,]  1.3394770  1.5269470  1.5942843  1.6462712
#>  [829,]  5.5293957  6.4219463  6.8445426  6.9618303
#>  [830,]  7.0409967  7.4626375  7.8343421  8.1701876
#>  [831,]  4.0643975  5.4152632  6.6084889  7.3188758
#>  [832,]  6.0082595  7.0088067  7.9135085  8.3654387
#>  [833,]  1.2363712  1.4365378  1.5951489  1.6402014
#>  [834,]  7.1534952  8.8982531  9.0916017  9.2024453
#>  [835,]  4.0465137  5.1010682  5.9550179  6.5899252
#>  [836,]  3.8720641  5.0491455  5.7702954  6.0048421
#>  [837,]  9.0013866  9.8509270 10.5523371 10.9544147
#>  [838,]  2.5243511  3.2980534  3.6281271  3.6783055
#>  [839,]  4.2384377  5.1313435  5.1866819  5.1911678
#>  [840,]  2.8462012  3.2068314  3.5433369  3.6019736
#>  [841,]  7.6030986  8.0006022  8.0227256  8.0357614
#>  [842,]  3.5560195  4.0123680  4.1872085  4.2535138
#>  [843,]  3.4364426  3.8281613  4.1038258  4.1490835
#>  [844,]  4.7608725  5.0926991  5.2481140  5.2931412
#>  [845,]  8.2292173 10.5680136 11.4502414 11.5284431
#>  [846,] 11.5153186 12.2455405 12.6113809 12.7714220
#>  [847,]  1.8511287  2.5072808  3.0405906  3.0611538
#>  [848,]  1.9668023  2.2028862  2.3504913  2.4205157
#>  [849,]  4.1881429  5.3878357  6.1884932  6.5087165
#>  [850,]  3.7944416  4.6880035  5.2862142  5.8032976
#>  [851,]  1.6082849  1.7686333  1.9109642  1.9847693
#>  [852,]  1.8068204  2.2174189  2.5968631  2.7794908
#>  [853,]  3.1340176  3.3096523  3.3780117  3.4418395
#>  [854,]  9.3424582  9.9293129 10.3602303 10.4641646
#>  [855,]  1.4225724  1.6867341  1.9359449  2.1335674
#>  [856,]  6.3102597  7.3672232  8.4240352  8.8341404
#>  [857,]  3.5284431  3.6669630  3.7282016  3.7860890
#>  [858,]  4.3967268  6.4003431  6.5108154  6.5707374
#>  [859,]  7.4924877  8.0078934  8.5047097  8.6910812
#>  [860,]  1.9122885  2.4106836  2.6445629  2.6495521
#>  [861,]  5.0854811  6.9187212  7.0453563  7.1351962
#>  [862,]  3.7332590  4.0736808  4.2487827  4.2567661
#>  [863,]  2.8490797  3.6983436  4.3702400  4.9790302
#>  [864,]  5.1437623  5.7427255  6.3303452  6.4684065
#>  [865,]  1.4682210  1.5886352  1.6724564  1.6961397
#>  [866,]  5.8327921  6.7696257  7.0413099  7.0439887
#>  [867,]  3.6594321  3.8928530  4.0971824  4.1748197
#>  [868,]  4.8878725  5.1045643  5.1286878  5.1467692
#>  [869,]  2.7430728  3.4489192  4.0019865  4.3964118
#>  [870,]  8.3899277 10.3336986 11.1632395 11.4115289
#>  [871,]  6.1103578  6.8191136  7.4725025  7.7461997
#>  [872,]  4.9609367  6.2083506  6.9133992  7.0103234
#>  [873,]  1.2967950  1.3586863  1.3985237  1.4174092
#>  [874,]  1.8671978  2.4075138  2.5787325  2.6767919
#>  [875,]  3.3526532  3.8357259  3.9495284  4.0109738
#>  [876,]  2.5660030  3.2282154  3.5145616  3.6323763
#>  [877,]  5.5635777  7.5043819  8.3519143  8.7709939
#>  [878,]  2.5375142  2.9787417  3.0464862  3.0839323
#>  [879,]  2.1182466  2.3307643  2.4259558  2.4449660
#>  [880,]  1.4381232  1.8355928  2.1645423  2.2454611
#>  [881,]  3.2169515  4.0855034  4.5372305  4.5823423
#>  [882,]  1.5367616  1.7568695  1.7870909  1.7959302
#>  [883,]  5.5142699  5.8328799  5.9678304  6.0126084
#>  [884,]  2.0818576  3.0275256  3.5570617  3.7201448
#>  [885,]  3.5622538  3.9732777  4.3023269  4.4429965
#>  [886,]  3.4942974  3.6481913  3.7679463  3.8086351
#>  [887,]  5.8115805  6.8090327  7.1637875  7.4393950
#>  [888,] 10.2495654 12.2195752 12.6457475 12.6858133
#>  [889,]  4.1679256  5.5501843  6.2424721  6.4011342
#>  [890,]  2.9237833  4.2111530  5.3675927  5.3752515
#>  [891,]  3.7094333  5.4020764  5.5465232  5.5516844
#>  [892,]  6.1790502  6.9398753  7.1375067  7.2677389
#>  [893,]  3.5851930  3.7768763  3.8903285  3.9400039
#>  [894,]  5.3722970  6.0878945  6.5855334  7.0250214
#>  [895,]  8.8137865 11.2023888 12.2229761 12.8613424
#>  [896,]  5.9992063  7.6239535  9.0721536  9.3695033
#>  [897,]  1.1930715  1.5201569  1.8176283  1.8231887
#>  [898,]  4.2383630  4.4954932  4.7442078  4.9224199
#>  [899,]  3.9535914  4.2188184  4.2277922  4.2306567
#>  [900,]  0.6888814  0.9628157  1.0964540  1.1145844
#>  [901,]  6.6258352  7.1484722  7.5604937  7.6056071
#>  [902,]  2.9947440  3.4898890  3.7694747  3.9239480
#>  [903,]  1.3674752  1.7690815  2.0895238  2.3742520
#>  [904,]  4.7510614  5.2476308  5.4981577  5.5808792
#>  [905,]  6.2890497  7.2152085  7.9413780  8.4289459
#>  [906,]  8.7742201 10.4342525 12.0183885 12.9261142
#>  [907,]  8.7862005  8.9088765  9.0151435  9.1000640
#>  [908,]  3.8848110  4.1525819  4.2463509  4.3076510
#>  [909,]  5.1448202  5.4001611  5.4749852  5.5248190
#>  [910,]  2.6257504  3.6454211  4.5296656  5.3952833
#>  [911,]  2.2213238  2.4006240  2.4499994  2.4883512
#>  [912,]  3.9965620  5.6535325  6.8559390  7.3913930
#>  [913,]  8.0773029  9.6994077  9.8602796  9.9870296
#>  [914,]  5.3622876  7.8177879  8.1454233  8.4339908
#>  [915,]  1.9178360  2.6294376  3.1813751  3.6033075
#>  [916,]  4.3838478  5.2052930  5.6423366  5.7457194
#>  [917,]  1.2337217  1.6185045  1.7830566  1.8031080
#>  [918,] 11.0976605 11.5353543 11.7557786 11.7587435
#>  [919,]  3.6508440  3.9618862  4.1981931  4.2407514
#>  [920,]  5.7847350  6.2903676  6.6154391  6.6593034
#>  [921,]  3.1050253  4.3770468  5.1683734  5.6650065
#>  [922,]  4.7849280  5.3747829  5.8462296  6.2259565
#>  [923,]  6.1609994  9.0877672 10.2320864 10.9633069
#>  [924,]  2.3635006  3.0075783  3.2509337  3.2955016
#>  [925,]  3.1762385  4.1108147  5.0195860  5.1391178
#>  [926,]  3.1598532  3.9072880  4.4740278  4.8895253
#>  [927,]  2.2311365  2.5323522  2.6895941  2.7245306
#>  [928,]  4.2798522  6.1432637  6.6292995  6.7296484
#>  [929,]  3.6233170  4.4466462  4.9649196  4.9705762
#>  [930,]  5.0791808  5.6469701  5.8076804  5.8381814
#>  [931,]  1.8346775  2.3319804  2.7324215  2.9553586
#>  [932,]  2.6570468  3.1009865  3.2629751  3.3045681
#>  [933,]  2.0797469  2.5025918  2.5123597  2.5181812
#>  [934,]  4.2011929  4.8764019  5.3798199  5.8534245
#>  [935,]  3.8133035  4.0678019  4.2014756  4.2374053
#>  [936,]  4.7860732  6.0107851  6.4783234  6.5880510
#>  [937,]  5.1145507  6.4854199  6.5534911  6.5849008
#>  [938,]  0.8573113  0.9678286  1.0743206  1.1567966
#>  [939,]  8.6334202 10.1187188 10.4103550 10.4723930
#>  [940,]  2.6574307  2.9951214  3.3251562  3.5167372
#>  [941,]  2.0767446  2.5369741  2.9464827  3.1857817
#>  [942,]  2.4506854  3.1521032  3.4969145  3.5135717
#>  [943,]  5.2745038  7.0704743  7.5722612  7.9658315
#>  [944,]  5.0257244  6.6693522  7.2818217  7.6297839
#>  [945,]  3.1954053  3.4399989  3.6430735  3.7574821
#>  [946,]  4.5546275  5.1409582  5.2545639  5.3172425
#>  [947,]  2.4876496  3.4843257  3.7141751  3.8879814
#>  [948,] 14.5904936 16.0089604 16.9380439 17.0234094
#>  [949,]  8.4019325  9.9286758 11.3124917 11.9721160
#>  [950,]  8.8258080 10.0252675 10.6660847 10.7191492
#>  [951,] 15.4834231 16.3576532 16.5007842 16.6268940
#>  [952,]  6.2902610  6.7790258  7.1253701  7.3306793
#>  [953,]  1.5548321  1.8238936  1.9360545  2.0315369
#>  [954,]  2.3893117  2.4980651  2.5579959  2.5801915
#>  [955,]  3.4884162  4.2782150  4.7452666  4.8071550
#>  [956,]  5.4185715  6.9077980  7.4288468  7.6635578
#>  [957,]  2.0340544  2.3355764  2.5363997  2.5709112
#>  [958,]  3.3857667  3.9340471  4.4489659  4.5072367
#>  [959,]  3.5486833  4.6988300  5.0387651  5.1201825
#>  [960,]  1.9611727  2.3143043  2.4733436  2.6096902
#>  [961,]  5.9874710  7.2594988  7.7195259  8.1200393
#>  [962,]  8.2133418  9.6381889 10.2367171 10.4105903
#>  [963,]  4.1597083  4.3135206  4.4274268  4.4679953
#>  [964,]  7.4025172 10.3494850 11.7268595 11.8059581
#>  [965,]  5.0435089  5.5568527  5.8018165  5.8030308
#>  [966,]  2.6155587  3.3201651  3.5371576  3.6685596
#>  [967,]  3.5665114  3.8450577  4.0537128  4.1587819
#>  [968,]  0.9580065  1.0961646  1.2088770  1.3081549
#>  [969,]  4.7980026  5.5751738  5.6098776  5.6414587
#>  [970,]  3.8782756  4.8329318  5.6112019  6.2512720
#>  [971,]  0.9379038  1.1862373  1.2800091  1.3004213
#>  [972,]  4.7634749  6.0829649  7.2094817  8.2118275
#>  [973,]  7.1536220  7.2678280  7.2908086  7.3054954
#>  [974,]  1.4770418  1.4988145  1.5082838  1.5110828
#>  [975,]  6.2599773  6.7551187  7.0179464  7.1989585
#>  [976,]  9.6469251 10.8794233 11.9008166 12.1460392
#>  [977,]  0.9159788  1.0807840  1.1292165  1.1549449
#>  [978,]  3.4693231  4.0318038  4.4906607  4.6171074
#>  [979,]  5.7247272  6.3422782  6.8269957  6.9098472
#>  [980,]  4.5941244  5.6734088  6.7059647  6.8663091
#>  [981,]  4.0669233  5.1827187  5.4728160  5.6953509
#>  [982,]  6.0737096  7.1711892  7.8871034  8.1797660
#>  [983,]  5.7847740  6.6499803  6.9936359  7.2828020
#>  [984,]  2.4186286  2.9373634  3.4053564  3.4056941
#>  [985,]  3.5822848  3.8762585  3.9098300  3.9392922
#>  [986,]  1.7320587  2.0291589  2.0409847  2.0495653
#>  [987,]  2.9600113  3.5461579  3.5944980  3.6192392
#>  [988,]  5.6372541  6.6394347  7.4891662  8.2581640
#>  [989,]  5.2521947  5.6483553  5.9829871  6.2618325
#>  [990,]  8.6724740  8.8983914  8.9966951  9.0862530
#>  [991,]  7.1162481  8.0141826  8.2854784  8.4457826
#>  [992,]  5.5559329  7.0674190  7.8946032  8.5835776
#>  [993,]  5.6799750  6.5578333  7.3882103  7.9014211
#>  [994,]  4.2721500  5.4701338  6.0172312  6.0427205
#>  [995,]  3.1802895  3.9778095  4.0867219  4.1113894
#>  [996,]  3.4194953  4.4050949  4.8495848  5.0027666
#>  [997,]  0.5120640  0.6819694  0.8005691  0.8050445
#>  [998,]  8.0705871  8.5772327  8.7334788  8.8168045
#>  [999,]  4.7183076  5.7031680  6.0484048  6.1043289
#> [1000,]  5.2027722  7.5406220  9.4151313 11.0005347