generate_null_distribution_T3

This function generate the null distribution of T_eta, should be run before calculating TWT. Please see Setp 2 in our manuscript for more information.

generate_null_distribution_T3(m, n, cov_mat, cutoff_value)

Arguments

m

Dimension.

n

Number of the generated sample size.

cov_mat

Covariance matrix for generating random samples.

cutoff_value

Truncated value, should be in 0 to 1 and must contain 1.

Value

Null distribution matrix with number of rows equal n and number of columns equal length of cutoff value minus 1.

Examples

ThreeWayTest::generate_null_distribution_T3(m=6,n=1000,
cov_mat=diag(nrow = 6, ncol = 6), cutoff_value=c(0.2,0.4,0.6,0.8,1))
#>               [,1]       [,2]       [,3]       [,4]
#>    [1,]  1.5984428  1.8113009  1.9515769  2.0459832
#>    [2,]  2.7431723  2.9568985  3.1416183  3.2820054
#>    [3,]  6.5692355  6.7661804  6.8978421  6.9678456
#>    [4,]  3.7740187  4.1180480  4.3248321  4.5305267
#>    [5,]  3.6751767  4.4173765  4.7422839  4.8380333
#>    [6,]  2.2051256  2.4654744  2.6833526  2.8830323
#>    [7,]  4.1689162  4.5156638  4.5452554  4.5541944
#>    [8,]  4.5940440  5.7640069  6.6567783  6.7656963
#>    [9,]  5.4828728  7.5488362  8.1051731  8.3107124
#>   [10,]  4.0785295  4.4006016  4.6075064  4.7963680
#>   [11,]  4.5083080  4.6797581  4.7752750  4.7915501
#>   [12,]  2.9148322  3.6152121  4.1697070  4.5333120
#>   [13,]  2.1542069  2.5011267  2.7136528  2.8794101
#>   [14,]  5.0587592  6.4591406  6.7723723  6.9879964
#>   [15,]  3.0509508  3.3208373  3.3759017  3.4023940
#>   [16,]  3.5486937  4.1662510  4.7345759  4.9628252
#>   [17,]  6.8313360  6.9707322  7.0607072  7.0748814
#>   [18,]  9.1946350 10.8979014 11.5195075 12.0119107
#>   [19,]  4.8240465  5.7226697  6.1379122  6.5509226
#>   [20,]  4.1856972  5.5984883  6.1442245  6.4846689
#>   [21,]  0.9467929  1.2893620  1.4814930  1.6209073
#>   [22,]  1.4343908  1.6489961  1.8040342  1.9263911
#>   [23,]  4.5343892  5.6126998  6.3273500  6.4690259
#>   [24,]  0.9789252  1.3768723  1.5708153  1.6760070
#>   [25,]  6.1126175  6.7096026  7.2378084  7.2679864
#>   [26,]  1.5328613  2.0221650  2.3521862  2.6428283
#>   [27,]  1.4504295  1.7963113  2.1286542  2.3209083
#>   [28,]  7.7200827  8.2585214  8.4755234  8.5067048
#>   [29,]  1.3884685  1.7973088  1.8255296  1.8502959
#>   [30,] 10.6932630 10.8518665 10.9398225 11.0169080
#>   [31,]  2.5273960  3.0948210  3.3310912  3.5422722
#>   [32,]  3.1542789  3.6615322  4.1650406  4.3439187
#>   [33,]  0.5863441  0.6748768  0.7568965  0.8170168
#>   [34,]  6.1551288  6.6889322  6.8474681  6.8818519
#>   [35,]  4.9400529  6.5650619  6.6886649  6.7572903
#>   [36,]  3.3089779  3.9100610  4.1315119  4.2637565
#>   [37,]  5.5578035  7.5491777  9.1368039  9.4855141
#>   [38,]  2.1454773  2.2145002  2.2566101  2.2950167
#>   [39,]  3.2367223  4.5702851  5.0191127  5.0633826
#>   [40,]  2.0136517  2.7817938  2.9209743  2.9225255
#>   [41,]  8.3631000  8.8406086  9.2723703  9.4387087
#>   [42,]  7.1042403  8.8803658  9.1176759  9.1338383
#>   [43,]  3.6375342  4.7103749  5.1287446  5.2861715
#>   [44,]  4.8416007  5.7534176  6.5277398  6.6630524
#>   [45,]  4.6154318  6.0065908  7.2335931  8.3314227
#>   [46,]  2.4528875  2.7382730  3.0009184  3.1211855
#>   [47,]  5.2525511  6.1534025  6.3692568  6.3715640
#>   [48,]  8.2339029  9.4714272  9.7019206  9.7354483
#>   [49,]  2.9533356  3.4669834  3.5609176  3.5678313
#>   [50,] 16.0484273 17.1317452 17.8325307 17.8966930
#>   [51,]  4.3476600  5.0538629  5.6606681  5.9082082
#>   [52,]  9.8876248 11.6134039 12.1635146 12.3949672
#>   [53,]  1.4407384  1.5078658  1.5596309  1.6020991
#>   [54,]  6.6490203  8.0167157  8.2338431  8.4094290
#>   [55,]  2.9244500  3.1476921  3.2074573  3.2641880
#>   [56,]  3.0872530  3.1787231  3.2686286  3.2752743
#>   [57,]  2.6078891  3.7080495  4.3854396  5.0252069
#>   [58,]  8.2183564 10.4628907 10.8227655 10.8365233
#>   [59,]  5.0481973  5.6924381  6.0242597  6.0632081
#>   [60,]  4.2181730  5.6157930  6.5267568  6.8564309
#>   [61,]  2.1714927  2.2297173  2.2854155  2.3010295
#>   [62,]  6.8064541  7.4877814  7.7483771  7.9956759
#>   [63,]  6.8830530  8.9781637  9.9031904 10.3750519
#>   [64,]  3.8240261  4.1812022  4.5332043  4.6828159
#>   [65,]  7.1038947  8.3259574  8.8073548  8.9686976
#>   [66,]  6.3387040  7.5582664  8.1750406  8.6836764
#>   [67,]  2.7076914  3.4197959  3.8247558  4.1920812
#>   [68,]  3.7435852  4.2389295  4.5446493  4.5558107
#>   [69,]  7.5885316 10.4668359 10.9778114 11.0220593
#>   [70,]  1.3085387  1.3886893  1.4290212  1.4690646
#>   [71,]  2.7785710  3.2110836  3.3687329  3.3939217
#>   [72,]  3.1224620  3.2255984  3.2613699  3.2881513
#>   [73,]  3.3367257  4.0843985  4.4001733  4.4213148
#>   [74,]  3.9232766  4.6678068  5.1628928  5.5976375
#>   [75,]  4.1372931  5.1956548  5.5610662  5.6906299
#>   [76,]  1.2008080  1.3285207  1.3534364  1.3569930
#>   [77,]  3.6752274  3.9287067  4.1258972  4.2357166
#>   [78,]  2.2118781  2.7964137  3.0012892  3.0949081
#>   [79,]  1.0921836  1.3381814  1.4894080  1.6253055
#>   [80,]  1.5068323  1.7190200  1.8905096  1.9646929
#>   [81,]  2.8523401  3.2253211  3.3707485  3.3726994
#>   [82,] 13.9181596 14.4565990 14.6071330 14.6282474
#>   [83,]  6.0547662  7.0403329  7.9043080  8.7505943
#>   [84,]  2.9114815  3.3533259  3.7115289  3.7570306
#>   [85,]  4.1383347  5.7372480  7.1523557  7.5663929
#>   [86,]  2.9123680  3.8818003  4.3892512  4.8579202
#>   [87,]  1.6804336  2.3324017  2.5076036  2.5531357
#>   [88,]  6.9578149  7.8786918  8.0854711  8.1373081
#>   [89,]  4.2951924  5.2547007  5.8416733  5.8836567
#>   [90,]  7.4699966  8.0172187  8.0445293  8.0577427
#>   [91,]  3.6138859  4.9064597  5.1044814  5.2849445
#>   [92,]  7.2235380  8.0652499  8.2552079  8.3998687
#>   [93,]  4.4784284  5.8391171  6.8918668  6.9526393
#>   [94,]  4.7582361  4.8536884  4.8761364  4.8926848
#>   [95,]  3.0829703  4.0689158  4.2686773  4.4070621
#>   [96,]  4.2784924  5.4488846  5.8341255  6.0665717
#>   [97,]  8.2423824  8.9127012  9.3173044  9.3226981
#>   [98,]  1.4435665  1.7504811  1.9831170  2.0062620
#>   [99,]  3.6784630  4.0786219  4.1748441  4.2532428
#>  [100,]  1.5666418  1.9193926  2.1438035  2.2194980
#>  [101,]  8.5033463 10.7795245 11.3292512 11.3518160
#>  [102,]  7.1096362  8.5488310  9.2099339  9.4236105
#>  [103,]  4.5781936  5.5496756  5.9024938  6.0279736
#>  [104,]  5.2545137  6.0291372  6.7474239  6.9974162
#>  [105,]  7.8635730  8.9404697  9.9789626 10.8826552
#>  [106,]  3.1997504  3.8974781  4.3845910  4.4186319
#>  [107,]  7.8349335  8.7368882  9.3140827  9.3811544
#>  [108,]  1.6980118  1.8169512  1.9288900  1.9777920
#>  [109,] 10.9353176 12.2857158 13.5518185 14.3865367
#>  [110,]  1.4342879  1.9183312  2.3287266  2.7378362
#>  [111,]  4.8080824  5.1315542  5.4256406  5.6110966
#>  [112,]  9.7662882 10.9058432 11.2477126 11.3576772
#>  [113,]  6.5568563  6.6383320  6.7128733  6.7721039
#>  [114,]  7.5952941  8.0916948  8.1401876  8.1492750
#>  [115,]  4.6552933  5.2722367  5.5067176  5.6037020
#>  [116,]  2.0125511  2.4465837  2.5454842  2.5807096
#>  [117,]  4.3882394  4.7221907  5.0376355  5.2000875
#>  [118,]  5.5896476  7.9835339  9.5846423  9.9339560
#>  [119,]  4.3145114  5.2841045  6.0339995  6.2859938
#>  [120,]  2.4289063  2.5508017  2.6526994  2.7228337
#>  [121,]  5.5371111  5.8760410  5.9735467  5.9833766
#>  [122,]  4.2872106  4.9539519  5.3377564  5.6016681
#>  [123,]  2.6697428  2.7504610  2.7719575  2.7836225
#>  [124,]  8.4927851  9.1232497  9.6240089  9.9397351
#>  [125,]  2.9644492  3.9913351  4.8535134  5.0757686
#>  [126,]  1.6517547  1.9713272  2.1608395  2.1913268
#>  [127,]  2.4355756  2.5581472  2.6617855  2.7145291
#>  [128,]  1.9470783  2.8104748  3.2173367  3.3084503
#>  [129,]  7.8949586 10.5144641 11.3229439 11.9548118
#>  [130,]  4.4829170  5.3025545  5.9563469  6.3053460
#>  [131,]  1.3896776  1.5042027  1.5819379  1.6397871
#>  [132,]  7.9618878 10.2803634 11.3564632 11.6629331
#>  [133,]  5.4491955  5.7161252  5.9805994  5.9809848
#>  [134,]  5.1304906  7.4888234  8.2787508  8.4682596
#>  [135,]  7.2464000  9.3070530  9.4098391  9.4885318
#>  [136,]  6.2251774  6.5426053  6.7192240  6.8798153
#>  [137,]  3.1136981  4.1705350  4.6748520  5.0549460
#>  [138,]  6.9856847  8.7240772  8.9508660  9.0997236
#>  [139,]  4.7891725  5.7185933  6.3451401  6.6001380
#>  [140,]  1.8149514  2.0582600  2.1569125  2.1895175
#>  [141,]  2.5113376  2.8014968  3.0144191  3.1557951
#>  [142,]  3.2470468  4.6621491  5.4854055  5.8179430
#>  [143,]  4.7757069  6.0830967  7.1493488  7.9717399
#>  [144,]  8.6341291  9.4259487 10.0735836 10.2886293
#>  [145,]  3.1079065  3.4189063  3.5811404  3.6494383
#>  [146,] 14.9252071 16.9997448 18.6596524 19.0109914
#>  [147,]  3.2757666  4.0007220  4.6661148  5.0043035
#>  [148,]  6.4558628  7.8285435  8.9760268  9.1947856
#>  [149,]  2.7206575  3.3242795  3.3515610  3.3657226
#>  [150,]  2.2162252  2.7739499  2.9156642  3.0387089
#>  [151,]  9.5383715 11.0361445 11.9892538 12.4430520
#>  [152,]  3.6196026  4.6178368  5.0502772  5.0552378
#>  [153,]  1.1533421  1.4139943  1.6701509  1.7002197
#>  [154,]  8.0539000  9.0554755  9.9837753 10.2725334
#>  [155,]  4.6876705  6.2694355  6.5535579  6.6276423
#>  [156,]  4.6485974  5.8153047  6.3839204  6.5188244
#>  [157,]  3.5716504  4.5664252  5.3206718  5.9020911
#>  [158,]  9.0078513 10.4598049 11.1707779 11.7103937
#>  [159,]  3.8892657  4.6263840  4.8982898  5.0010113
#>  [160,]  2.7273817  3.9789335  5.0220344  5.4501370
#>  [161,]  7.0310918  7.2525640  7.2889008  7.2897027
#>  [162,]  3.5797997  4.6003726  4.7337006  4.8460344
#>  [163,] 11.0738038 12.1957723 12.4147041 12.6258146
#>  [164,]  4.2498402  4.5154312  4.6831997  4.6997678
#>  [165,]  2.3461326  2.9252640  3.2635908  3.4532090
#>  [166,]  6.2837676  7.2426044  7.7767030  7.9710937
#>  [167,] 11.1632453 12.6114524 13.7432095 14.4348509
#>  [168,]  1.0341176  1.1287564  1.2222372  1.2603117
#>  [169,]  3.7982495  5.3853692  6.0605146  6.1822694
#>  [170,]  2.6803806  2.9531093  3.1086458  3.1956621
#>  [171,]  3.5544409  4.3815720  4.5266271  4.6229439
#>  [172,]  2.4597256  2.5014384  2.5317057  2.5600866
#>  [173,]  3.8645883  4.8633639  5.5759297  5.6317915
#>  [174,]  9.0595102  9.9580483 10.3974772 10.6096935
#>  [175,]  0.7702262  0.9238157  0.9868901  1.0134771
#>  [176,]  5.9463964  6.6117215  6.9225815  7.2147966
#>  [177,]  1.8158624  2.1290833  2.2941112  2.3129901
#>  [178,] 11.1158508 13.2278346 14.3719520 14.5863247
#>  [179,]  1.6654660  2.1799077  2.6437131  2.9590948
#>  [180,]  2.0015627  2.7828468  3.3689913  3.4484124
#>  [181,]  1.7107663  1.7667552  1.7896674  1.7972683
#>  [182,]  3.9496392  4.6859531  4.9326195  5.1786825
#>  [183,]  1.9639010  2.2908340  2.3378683  2.3662716
#>  [184,]  3.8133398  4.1041952  4.1392984  4.1614304
#>  [185,]  5.7565990  6.3684833  6.7351700  6.7602796
#>  [186,]  4.1168606  5.4358213  6.6019377  6.8453400
#>  [187,] 10.1347247 12.0550546 13.2259380 13.2637394
#>  [188,]  6.7494948  8.2749185  8.3540522  8.4294986
#>  [189,]  6.3724578  7.1104182  7.2850323  7.4409523
#>  [190,]  4.1670537  4.9565723  5.4406312  5.7721913
#>  [191,]  5.8537380  7.7940429  8.8612508  9.3960532
#>  [192,]  4.1198297  5.8024354  6.6849704  6.6874980
#>  [193,]  2.9726758  3.7215807  4.3408323  4.8640513
#>  [194,]  6.0306433  7.3982805  8.0203169  8.3277958
#>  [195,]  5.3778063  6.3046306  6.6626091  7.0144473
#>  [196,]  6.2479657  6.6718607  6.9590567  7.1789017
#>  [197,]  7.5331068  7.9397669  8.2710541  8.5997021
#>  [198,]  2.2929242  2.6819682  3.0011815  3.2858036
#>  [199,]  1.5985613  2.0502278  2.4325060  2.6752957
#>  [200,]  1.6214544  1.7411351  1.7467536  1.7492203
#>  [201,]  2.5747286  3.4670361  3.5611067  3.6288868
#>  [202,]  1.6618733  2.1619527  2.3704337  2.3769042
#>  [203,]  6.7943822  7.8355536  8.0890892  8.3301617
#>  [204,]  1.6875454  2.3790070  2.7887296  3.1773340
#>  [205,]  1.8170103  2.0683325  2.0848474  2.0940162
#>  [206,]  6.5760938  7.6454649  8.4636885  8.5324664
#>  [207,]  3.4197945  4.7330669  5.8152227  6.0085851
#>  [208,]  0.3331253  0.4168665  0.4683939  0.4833088
#>  [209,]  2.8241931  4.0478128  4.2276154  4.2935736
#>  [210,]  4.8583053  5.7079054  6.1882810  6.2392750
#>  [211,]  1.8169510  2.1004094  2.3255581  2.3849855
#>  [212,]  4.0918867  4.4266275  4.6050202  4.6155954
#>  [213,]  2.5953290  3.3456928  3.8282324  4.1638409
#>  [214,]  2.5285846  3.2863469  3.8451034  4.1655138
#>  [215,] 10.8553618 12.2378682 12.9494709 13.4795481
#>  [216,]  5.9937386  6.5008110  6.8464792  7.1112000
#>  [217,]  9.0086457 10.5820332 11.9767946 12.4279516
#>  [218,] 14.0922474 14.7468041 15.1352036 15.3905275
#>  [219,]  5.7103776  5.9307851  6.1045062  6.2401132
#>  [220,]  6.0028491  6.9606548  7.8109817  7.9497511
#>  [221,]  2.5154538  3.3661260  3.5186237  3.6554228
#>  [222,]  3.5516460  3.9172479  4.1853211  4.2942784
#>  [223,]  3.2538180  3.6293462  3.9862344  4.1343041
#>  [224,]  3.5153912  4.1993073  4.7839830  5.2824791
#>  [225,]  6.0019771  7.0004793  7.8668409  8.5994778
#>  [226,]  3.5559557  3.8766491  4.1478111  4.2717200
#>  [227,]  6.4995251  7.1175859  7.4096538  7.4719125
#>  [228,]  3.3938286  3.6474387  3.8184055  3.8476759
#>  [229,]  1.6989366  2.1935497  2.5780434  2.9318041
#>  [230,]  2.1301802  2.9286418  2.9690142  2.9730664
#>  [231,]  4.3516647  5.8456290  5.8895939  5.9202593
#>  [232,]  2.4989992  3.0501957  3.2438680  3.2966323
#>  [233,]  9.6889298 10.1235190 10.5361431 10.6871746
#>  [234,]  3.4217250  4.4930081  5.0900721  5.3907346
#>  [235,]  7.2236182  8.5315519  9.7293192  9.9999441
#>  [236,] 12.1535978 12.2663117 12.3654139 12.3748933
#>  [237,]  6.0754750  6.2842876  6.4220118  6.5447736
#>  [238,]  5.8521045  7.0787685  7.6958515  8.2300674
#>  [239,]  3.2864765  4.4348288  5.5593993  6.3378099
#>  [240,]  1.6605018  1.7183996  1.7420439  1.7535811
#>  [241,] 12.4649489 12.9824661 13.3545906 13.3833585
#>  [242,]  2.9290881  3.2970148  3.6552733  3.7149182
#>  [243,]  4.9302320  5.6881320  6.1717372  6.3557797
#>  [244,]  1.5782631  2.1880157  2.4863615  2.5087154
#>  [245,]  3.5481802  3.8806453  3.9925137  4.0564493
#>  [246,]  2.6894529  3.0219874  3.2972980  3.5361551
#>  [247,]  4.9954617  6.7915507  8.2336553  8.4193240
#>  [248,]  6.7573250  8.7294001 10.3028133 11.2164611
#>  [249,]  5.0960897  6.2990643  6.9959062  7.1385094
#>  [250,]  1.6455080  1.8975487  2.1358711  2.3574706
#>  [251,]  8.7424959  9.4607233  9.8622404 10.1413192
#>  [252,]  8.5282925  9.1632425  9.3561331  9.4171354
#>  [253,] 15.0578054 16.3415943 17.5500804 17.9694544
#>  [254,]  1.5303286  2.2254442  2.7966445  2.9131800
#>  [255,] 11.4268804 13.9994641 14.3966901 14.6551027
#>  [256,]  4.1968431  5.2311575  5.3027925  5.3321138
#>  [257,]  2.5750129  3.2431970  3.5734372  3.5856270
#>  [258,]  7.8070886  8.5547912  9.0312604  9.0656378
#>  [259,]  1.7938262  2.0101097  2.1340164  2.1397635
#>  [260,]  4.2008575  5.6773546  6.4470027  6.8745055
#>  [261,]  2.3467086  2.9592172  3.0727111  3.0980281
#>  [262,]  1.5592427  2.1028214  2.4296865  2.5856237
#>  [263,]  4.5258860  6.4287410  6.9399921  7.2679985
#>  [264,]  1.1455640  1.2551934  1.3442501  1.3574857
#>  [265,]  5.0942783  6.5554830  7.0705031  7.2623465
#>  [266,]  1.8390428  2.3491219  2.4972937  2.4982497
#>  [267,] 10.0561440 10.9073696 11.6970219 11.7700671
#>  [268,]  7.4288530  9.3675286 10.0754747 10.7041630
#>  [269,]  1.8078804  2.1775816  2.4822815  2.5183049
#>  [270,]  2.4192171  2.8512381  3.1088762  3.1919575
#>  [271,]  3.8047821  3.8809997  3.8987604  3.9071333
#>  [272,]  8.4058113 10.5995643 11.9204592 12.3965081
#>  [273,]  2.4916406  2.7365911  2.8244230  2.8925997
#>  [274,]  4.8112511  4.9689513  5.0723842  5.1658995
#>  [275,] 11.7148231 12.8528555 13.6862016 13.7388815
#>  [276,]  2.7337751  3.4383336  3.5090100  3.5190095
#>  [277,] 10.4909032 12.1086043 13.3240100 14.3173183
#>  [278,]  1.5806180  1.8382142  2.0359411  2.1476982
#>  [279,]  2.7670731  3.8028891  4.3682415  4.4599510
#>  [280,]  8.1848526  9.5439102 10.0699229 10.4278528
#>  [281,]  4.2533456  5.5868428  5.9246629  6.0469219
#>  [282,]  2.5678807  3.0513315  3.1615793  3.2051405
#>  [283,]  1.5332429  1.8551944  1.8974069  1.9166915
#>  [284,]  6.1197725  6.9590190  7.6459979  7.9464749
#>  [285,]  1.9251282  2.1185497  2.1606421  2.1953201
#>  [286,]  9.8469405 11.0308922 11.2285926 11.2439611
#>  [287,]  3.0839948  3.6449215  4.0126628  4.0434571
#>  [288,]  2.4796158  3.0817342  3.2369725  3.3288921
#>  [289,]  6.4427375  6.9291906  7.1021978  7.2677756
#>  [290,]  2.5279178  2.8610398  3.1168829  3.3694891
#>  [291,]  2.5200426  2.6347012  2.6693428  2.6848111
#>  [292,]  4.6598036  5.3374105  5.8163097  6.1629159
#>  [293,] 10.4856348 11.5270987 12.1308531 12.6637608
#>  [294,]  6.4329093  7.6818865  8.6079809  8.6433386
#>  [295,]  3.7652816  4.7946362  5.3712215  5.7032007
#>  [296,]  3.3745696  4.2858076  4.6991485  4.9993897
#>  [297,]  3.9934671  4.5620879  5.0206536  5.4547713
#>  [298,]  2.1691494  2.7682564  3.1184353  3.2760224
#>  [299,]  4.0253078  4.4063005  4.7709776  5.1302480
#>  [300,]  7.2461436  8.2994870  9.0818675  9.5198869
#>  [301,]  2.1592063  2.5993865  2.7047180  2.7261023
#>  [302,]  3.2580381  3.7358856  4.0778320  4.1826747
#>  [303,]  5.3761309  6.3253625  6.4695551  6.4817002
#>  [304,]  2.1260869  2.2649473  2.3786352  2.4533817
#>  [305,]  2.7231657  3.0070794  3.1353450  3.2403505
#>  [306,]  0.5398835  0.6310514  0.6420235  0.6442429
#>  [307,]  2.5797624  3.2884818  3.3957325  3.4826702
#>  [308,]  2.9381561  3.5479840  4.1515038  4.5058856
#>  [309,]  9.6042097 12.4763503 14.5113055 16.0851937
#>  [310,]  4.8342751  5.1830341  5.2136192  5.2423948
#>  [311,]  2.3297529  2.5732961  2.7383015  2.8622847
#>  [312,]  5.4053495  6.8900686  7.4668871  7.6879862
#>  [313,]  4.0611530  4.8136401  5.3682963  5.3694295
#>  [314,]  4.6604100  6.5238069  7.2082363  7.5524509
#>  [315,]  8.6449227  9.4454511  9.5558650  9.5905606
#>  [316,]  1.8223261  2.2106610  2.5195820  2.5812645
#>  [317,]  3.3753688  4.0365560  4.6011261  4.8349723
#>  [318,]  2.7027345  3.0524205  3.3608187  3.4735410
#>  [319,]  5.2832119  6.0559434  6.5426995  6.9668233
#>  [320,] 15.9175152 21.6014297 23.1683339 23.3866608
#>  [321,]  6.1227919  8.2740334  9.0384529  9.7242821
#>  [322,]  3.9660900  4.6889239  4.7810483  4.8368923
#>  [323,]  4.0647674  5.1731405  6.1722208  6.4256251
#>  [324,]  1.4142209  1.4396979  1.4612156  1.4824284
#>  [325,]  7.5890996  9.6739599 11.3209350 12.1948401
#>  [326,]  2.8071949  3.4806275  4.0133384  4.1612021
#>  [327,]  4.8821269  5.5341313  5.7555442  5.8128602
#>  [328,]  1.9857931  2.2017659  2.2523302  2.2953066
#>  [329,]  3.4008747  3.5282920  3.6331855  3.7104769
#>  [330,]  8.1785035  9.4494239 10.3112170 10.4743475
#>  [331,]  1.1825939  1.6386496  2.0043340  2.0431014
#>  [332,]  4.2093618  5.7916886  6.9797352  8.1314208
#>  [333,]  6.2607811  6.4797763  6.6040702  6.6943612
#>  [334,]  7.3615726  8.4970880  8.8482680  8.9121406
#>  [335,]  1.9634296  2.3723100  2.7194349  2.9879989
#>  [336,]  2.9458551  3.6596513  4.2130990  4.7376259
#>  [337,]  6.4582366  7.6689062  8.1046008  8.4140506
#>  [338,]  4.4540921  4.7723676  4.9709074  5.1599875
#>  [339,]  2.3825601  3.2289712  3.5179231  3.6911214
#>  [340,]  3.7730189  4.8878565  5.5573396  6.0073101
#>  [341,]  6.5737107  7.4439525  8.0139113  8.4472076
#>  [342,] 15.9059093 18.3936193 20.1638799 20.4104096
#>  [343,]  3.1620234  3.4299765  3.5889839  3.7228845
#>  [344,]  4.8699637  5.9488590  6.9271681  7.1789272
#>  [345,]  6.7969975  7.8834445  8.8505423  9.0254734
#>  [346,]  6.7101805  8.1712862  9.0902989  9.4098611
#>  [347,]  2.4550057  3.0914176  3.6431501  3.7570181
#>  [348,]  3.9125555  4.3810161  4.7301690  5.0676097
#>  [349,]  4.6659029  5.8113469  6.8425523  6.9935957
#>  [350,]  4.6322996  5.4476865  5.5898114  5.6406443
#>  [351,]  2.8754464  3.2237273  3.5397274  3.7209811
#>  [352,]  6.0712803  8.1828793 10.0429099 10.7901275
#>  [353,] 10.0213425 11.1058523 11.9907388 12.5233897
#>  [354,]  6.8046803  8.3057947  9.3063046  9.6810975
#>  [355,]  1.4387435  1.6367930  1.8101166  1.9120104
#>  [356,]  1.5492385  2.1439025  2.6362298  3.0252453
#>  [357,]  9.6133344 10.4778913 11.0457537 11.5151525
#>  [358,]  4.1935272  4.3434130  4.3995692  4.4267083
#>  [359,]  4.1458715  5.3975246  6.0268221  6.3199144
#>  [360,]  6.2875987  8.3314431  9.7901797 10.4688495
#>  [361,]  5.1365011  6.1373365  6.1496915  6.1519652
#>  [362,]  4.7202620  5.4962127  5.9470288  6.0326073
#>  [363,]  6.3324486  7.2305347  7.5750145  7.8591881
#>  [364,]  0.6548225  0.6753838  0.6891487  0.6942007
#>  [365,]  3.2590357  3.8292028  4.1112977  4.1399335
#>  [366,]  1.8315261  1.8943386  1.9493178  1.9509805
#>  [367,]  2.1469893  2.3077359  2.4021175  2.4386880
#>  [368,]  5.0766073  6.5113422  7.4454505  7.6056664
#>  [369,]  1.5176326  1.6550933  1.6776903  1.6819400
#>  [370,]  2.8544613  3.6317704  4.3035764  4.3091494
#>  [371,]  4.3213948  4.7978820  5.0247143  5.0292279
#>  [372,]  4.1823225  5.6787980  5.9710998  6.0473047
#>  [373,]  5.4497881  6.0171296  6.5466703  6.8888595
#>  [374,]  1.7277087  2.2141090  2.3322783  2.4188292
#>  [375,]  3.5670579  4.3780180  5.1357996  5.5425491
#>  [376,]  2.2604822  3.0207963  3.7484844  3.8833093
#>  [377,]  8.7111528  9.2029465  9.5883413  9.8818600
#>  [378,]  1.9487018  2.4998183  2.9431501  2.9606678
#>  [379,]  8.2609136  8.9492516  9.1437349  9.2609117
#>  [380,]  4.1379945  5.8314667  6.4200020  6.6177915
#>  [381,]  3.2808401  4.1022627  4.7230223  4.9076744
#>  [382,]  2.2082056  2.8783218  3.1012878  3.1233651
#>  [383,]  3.0562486  4.0561622  4.5111565  4.7501214
#>  [384,]  8.8854921 11.2052152 11.6697010 11.6959133
#>  [385,]  4.9533903  6.3028286  7.3115949  7.5524686
#>  [386,]  1.3456377  1.7544051  1.9259916  1.9803616
#>  [387,]  2.5990856  3.1075149  3.5124602  3.7936155
#>  [388,]  2.8566253  3.4465137  3.8465292  4.0635625
#>  [389,]  8.3630948 10.3494885 11.9555354 13.3150439
#>  [390,]  7.2419144  8.0216227  8.7736871  9.1951739
#>  [391,]  6.0249744  6.5636593  6.9377266  7.2344059
#>  [392,]  2.8923063  4.2875248  5.0660427  5.3226771
#>  [393,]  7.0833360  7.9002242  8.6182004  8.8981623
#>  [394,]  3.3882655  4.6752606  5.5346650  5.7136921
#>  [395,]  6.1331936  7.0326236  7.3912515  7.5735550
#>  [396,]  1.7771285  2.1655755  2.3769744  2.4883400
#>  [397,]  4.9713261  5.8076750  6.3151955  6.5644838
#>  [398,]  2.7843067  3.4272894  3.8778798  4.0812880
#>  [399,]  2.0083794  2.5022264  2.6649157  2.6687683
#>  [400,]  5.7146093  6.2092031  6.7036815  7.0370023
#>  [401,]  8.2686150 11.3904097 12.9742087 14.4776814
#>  [402,]  4.4996748  5.8137910  6.0882436  6.1588082
#>  [403,]  4.9431088  5.8752045  6.3597977  6.6777728
#>  [404,]  2.7171417  3.0083585  3.2774969  3.4030274
#>  [405,]  3.2567497  4.2011903  4.9892611  5.3717232
#>  [406,]  3.6047147  3.7466364  3.7581583  3.7615673
#>  [407,]  4.1481165  4.8125260  5.4392351  5.6489296
#>  [408,]  5.4917008  6.0175820  6.3539626  6.3898342
#>  [409,]  4.6441598  5.2910111  5.3446146  5.3933490
#>  [410,]  1.8053819  2.0650543  2.2480131  2.3274385
#>  [411,]  3.7256465  4.7639117  5.0094726  5.1276707
#>  [412,]  6.5549719  7.7824290  8.4577183  8.5514843
#>  [413,]  8.7623748 10.0137078 10.9634920 11.5728401
#>  [414,]  9.5148729 12.3515409 14.2529530 14.9462593
#>  [415,]  4.7766024  5.0043693  5.1277597  5.1586361
#>  [416,]  5.8477025  6.3384238  6.3583900  6.3701216
#>  [417,]  5.2341899  5.7027990  6.0561700  6.1170743
#>  [418,]  1.1905989  1.3668344  1.5252209  1.6336639
#>  [419,]  2.1185223  2.5826067  2.9614523  3.0381696
#>  [420,]  5.4401287  6.1437602  6.7892975  7.2182222
#>  [421,]  7.3384179  7.5971057  7.7727890  7.8295346
#>  [422,]  3.0249485  4.1884285  4.2835887  4.3490831
#>  [423,]  2.0125045  2.4630311  2.7202367  2.8922476
#>  [424,]  4.0668990  4.3431464  4.4404271  4.4907035
#>  [425,]  5.4534995  5.9176624  6.2433012  6.3235743
#>  [426,]  6.7515556  7.5453443  8.2109768  8.5395186
#>  [427,]  3.6609095  4.1591808  4.2449218  4.2817651
#>  [428,]  8.2815660  9.1723982  9.6156741  9.6638311
#>  [429,]  1.6399596  1.8669197  2.0193303  2.0747205
#>  [430,]  2.8748729  3.5131522  3.5833449  3.6286369
#>  [431,]  1.0208925  1.2063438  1.3144644  1.3270755
#>  [432,]  4.0868934  4.8029742  5.2652394  5.4096324
#>  [433,]  4.9095112  5.7243384  6.2960698  6.7420384
#>  [434,]  3.7890386  4.5577056  4.5720836  4.5862093
#>  [435,]  1.3198659  1.8210089  1.8684479  1.8963170
#>  [436,]  1.2240242  1.6876587  1.9181652  2.1378440
#>  [437,]  2.7374934  3.1098893  3.2787527  3.3296358
#>  [438,]  1.6905014  2.1355050  2.5626914  2.8920835
#>  [439,]  1.7129754  2.2851168  2.5553993  2.7708823
#>  [440,]  1.5960128  2.0896222  2.1617933  2.1797623
#>  [441,]  3.7738311  4.7788777  5.7502360  5.8578415
#>  [442,]  7.2494881  9.7006402 11.8320313 12.9285931
#>  [443,] 13.4245565 14.7282648 15.7336720 16.3721663
#>  [444,]  8.5803526  9.4841284 10.3504814 10.4441020
#>  [445,]  2.9009730  3.0351147  3.1245821  3.1289222
#>  [446,]  3.1805011  4.0416518  4.4265694  4.7577273
#>  [447,]  4.0013088  4.5768653  5.0590055  5.4152707
#>  [448,]  5.6301679  6.3075824  6.7736612  7.1343612
#>  [449,]  3.3377393  4.2991392  4.6231901  4.9423505
#>  [450,]  5.6414546  6.3816692  6.9182291  7.1609280
#>  [451,]  2.6862726  3.2867486  3.8823908  3.8969574
#>  [452,]  1.6775436  2.1030407  2.4666552  2.5213006
#>  [453,]  2.0023156  2.5363531  2.9705532  3.3589979
#>  [454,]  3.2624458  3.8102684  4.1601909  4.3188628
#>  [455,]  7.6008415  8.6579443  9.4994869  9.7671562
#>  [456,]  3.8563435  5.2569708  5.7711398  6.0432003
#>  [457,]  6.1740406  7.1897734  7.9671702  8.0139434
#>  [458,]  3.7883826  4.2304530  4.6571455  4.7942053
#>  [459,]  8.3868698 10.2164356 11.5068752 12.1344648
#>  [460,]  1.9456526  2.1738314  2.3961458  2.4649144
#>  [461,]  1.0791998  1.3893831  1.4847933  1.5449625
#>  [462,]  3.3120700  3.5107453  3.6853568  3.8499735
#>  [463,]  5.3214066  7.5119100  9.5345107  9.7387437
#>  [464,]  1.8832214  2.1948211  2.3496627  2.4842586
#>  [465,]  2.9571912  3.8108925  4.2200998  4.2471643
#>  [466,]  4.1685225  5.4443875  6.5240426  6.9750889
#>  [467,]  2.6036242  3.3300341  3.7391211  3.9260607
#>  [468,]  7.7227246  8.9505975  9.7889547  9.8896233
#>  [469,]  2.1741272  2.8390686  2.9498948  3.0231046
#>  [470,]  1.8971793  2.1925732  2.3160718  2.4281301
#>  [471,]  1.9677750  2.1619713  2.3388825  2.4395422
#>  [472,]  6.6449601  8.6504295 10.0997246 10.7190602
#>  [473,]  6.3238058  8.8921036  9.7347531  9.8341967
#>  [474,]  2.5493377  3.1794987  3.4907166  3.5630103
#>  [475,]  6.9262116  7.9847764  8.0660857  8.0678237
#>  [476,]  4.4066653  6.2424002  6.5692165  6.6571613
#>  [477,]  3.1383515  3.9044438  4.4613120  4.9189482
#>  [478,]  0.9319563  1.2911340  1.4114099  1.4165768
#>  [479,]  3.8629633  4.8993084  5.8642777  6.7391103
#>  [480,]  2.3839743  2.8008036  2.8233125  2.8387311
#>  [481,]  1.9673361  2.5563672  3.0245120  3.4588864
#>  [482,]  2.9111918  2.9744257  3.0031311  3.0176561
#>  [483,]  2.5254200  3.0243221  3.5213343  3.5817824
#>  [484,]  6.3554318  6.8448701  7.0630844  7.1941171
#>  [485,]  5.1688913  5.2935195  5.3813823  5.4000684
#>  [486,]  2.5961114  3.2672146  3.5151699  3.5645899
#>  [487,]  4.8171283  5.9917284  6.1720888  6.2067198
#>  [488,]  1.8855389  2.2375628  2.3849277  2.4231129
#>  [489,]  4.0196644  4.1564029  4.2867546  4.3637765
#>  [490,] 11.6920094 13.7064669 15.2573618 15.6841763
#>  [491,]  7.4301729  7.9472010  8.4559688  8.5629480
#>  [492,]  2.1970638  2.5911329  2.8136907  3.0044128
#>  [493,]  0.9479011  1.1842743  1.2920838  1.3058314
#>  [494,]  9.4854386  9.9204382 10.2159205 10.3525052
#>  [495,]  1.8737096  2.2756312  2.3719804  2.3880338
#>  [496,]  6.3410708  7.6371078  8.4535591  9.0026957
#>  [497,]  3.4658209  3.8202120  4.1157886  4.2801761
#>  [498,]  4.4353981  4.7602877  4.9632223  4.9774724
#>  [499,]  1.4628244  1.8189320  2.1090477  2.3886611
#>  [500,] 15.2416620 16.8964433 18.2164688 18.4357454
#>  [501,] 10.5622469 13.8857847 16.7655435 17.9999264
#>  [502,]  3.6297273  4.2838675  4.4816549  4.6186757
#>  [503,]  1.1503411  1.2506297  1.3341995  1.3992029
#>  [504,]  5.9202260  6.7538959  6.9674655  7.1716502
#>  [505,]  5.2242225  7.5089241  7.7378351  7.7502121
#>  [506,]  7.8497127  9.1638987  9.9294202 10.0190161
#>  [507,]  8.8502194 11.2734567 13.1249824 13.3461860
#>  [508,]  7.1284726  8.9316372  9.4001644  9.6550577
#>  [509,]  1.8136622  2.0926260  2.2452473  2.3212667
#>  [510,]  2.6192108  2.6562377  2.6779612  2.6936285
#>  [511,]  5.5776858  6.5997248  6.8127699  6.8608313
#>  [512,] 10.5241057 13.0897461 14.7584057 14.9477552
#>  [513,]  1.7928949  2.3904568  2.9782902  3.2114975
#>  [514,]  5.7701604  5.8954317  5.9765176  5.9806052
#>  [515,]  3.5321534  3.7755189  3.9835652  4.0232820
#>  [516,]  7.9954401  9.5170876  9.8683110  9.9764028
#>  [517,] 12.4592475 13.0493246 13.1345097 13.1563526
#>  [518,]  9.7334444 12.3471773 14.3519444 14.4623177
#>  [519,]  5.0422738  5.2369283  5.2877880  5.2931075
#>  [520,]  4.0263249  4.6861463  5.0523623  5.0821558
#>  [521,]  1.3402923  1.7455293  2.1473141  2.3141826
#>  [522,]  6.9290539  7.0457248  7.0727159  7.0971855
#>  [523,]  4.2222168  5.7313022  6.1011706  6.4167024
#>  [524,]  1.5140612  1.5898802  1.6353556  1.6800014
#>  [525,]  1.6571909  1.9021566  2.0273703  2.0929243
#>  [526,]  5.9629124  6.9959609  7.3298534  7.3868990
#>  [527,]  4.5679985  5.6430575  6.0517841  6.0618019
#>  [528,]  7.6600900  9.0093938  9.2161543  9.3711702
#>  [529,]  7.1266927  7.3522014  7.5538833  7.6698123
#>  [530,]  5.1896598  6.3564495  7.0874247  7.3826485
#>  [531,]  6.0015937  6.2675153  6.3441373  6.3834514
#>  [532,]  7.3350528  7.7914452  8.1692677  8.4083200
#>  [533,]  8.3836584  9.8132355 10.7545996 10.7861643
#>  [534,]  2.9409087  3.1918344  3.2921777  3.2946068
#>  [535,]  1.8237110  2.1620720  2.3497020  2.3635735
#>  [536,]  4.4140582  5.7548964  6.2756407  6.5133582
#>  [537,]  4.2265517  5.3276273  6.0734710  6.3962426
#>  [538,]  9.6677321 11.9416911 12.1071939 12.1574005
#>  [539,]  4.3323279  5.2081723  5.8945078  6.2430447
#>  [540,]  1.5030631  1.9633692  2.0290395  2.0741036
#>  [541,]  7.4862286 10.1318385 11.8361932 12.9416816
#>  [542,]  1.6071241  2.0949865  2.4773413  2.5947475
#>  [543,]  2.2169836  2.6140699  2.7335066  2.7821492
#>  [544,]  3.1725655  3.5367590  3.6568541  3.6793747
#>  [545,]  7.3662466 10.4169543 12.4181387 13.3421847
#>  [546,]  4.9798416  6.0083378  6.7336793  7.0171750
#>  [547,]  3.0916506  4.1498062  4.8689878  5.1926024
#>  [548,]  9.3105330 11.8452572 12.8343551 12.9000501
#>  [549,]  7.9760260  8.8950172  9.1667696  9.3573656
#>  [550,]  2.6622397  2.9350556  3.1887200  3.2548153
#>  [551,]  6.9451606  7.7960387  8.0310761  8.1586926
#>  [552,]  3.2658356  4.1672122  4.6545990  4.7431468
#>  [553,]  2.7333980  3.8974041  4.3973886  4.6015456
#>  [554,]  2.8108921  3.8266281  4.6180055  5.0278608
#>  [555,]  3.7837501  4.5514240  4.9178430  5.0216544
#>  [556,]  7.1540533  7.8031488  8.3948564  8.5223629
#>  [557,]  5.8508421  5.9857556  6.0911598  6.1264229
#>  [558,]  6.0906993  6.9973295  7.7595648  7.7642436
#>  [559,]  5.3204089  7.3695209  7.7687429  7.7882717
#>  [560,]  6.5695193  7.2081388  7.5680819  7.6214489
#>  [561,]  7.6619059  8.3203961  8.8111173  9.0514290
#>  [562,] 10.0176975 12.2216663 12.5493951 12.8492293
#>  [563,]  2.4368496  2.9293909  2.9861428  3.0214809
#>  [564,]  2.4842622  3.3709776  3.7331862  3.9856203
#>  [565,]  0.8623073  1.0048637  1.0871222  1.1099209
#>  [566,]  5.7301248  6.3005264  6.6120613  6.6904873
#>  [567,]  2.0885667  2.8987550  3.2258781  3.5526594
#>  [568,]  2.0540883  2.2381176  2.3811017  2.4140114
#>  [569,]  7.3727744  7.8411763  8.1610418  8.4299803
#>  [570,]  4.4807596  5.0045154  5.3478679  5.4177423
#>  [571,]  3.7161867  4.1116090  4.2845618  4.4223702
#>  [572,] 10.0737425 10.9186525 11.3378582 11.5992756
#>  [573,]  0.7478587  0.8157925  0.8822688  0.9036944
#>  [574,]  1.0011680  1.2916905  1.3984124  1.4268249
#>  [575,]  3.7352274  4.2953959  4.3850436  4.4677087
#>  [576,]  2.8762876  4.0233272  4.6594602  4.9167497
#>  [577,]  8.6739838  9.9773584 10.2061207 10.2704505
#>  [578,]  4.9264715  6.3733116  7.6762670  8.4376121
#>  [579,]  5.4056641  6.4196038  7.2965367  7.2984152
#>  [580,]  2.4136326  3.4918515  4.0764185  4.3567053
#>  [581,]  3.0542981  3.5311829  3.6499408  3.6586426
#>  [582,]  3.0752892  4.0744149  4.9442828  5.0809582
#>  [583,]  0.9760454  1.1032450  1.2303673  1.2574918
#>  [584,]  3.4874820  4.1425422  4.5957291  4.9210986
#>  [585,]  5.2510156  7.0919447  7.7578249  7.9712390
#>  [586,]  8.9889934 10.4942511 11.0259296 11.1195510
#>  [587,]  6.1774610  6.3533013  6.4010453  6.4433850
#>  [588,]  7.0261414  8.3774895  8.4585454  8.4698395
#>  [589,]  4.2541095  5.3902692  6.0669658  6.3276250
#>  [590,]  1.5184752  1.7107719  1.8048412  1.8921614
#>  [591,] 13.8669963 15.6537081 15.8204349 15.8907472
#>  [592,]  0.6519811  0.8285326  0.9254872  0.9258290
#>  [593,]  6.6627659  7.9674272  8.3027078  8.4518985
#>  [594,]  1.7660442  1.8013193  1.8264566  1.8404148
#>  [595,]  4.7132895  5.5227309  6.3314738  6.3489645
#>  [596,]  2.9357207  3.0047823  3.0695890  3.1320090
#>  [597,]  1.5105905  1.5737818  1.5929182  1.5956164
#>  [598,]  1.0073463  1.2513971  1.4382214  1.5210609
#>  [599,]  9.2292829 10.4362376 10.8933039 11.2628834
#>  [600,]  4.2855558  5.5631895  6.6213142  7.0345182
#>  [601,]  4.7986988  5.6306072  6.2801665  6.4820334
#>  [602,]  3.0606729  4.4537631  5.2637857  5.6975001
#>  [603,] 11.0989684 11.9756064 12.7618884 12.7729811
#>  [604,]  5.3543101  6.8232959  7.8339941  8.1280952
#>  [605,]  8.7572014  9.2256120  9.5507409  9.7962139
#>  [606,]  4.8967299  6.0403323  6.6381793  6.7896035
#>  [607,]  6.8939858  7.4118119  7.5578852  7.5684595
#>  [608,]  3.8386625  4.6961977  5.3392802  5.7845517
#>  [609,]  2.3831087  3.1054019  3.7823518  3.9860024
#>  [610,]  6.9993652  8.3835643  9.5882566  9.8626800
#>  [611,]  3.9940200  5.1069140  5.8438785  6.2926756
#>  [612,]  3.9070268  4.7105767  5.4463061  5.9326876
#>  [613,]  4.2108016  4.3753053  4.4057005  4.4092826
#>  [614,]  3.3238710  4.1479717  4.6226728  4.9293670
#>  [615,] 11.0266963 11.8530229 12.3921934 12.9248942
#>  [616,]  6.3793271  8.0118995  9.3117391  9.4231322
#>  [617,]  1.8891631  2.0636984  2.2119685  2.2977902
#>  [618,]  3.6329432  4.8230821  5.5212428  6.0255862
#>  [619,]  0.9135429  1.2043924  1.4539719  1.6602174
#>  [620,]  8.3978123  8.9237361  9.3368460  9.3432396
#>  [621,]  8.4499747  9.7593933 10.5534517 10.5643361
#>  [622,]  6.6359579  9.0014638 10.9695198 12.2055048
#>  [623,]  1.6114291  1.8514286  1.9095009  1.9583670
#>  [624,]  2.0069246  2.7074018  3.2894281  3.4945526
#>  [625,]  5.7546518  6.2692750  6.2939103  6.2955601
#>  [626,]  1.8512035  2.0099644  2.1236823  2.1273014
#>  [627,]  2.4227656  2.8465125  3.2554583  3.3211025
#>  [628,]  4.4411388  5.6007259  6.3989132  6.5063330
#>  [629,]  4.7909909  5.7124434  6.1813274  6.4645975
#>  [630,]  3.2517803  3.6007757  3.8223290  4.0184462
#>  [631,]  4.5335178  5.5196215  6.0713182  6.1649353
#>  [632,]  7.1665782  8.5419262  9.0327532  9.0970090
#>  [633,]  9.2370671 10.5391961 11.2357330 11.8417972
#>  [634,]  5.2769010  7.2664044  8.8678103  9.8047233
#>  [635,]  6.8433411  7.6855910  7.8378749  7.9859709
#>  [636,]  2.3692127  2.7900373  2.9215320  2.9904450
#>  [637,]  1.3365154  1.4712257  1.5534571  1.5685637
#>  [638,]  4.4024823  5.0658038  5.3243753  5.4944823
#>  [639,]  0.6393595  0.8511101  1.0483926  1.2039563
#>  [640,]  4.5527929  6.2359905  7.2859242  7.8628728
#>  [641,]  2.6755167  3.7697199  4.3811753  4.8423763
#>  [642,]  4.6931744  6.2656727  6.9856780  7.4133168
#>  [643,]  3.8735903  4.3282756  4.5236435  4.6231894
#>  [644,]  3.3705586  3.9387509  4.2842360  4.4872762
#>  [645,]  3.9355285  5.3543787  5.9214893  6.2535023
#>  [646,]  7.7580313  8.6312604  9.2353611  9.4800823
#>  [647,]  7.4059500  8.2664437  9.0121867  9.6662100
#>  [648,]  4.4173770  5.4435555  5.9944060  6.2146277
#>  [649,]  2.2800767  2.5066939  2.6832388  2.7654863
#>  [650,]  3.2197986  4.4745265  5.3029209  5.8117472
#>  [651,]  3.1167160  3.9045371  4.5058937  4.9593890
#>  [652,]  3.6949069  4.9794170  5.7664264  6.2129696
#>  [653,]  5.8748077  7.0278997  7.6927444  7.7598977
#>  [654,]  4.8214158  5.8961549  6.4349251  6.5471306
#>  [655,]  6.4680386  8.4246528 10.1632546 10.2112116
#>  [656,]  2.3396541  2.8759889  3.2053918  3.3482666
#>  [657,]  4.9296011  5.9261015  6.7414406  7.1177171
#>  [658,]  2.1051211  2.5106533  2.8033116  2.8751836
#>  [659,]  7.2033312  8.1186229  8.3677009  8.4107057
#>  [660,]  1.4350925  1.6707463  1.8373593  1.8454280
#>  [661,]  2.9880132  3.1399766  3.2355806  3.2959708
#>  [662,]  5.6286334  7.3109061  8.4135381  9.5088571
#>  [663,]  5.0463455  5.4444816  5.6524089  5.8510829
#>  [664,]  8.3516170  8.6758797  8.7942834  8.8299547
#>  [665,]  3.1872850  3.8048330  4.3539468  4.3722997
#>  [666,]  4.0121151  4.5084003  4.6599947  4.7701917
#>  [667,]  5.7570994  6.0282857  6.1650871  6.2047721
#>  [668,] 17.8335215 18.9655737 19.2040917 19.2066884
#>  [669,]  1.6155654  2.1742424  2.5669097  2.6767630
#>  [670,] 12.0609878 12.9150391 13.4553156 13.5179782
#>  [671,]  3.7274957  3.9170037  4.0497030  4.1366051
#>  [672,]  4.4288635  4.6934014  4.7823132  4.8186649
#>  [673,]  5.8487972  7.2058064  7.5086853  7.7333538
#>  [674,]  4.0252701  4.1831037  4.2246730  4.2574515
#>  [675,]  5.0031435  5.6574729  5.8934456  5.9250428
#>  [676,]  2.7869961  3.0012875  3.0868881  3.0922584
#>  [677,]  2.8898662  3.5050332  3.8346082  3.9748088
#>  [678,]  4.3863811  4.7430593  4.8227561  4.8797724
#>  [679,]  1.8650688  2.1907552  2.3147706  2.3287100
#>  [680,]  2.8899958  3.5237157  4.0862433  4.1930203
#>  [681,]  2.6922738  3.0248971  3.2455444  3.3631779
#>  [682,]  4.9160780  5.8410551  6.4500117  6.4900228
#>  [683,]  2.3536209  2.9972450  3.4991431  3.7300450
#>  [684,]  6.9357613  7.1442378  7.1810089  7.2162918
#>  [685,]  3.5478706  4.0684156  4.2900212  4.2917867
#>  [686,]  3.0075504  3.5277346  4.0380442  4.4348029
#>  [687,]  2.3750114  2.9544987  3.4837778  3.9276761
#>  [688,]  5.4478100  6.3765541  6.9690156  7.2641909
#>  [689,]  5.7592275  6.6466180  7.0057741  7.3354580
#>  [690,]  3.8772508  4.4658001  4.6239167  4.6746009
#>  [691,] 11.9176004 12.6039461 12.9816267 13.1265095
#>  [692,]  1.6073868  1.9376053  2.1549678  2.3484922
#>  [693,]  3.8166184  4.0676598  4.2335396  4.2905768
#>  [694,]  2.5939905  2.9661748  3.3356670  3.3393640
#>  [695,]  5.8885772  7.0235998  7.6339067  7.7020833
#>  [696,]  3.4541077  3.9997846  4.2695366  4.4225284
#>  [697,]  5.8849432  7.4107778  8.8631846  9.2793074
#>  [698,]  4.2248905  4.6614471  4.8792055  4.9101965
#>  [699,]  7.7248582  8.4298358  8.7447331  9.0294256
#>  [700,]  3.1901719  3.4979690  3.6658544  3.8264523
#>  [701,]  4.5045820  4.6382059  4.7408107  4.7695719
#>  [702,]  1.7530040  2.2022571  2.5808547  2.6172935
#>  [703,]  3.7084716  4.2613354  4.5413717  4.6908966
#>  [704,]  2.8407146  3.3291761  3.6213639  3.6372236
#>  [705,]  2.4286544  2.7586697  2.9913610  3.2226506
#>  [706,]  0.4260195  0.5474410  0.5596559  0.5700729
#>  [707,]  5.9273885  6.5243470  6.9592589  7.0212407
#>  [708,]  2.4836982  3.5925754  4.4505786  5.0486209
#>  [709,]  1.9300079  2.3121126  2.5652306  2.6582968
#>  [710,]  4.3181618  5.3957001  5.8849334  6.0968817
#>  [711,]  2.9191412  3.8841011  4.1394191  4.2747784
#>  [712,]  2.4035131  2.6313320  2.7648400  2.8261816
#>  [713,]  2.6548157  3.5363604  4.3202388  4.3801026
#>  [714,]  1.8947392  2.4722169  2.9401499  3.2032581
#>  [715,] 10.0079585 10.9597884 11.5055735 11.9569412
#>  [716,]  7.9599079  8.0766543  8.1325729  8.1568380
#>  [717,]  5.4037001  5.8381959  6.0858020  6.3152737
#>  [718,]  2.6179081  3.1742247  3.2725449  3.3340763
#>  [719,]  3.5919237  4.3293551  4.9012466  5.1306311
#>  [720,]  4.7173487  5.0018833  5.2490769  5.2866661
#>  [721,]  3.4963572  4.1496332  4.7432628  5.1030741
#>  [722,]  3.4577932  4.2027792  4.9444404  5.0749384
#>  [723,]  3.1066355  3.7659728  4.0753965  4.0914291
#>  [724,] 10.5963255 11.4729928 11.9100922 12.2513402
#>  [725,]  1.8953189  2.2995866  2.5375772  2.5802263
#>  [726,]  1.3006067  1.6446290  1.8075124  1.9699496
#>  [727,]  3.9023572  4.2073859  4.4940018  4.7410234
#>  [728,]  3.5870713  4.0843409  4.5619143  4.7642175
#>  [729,]  2.8944938  3.2926600  3.4896889  3.5113740
#>  [730,]  1.4127746  1.9038299  2.3267382  2.3366205
#>  [731,]  1.4970204  2.0023683  2.0396516  2.0410196
#>  [732,]  4.5130352  5.1464738  5.7584419  6.2520067
#>  [733,]  6.1589292  8.0030838  9.3304820  9.6152138
#>  [734,] 21.3991955 22.9223326 23.1680397 23.3945689
#>  [735,]  3.0450362  3.7264884  3.8750480  4.0088316
#>  [736,]  3.6881999  4.2781564  4.5445016  4.6606670
#>  [737,]  4.6239590  6.1827070  6.5684291  6.7482029
#>  [738,]  3.3407863  4.1973471  4.7935580  4.9282659
#>  [739,]  3.4534569  3.9804990  4.2404186  4.4570134
#>  [740,]  4.4106012  5.6752123  6.3006921  6.7368262
#>  [741,]  2.1650381  3.0571149  3.6793800  3.8860890
#>  [742,]  2.3855376  2.6434609  2.8207169  2.8997638
#>  [743,]  9.1978176  9.7421884 10.0190939 10.2859000
#>  [744,]  1.4478650  1.7562868  1.8574447  1.8872766
#>  [745,]  1.2892539  1.4988505  1.6740840  1.8146802
#>  [746,]  1.7604713  1.8856071  1.9864116  2.0188377
#>  [747,]  5.1581001  7.2952285  7.8078143  7.8593122
#>  [748,]  2.6365753  2.8790972  2.9821398  3.0423082
#>  [749,]  2.1502502  2.4478680  2.4796963  2.4969474
#>  [750,]  4.0584593  5.2882868  5.6869347  5.8097262
#>  [751,]  2.5048995  3.0584207  3.4961819  3.5122000
#>  [752,]  2.7919605  3.6043958  3.9616569  4.0738335
#>  [753,]  1.8849843  2.2773189  2.6160116  2.6588003
#>  [754,]  1.6393416  1.9589010  2.0679405  2.0907089
#>  [755,]  7.5346138  8.0679929  8.4852203  8.6471894
#>  [756,]  0.9218050  1.2559807  1.4916037  1.5699438
#>  [757,]  7.8271314 10.2673304 11.5903963 11.7397191
#>  [758,]  4.5798057  4.8377198  4.8932328  4.9326224
#>  [759,]  3.0967291  4.4107438  5.6853489  5.9777164
#>  [760,]  3.0030880  3.3780648  3.6216971  3.7833677
#>  [761,]  3.3103077  3.5065665  3.6176956  3.6397505
#>  [762,]  6.9404789  8.9952078  9.3343119  9.3484516
#>  [763,]  3.2055358  3.6615888  3.9089196  4.0554680
#>  [764,]  2.4592263  2.7779473  3.0495991  3.1494787
#>  [765,]  6.3084177  7.2522476  8.1501017  8.3006363
#>  [766,]  6.9504281  8.5228685  8.8439735  9.1365644
#>  [767,]  4.7952150  4.9782142  5.1174876  5.1385177
#>  [768,]  3.6836153  5.3286485  6.8483548  8.2123746
#>  [769,] 12.1451983 13.8634380 15.5391415 16.7030283
#>  [770,]  2.1656702  2.4887574  2.7548098  2.8702590
#>  [771,]  7.6664823  9.3714595 10.8713862 10.9717213
#>  [772,]  2.3231906  2.9530882  3.5658555  3.6465808
#>  [773,]  5.0270237  5.6564281  5.9974779  6.3359852
#>  [774,]  5.8080349  7.2852023  8.0724803  8.8086323
#>  [775,]  4.2408391  4.5079756  4.7060049  4.7399880
#>  [776,]  1.3799067  1.8536869  2.2042008  2.2764287
#>  [777,]  6.3593643  8.0437840  8.1901235  8.3003780
#>  [778,]  5.2946945  7.0269603  7.3945846  7.6745144
#>  [779,]  8.8574605  9.8118870 10.0564582 10.2285766
#>  [780,]  7.9468890  9.9609037 11.5058639 11.9190910
#>  [781,]  3.3825900  3.6549351  3.8883845  3.9639515
#>  [782,]  3.1674168  3.4331501  3.6185438  3.6207119
#>  [783,]  8.6415852  8.8944565  9.1382032  9.2260482
#>  [784,]  6.9018675  7.7341935  8.1700737  8.2247946
#>  [785,]  9.7979952 11.0489362 11.7321894 11.9387106
#>  [786,]  3.7435268  4.1684535  4.3691634  4.4710793
#>  [787,]  4.6526411  5.5972110  6.0668217  6.2158134
#>  [788,]  3.6226900  4.1844610  4.3388087  4.3763557
#>  [789,]  4.8244589  5.2665233  5.4296835  5.4919310
#>  [790,]  4.5305567  5.3801456  6.0910172  6.5426193
#>  [791,]  6.5745504  9.1019708  9.7569938  9.7960906
#>  [792,]  2.6196406  3.4670096  4.1888747  4.5393133
#>  [793,] 10.3703810 12.2082166 13.8911661 14.7692088
#>  [794,]  8.4020928  9.8833342 10.6055799 10.7294322
#>  [795,]  3.7958862  4.2558878  4.4453353  4.5236453
#>  [796,]  4.9407689  6.2615585  6.6362274  6.6781572
#>  [797,]  2.9121288  3.7600684  3.9505430  4.0055248
#>  [798,]  6.7903988  7.2469629  7.5948694  7.7558479
#>  [799,]  4.7657134  6.1586874  6.2182477  6.2391089
#>  [800,]  1.7398461  2.1243817  2.4797550  2.7295558
#>  [801,]  1.0811123  1.4254038  1.6579579  1.8350910
#>  [802,]  4.4019100  5.3638830  5.7960442  5.9314290
#>  [803,]  4.2309647  4.6336744  4.6553991  4.6636680
#>  [804,]  2.2571808  2.4850906  2.6966360  2.7851922
#>  [805,]  1.8982571  2.0958607  2.2199670  2.3305496
#>  [806,]  2.6496562  3.3915598  4.1261854  4.2297592
#>  [807,]  1.9439328  2.5368582  2.8880989  2.8940600
#>  [808,]  5.0505595  5.4271933  5.4429653  5.4585293
#>  [809,]  2.1172354  2.4233438  2.6811279  2.8466667
#>  [810,]  3.7917221  4.4796719  4.8352516  4.8620691
#>  [811,] 12.5315355 13.2721326 13.8785655 14.2619037
#>  [812,]  4.8136298  5.2001414  5.4094125  5.5723322
#>  [813,]  9.1258734 10.5891483 11.7043945 11.7968388
#>  [814,]  5.3120035  7.2324423  8.5647471  9.0887057
#>  [815,]  1.7933988  1.9745697  2.0571426  2.0624733
#>  [816,]  2.8807454  3.2300464  3.4215658  3.4550423
#>  [817,]  2.2535940  2.3650727  2.4313428  2.4466074
#>  [818,]  5.9754771  6.5399750  7.1031860  7.4065629
#>  [819,]  4.5806669  5.4589904  5.7318690  5.9727762
#>  [820,]  1.1269590  1.2293391  1.3045593  1.3214341
#>  [821,]  0.8898229  1.0556172  1.1705487  1.2254642
#>  [822,]  2.5019017  3.3668804  3.7553394  3.9462307
#>  [823,]  5.6947315  5.8822786  5.9896206  5.9950899
#>  [824,]  3.1859713  4.4107984  4.6526973  4.8603519
#>  [825,] 10.1869001 13.7999225 13.9561835 14.0218633
#>  [826,]  6.7060537  7.8477273  8.6023715  9.0358120
#>  [827,]  3.2437401  3.4704768  3.6860015  3.8464140
#>  [828,]  6.8679913  8.1775330  8.4720483  8.6944098
#>  [829,]  3.2473968  3.8735687  4.2260766  4.3265783
#>  [830,]  1.4496204  1.9491625  2.3797756  2.4217889
#>  [831,]  3.0879709  3.7777767  4.3184684  4.8491003
#>  [832,]  8.1740759 10.1921661 10.8442207 11.0212909
#>  [833,]  8.0283981  9.0973455  9.1580383  9.1952205
#>  [834,]  6.2130292  7.6578609  8.4273248  8.4865596
#>  [835,]  4.0129424  4.4131099  4.7772451  4.8083170
#>  [836,]  4.9559662  6.4125473  7.1955135  7.7124883
#>  [837,]  2.2869165  2.5086033  2.6466712  2.7219086
#>  [838,]  2.2794086  2.6526215  2.8154125  2.9405660
#>  [839,]  3.0717913  3.9971264  4.6446123  5.0515644
#>  [840,]  7.0930746  9.1414042 10.9720940 11.0988689
#>  [841,]  3.3038895  4.0534657  4.1036069  4.1110503
#>  [842,]  7.6042323 11.0277342 12.3124943 12.5145253
#>  [843,]  7.5251837  8.1774994  8.4165675  8.6072803
#>  [844,]  7.3934261  8.0577701  8.2439095  8.3335787
#>  [845,]  0.9327401  1.1814809  1.2987747  1.3036972
#>  [846,]  5.7374427  6.1484753  6.3871102  6.4015767
#>  [847,]  3.2802046  3.4074738  3.5294956  3.6488868
#>  [848,]  2.6912340  4.0035117  4.7321121  5.0653292
#>  [849,]  9.6225385 11.4257947 12.2098814 12.6189639
#>  [850,]  5.0850635  5.6176947  5.9153804  6.0387408
#>  [851,]  1.6617668  1.9282565  2.1439288  2.1571671
#>  [852,]  4.9964291  5.9582841  6.7370135  6.7768431
#>  [853,]  4.6019081  5.7144109  6.3516654  6.9527087
#>  [854,]  1.9054125  2.4346440  2.9501841  3.3160077
#>  [855,]  2.1187916  2.4139350  2.6958996  2.8363467
#>  [856,]  3.6379782  4.0193059  4.2619819  4.4785516
#>  [857,]  3.2599864  4.1956931  4.8676787  5.2985001
#>  [858,] 10.1814581 11.7402395 12.3985163 12.4447928
#>  [859,]  3.4144657  3.8403082  4.1255332  4.1540337
#>  [860,]  1.1535731  1.5994625  1.7683834  1.8754800
#>  [861,]  3.7210202  4.0383454  4.1428334  4.2270838
#>  [862,]  1.0806872  1.4860482  1.5611758  1.6044478
#>  [863,]  1.5068616  1.8212248  1.9691903  2.1136692
#>  [864,]  2.4929064  3.4042898  3.4492940  3.4722630
#>  [865,]  5.6138047  6.4511672  6.6086122  6.6403622
#>  [866,]  7.7248402  9.5535410 10.0949638 10.3129082
#>  [867,]  4.2943779  6.1356852  7.6941018  9.2091714
#>  [868,]  2.9317996  2.9892407  3.0389334  3.0406415
#>  [869,]  0.8878997  1.0905920  1.1000984  1.1036478
#>  [870,]  1.9132416  2.1411577  2.2218079  2.2597959
#>  [871,] 13.2640531 15.1677804 15.3754161 15.5561970
#>  [872,]  4.7972471  6.5707702  6.6459731  6.7150960
#>  [873,]  2.0288074  2.4951341  2.5400423  2.5489893
#>  [874,]  5.0975276  5.5458266  5.9639649  6.0727133
#>  [875,]  6.0315542  6.8449645  7.6411770  8.0410019
#>  [876,]  1.2457186  1.6795920  2.0640245  2.2296051
#>  [877,]  2.6804356  3.1424440  3.4632248  3.7150970
#>  [878,]  8.0976086 11.4355679 14.4296878 15.3669586
#>  [879,]  3.0558536  4.2692213  5.1973781  5.2783668
#>  [880,] 11.5554735 13.5336676 14.3495679 14.3837690
#>  [881,]  9.1627112 11.1755760 12.9635387 13.0581497
#>  [882,]  1.4359644  1.5737402  1.6940764  1.7419890
#>  [883,]  4.4563160  5.0200139  5.4410460  5.5896991
#>  [884,]  3.6295683  4.7253934  4.9296506  5.0649117
#>  [885,]  3.6251648  5.2236989  6.7887058  8.2542531
#>  [886,]  2.8687036  3.4116666  3.5254847  3.6310363
#>  [887,]  4.0317968  4.6688336  5.1466894  5.4051405
#>  [888,]  3.1139701  3.7515965  4.2050720  4.2596267
#>  [889,]  0.5428439  0.7550482  0.8641948  0.9628386
#>  [890,]  3.9357610  4.4530081  4.7898162  4.8316199
#>  [891,]  1.7684970  2.1048507  2.1524533  2.1942821
#>  [892,]  3.4950157  4.1939018  4.7458533  4.9425033
#>  [893,]  2.7681535  3.3701320  3.4324835  3.4891423
#>  [894,]  9.5827890 10.5794060 11.3247600 11.6266147
#>  [895,]  4.3463370  5.4285760  6.1118003  6.4930971
#>  [896,]  5.8307600  7.5000452  7.8790223  7.9241009
#>  [897,]  2.3937549  2.5912595  2.7865551  2.8786661
#>  [898,] 10.8602428 12.6827310 12.7737347 12.8493918
#>  [899,]  5.0369088  5.6445794  5.8196020  5.9754583
#>  [900,]  6.9580137  7.5874240  7.9339491  8.0526320
#>  [901,]  4.7017220  5.9517616  6.8043709  7.2197001
#>  [902,]  6.5707694  7.0658493  7.1537219  7.1760519
#>  [903,]  1.6461852  2.3136613  2.9029363  2.9728576
#>  [904,]  1.9921437  2.1322428  2.1703278  2.1880124
#>  [905,]  4.4159595  5.9973460  6.5635601  6.9323170
#>  [906,]  3.2944456  3.3677476  3.4273596  3.4674372
#>  [907,]  3.6446429  3.8495293  3.9298440  3.9950085
#>  [908,]  3.2074301  3.4443959  3.5928258  3.6584800
#>  [909,]  4.3186910  5.0097567  5.5005995  5.7867372
#>  [910,]  2.6187332  2.9525204  3.1472294  3.2880564
#>  [911,]  3.0418444  3.5346792  3.7770877  3.9113060
#>  [912,]  7.4216230  8.2129682  8.5530379  8.7580887
#>  [913,]  6.6317922  8.3446230  9.0015253  9.1046579
#>  [914,]  3.6602665  4.9260237  5.9018330  5.9876823
#>  [915,]  4.4276191  4.9606716  5.4055664  5.5161687
#>  [916,]  8.3956348  9.7708230 11.0966848 12.3952885
#>  [917,]  4.6167830  5.1596112  5.6687469  5.8141798
#>  [918,]  6.1277424  6.8130426  7.1066413  7.3658699
#>  [919,]  2.7419938  2.8395080  2.9119131  2.9584746
#>  [920,]  6.7763030  8.6062292  9.8698935  9.8925684
#>  [921,]  3.4341335  3.5958244  3.7067428  3.7141628
#>  [922,]  5.7870435  6.4889061  7.1088656  7.5663594
#>  [923,]  5.7180662  7.7601963  9.0196837  9.1253827
#>  [924,]  1.9803626  2.3928702  2.7480938  2.9499175
#>  [925,]  3.4564636  4.1146337  4.3748330  4.5913382
#>  [926,]  2.7044095  2.9931598  3.2003746  3.3864563
#>  [927,]  4.2866626  4.9075043  5.1076264  5.1586210
#>  [928,]  3.3151696  3.6585401  3.9169758  4.0777743
#>  [929,] 10.3596337 10.6948242 11.0290683 11.2388357
#>  [930,] 13.8146401 14.2938660 14.5650684 14.6656842
#>  [931,]  2.1216313  2.5181542  2.6085459  2.6870442
#>  [932,]  2.4255239  2.6362933  2.7097449  2.7631626
#>  [933,]  2.2477281  2.3338185  2.3429164  2.3517273
#>  [934,]  1.0536141  1.0793718  1.0981441  1.1064653
#>  [935,]  2.0084501  2.3133518  2.5422752  2.5493928
#>  [936,]  3.3345445  3.8964493  4.3469655  4.4815120
#>  [937,]  4.8677038  6.2360693  6.8524180  7.0597400
#>  [938,]  2.5464022  3.1970955  3.2615581  3.3218702
#>  [939,]  6.4053645  7.3032863  7.3367235  7.3700801
#>  [940,]  8.4443903 10.7791765 11.5023413 11.9926140
#>  [941,]  1.6867706  2.3361262  2.5826112  2.8248946
#>  [942,]  4.2356138  4.8394153  5.3282997  5.5112676
#>  [943,]  6.3345683  7.8047740  8.1651244  8.2215142
#>  [944,]  3.0498339  3.7131973  3.8099296  3.8941969
#>  [945,]  4.7103683  5.0633997  5.2809847  5.3088299
#>  [946,]  9.0137774 10.4701977 11.0664626 11.6583146
#>  [947,]  3.8669984  5.1677016  5.2726555  5.3699347
#>  [948,]  2.3713455  2.8656921  3.1634156  3.1811765
#>  [949,]  3.9386174  4.3516043  4.5878219  4.7045909
#>  [950,]  2.0141320  2.4243351  2.5761906  2.6189274
#>  [951,] 22.7608670 23.1631168 23.2966254 23.3966707
#>  [952,]  3.6323209  5.1140949  6.1665038  6.2499001
#>  [953,]  7.4622079  8.1859127  8.6810671  8.7700262
#>  [954,]  7.3437234  8.8733485  9.2108213  9.5033192
#>  [955,]  2.4870784  3.1000887  3.2577507  3.2647988
#>  [956,]  1.6003028  1.9007079  2.0736006  2.2125944
#>  [957,]  2.8547442  2.9922552  3.1099959  3.2182865
#>  [958,]  4.9686581  6.6495936  7.6057136  8.5348480
#>  [959,]  5.7858192  7.0741593  7.8126024  8.0136008
#>  [960,] 15.1033061 15.7963971 16.4359646 16.8272267
#>  [961,]  3.9128393  5.2771073  5.6114573  5.7631971
#>  [962,]  7.1529567  7.4301177  7.6258228  7.7971066
#>  [963,]  4.4642916  4.7524985  5.0314557  5.1923424
#>  [964,]  9.4341051 12.5146007 13.0800209 13.6427476
#>  [965,]  3.8297046  4.3717885  4.7908760  4.9996511
#>  [966,]  2.8942250  3.4072317  3.5125900  3.5339106
#>  [967,] 12.5258599 13.4726703 14.2811776 14.9067548
#>  [968,]  1.5687187  2.0330902  2.2793704  2.4221862
#>  [969,]  3.0691680  4.3608005  4.3810104  4.3899000
#>  [970,]  3.4958684  3.7012046  3.8572343  3.9692417
#>  [971,]  1.9968834  2.4943352  2.6779865  2.7684269
#>  [972,]  0.9474587  1.3564323  1.7398770  2.0365502
#>  [973,]  2.4551773  2.8767244  3.0216501  3.0359044
#>  [974,]  5.3258445  6.1390024  6.5624570  6.6132750
#>  [975,]  5.1882563  5.2752018  5.3280040  5.3711407
#>  [976,]  3.9656038  4.9077791  5.6939688  6.1332752
#>  [977,]  5.7490938  6.5556559  6.6634304  6.6956901
#>  [978,]  8.0209704  9.0290311  9.4011482  9.5634502
#>  [979,]  2.7134802  3.0657179  3.3910969  3.5080704
#>  [980,]  4.3490058  5.1055768  5.5829897  6.0401261
#>  [981,]  5.1616929  7.1887872  7.9609591  8.5393432
#>  [982,] 11.6564350 15.6099252 15.8478592 15.9564059
#>  [983,]  5.9433299  6.7129566  7.0683641  7.2615636
#>  [984,]  3.1756243  3.2932041  3.3861623  3.4551733
#>  [985,]  1.8955783  2.2657971  2.3124320  2.3266150
#>  [986,]  1.8083295  2.2351411  2.4352718  2.6275877
#>  [987,]  2.0792440  2.9511767  3.5942043  4.1673104
#>  [988,]  7.1241968  8.9592143  9.4148144  9.7762695
#>  [989,]  1.6639164  2.3244655  2.3476529  2.3573212
#>  [990,]  3.1497291  4.6094687  5.0199402  5.1377107
#>  [991,]  8.5874404 10.6774903 11.1722033 11.2597241
#>  [992,]  3.0611560  4.0346478  4.9976960  5.2311872
#>  [993,]  4.0357760  5.1546773  5.2367162  5.3057621
#>  [994,]  2.6952164  3.1750715  3.6197173  3.7345011
#>  [995,]  7.8447928 10.0273760 11.5731057 11.7917752
#>  [996,]  8.7576172  9.1549094  9.4994703  9.7644568
#>  [997,]  5.0963579  5.4643604  5.6836826  5.7827746
#>  [998,]  6.0512356  7.1865172  7.7853340  7.9531219
#>  [999,]  3.2318068  3.9941193  4.3810354  4.5929066
#> [1000,]  2.5764317  2.7556456  2.8560726  2.9559862