generate_null_distribution_T3

This function generate the null distribution of T_eta, should be run before calculating TWT. Please see Setp 2 in our manuscript for more information.

generate_null_distribution_T3(m, n, cov_mat, cutoff_value)

Arguments

m

Dimension.

n

Number of the generated sample size.

cov_mat

Covariance matrix for generating random samples.

cutoff_value

Truncated value, should be in 0 to 1 and must contain 1.

Value

Null distribution matrix with number of rows equal n and number of columns equal length of cutoff value minus 1.

Examples

ThreeWayTest::generate_null_distribution_T3(m=6,n=1000,
cov_mat=diag(nrow = 6, ncol = 6), cutoff_value=c(0.2,0.4,0.6,0.8,1))
#>               [,1]       [,2]       [,3]       [,4]
#>    [1,] 12.5526738 13.8368821 14.5593030 14.8559705
#>    [2,]  1.7863762  2.0380575  2.1962468  2.2909566
#>    [3,]  2.0890830  2.3975297  2.6911724  2.9026040
#>    [4,]  4.3058028  4.7585378  5.1858253  5.5199010
#>    [5,]  3.1266166  3.2192908  3.2759670  3.2982679
#>    [6,]  5.6046470  6.4336270  7.2500802  7.3587669
#>    [7,]  8.6330888  9.9032661 10.5313709 10.6465983
#>    [8,]  4.2506037  4.8842731  5.2886021  5.6465966
#>    [9,]  7.5679651  8.3545720  8.4877889  8.5367174
#>   [10,]  4.1598183  5.1900301  5.5674388  5.6885479
#>   [11,]  3.2064967  3.5424537  3.7055723  3.7770699
#>   [12,]  4.7872659  6.4629709  7.0497520  7.0529334
#>   [13,]  8.8754804  9.5485650  9.9927151 10.0006729
#>   [14,]  3.4490238  3.7466288  4.0216744  4.2767346
#>   [15,]  3.3357693  4.4676736  5.2867428  5.6512615
#>   [16,]  1.7447046  1.9482491  1.9659409  1.9670431
#>   [17,]  3.4975872  4.0044848  4.3483185  4.3517287
#>   [18,]  3.2653416  4.0677220  4.7270626  4.8056500
#>   [19,]  2.4363696  3.1510392  3.3367587  3.5205647
#>   [20,]  2.5210159  2.9650869  3.1943153  3.2918407
#>   [21,]  2.2317969  2.4171896  2.5773132  2.6838710
#>   [22,]  1.4548982  2.0606964  2.5456038  2.6119514
#>   [23,]  2.0894783  2.6638622  2.9654383  2.9719472
#>   [24,]  2.4305642  2.8728220  3.0242551  3.0618830
#>   [25,]  4.8829416  6.4223103  7.6593963  8.1400461
#>   [26,]  1.3786636  1.5217176  1.6243920  1.7137744
#>   [27,]  7.0875210  7.7739522  8.0112322  8.1858293
#>   [28,]  2.9849431  3.5921881  3.8680246  4.0610997
#>   [29,]  4.7197884  5.7683943  5.9071244  5.9891599
#>   [30,]  5.6889862  6.9598411  7.7925958  7.9017266
#>   [31,]  3.0548337  3.6204482  3.9999839  4.2817134
#>   [32,] 10.3670282 13.2478239 15.1034428 15.5690849
#>   [33,]  4.6287189  5.3107330  5.3334331  5.3521239
#>   [34,]  5.1488493  6.6606769  7.0606694  7.3337667
#>   [35,]  5.2762413  6.2380250  6.6832833  6.9227809
#>   [36,]  0.7867041  0.9598235  1.0239182  1.0768035
#>   [37,]  3.0809489  3.4979044  3.6129187  3.7012024
#>   [38,]  2.8136167  3.5010451  4.0728107  4.3141982
#>   [39,]  6.7823718  7.7015462  8.0944282  8.3469851
#>   [40,]  6.9059573  9.4149067 11.8698026 12.2758237
#>   [41,]  1.2995267  1.4381592  1.5368083  1.5440409
#>   [42,]  4.9640183  6.0651651  7.1263008  7.6569641
#>   [43,]  1.5804808  2.0157526  2.3572090  2.5256345
#>   [44,]  3.0495400  3.7645260  4.3358940  4.7635952
#>   [45,]  6.8163516  8.3484676  9.3590867  9.8001516
#>   [46,]  4.5617154  6.2417552  6.5668834  6.7863198
#>   [47,]  5.4547273  7.1651592  7.4232810  7.4412402
#>   [48,]  4.3419275  5.5283424  5.8967490  6.2641910
#>   [49,]  3.3529886  3.9394846  4.1360164  4.2117283
#>   [50,]  3.9659124  4.7488105  5.4456338  5.9460122
#>   [51,]  5.5979507  6.0047844  6.0194313  6.0258867
#>   [52,]  9.1308732 11.2020982 12.2945106 12.9225275
#>   [53,]  1.7191866  2.2773292  2.6114586  2.9149663
#>   [54,]  8.1873708  9.7481687 11.2545809 11.5703291
#>   [55,]  4.7804044  5.4378085  5.8185140  5.8825594
#>   [56,]  3.3654397  4.2812713  4.4937715  4.6236571
#>   [57,]  1.5018107  2.1594449  2.6953929  3.2068800
#>   [58,]  6.5611716  7.1532733  7.6969796  7.7553578
#>   [59,]  1.1148037  1.2241525  1.2354091  1.2360744
#>   [60,] 11.2297321 12.4182522 12.7990832 12.8273814
#>   [61,]  4.1615951  5.1611355  5.9124058  6.1104405
#>   [62,]  3.5159072  4.7080428  5.1777499  5.4181027
#>   [63,]  2.1487229  2.2565773  2.2837325  2.2941712
#>   [64,]  2.9249718  3.1640618  3.3063827  3.4237777
#>   [65,]  2.8451468  3.6700721  4.0898393  4.2730835
#>   [66,]  2.9296761  3.4684088  3.6494261  3.7088339
#>   [67,]  1.8792670  2.1815531  2.3043190  2.3517070
#>   [68,]  6.7197600  6.9041933  7.0533292  7.1855621
#>   [69,] 17.2520889 18.7433277 19.6365541 19.7833927
#>   [70,]  3.2316573  3.3952241  3.4684998  3.5082900
#>   [71,]  6.6924540  7.8580726  8.6631017  8.8574415
#>   [72,]  6.9960234  9.1003912  9.5117758  9.6168631
#>   [73,] 12.8351648 14.2846564 14.3983611 14.4772174
#>   [74,]  3.6359142  4.7580815  5.4823394  5.6052838
#>   [75,]  0.9755376  1.2773833  1.4590431  1.4902244
#>   [76,]  5.5877770  6.5880484  7.5567981  7.5810042
#>   [77,]  2.2963416  3.1017675  3.5135568  3.7630464
#>   [78,]  2.6893925  3.6160893  4.1994165  4.5542468
#>   [79,]  1.2303367  1.2471712  1.2547298  1.2554928
#>   [80,]  1.8071734  1.8507639  1.8522928  1.8524239
#>   [81,]  2.9819007  4.3069318  5.4236849  5.6491224
#>   [82,]  2.2516177  2.8714926  3.0573631  3.2353656
#>   [83,]  1.7075616  2.1622948  2.3572979  2.5476815
#>   [84,]  4.3809287  5.0198100  5.2352780  5.4243496
#>   [85,]  2.5722455  3.4918666  4.3264715  4.3600027
#>   [86,]  2.5127354  2.8928090  3.1541067  3.3456446
#>   [87,]  7.2924628  8.7477523  9.8275336  9.8461078
#>   [88,]  6.8331895  7.6936722  8.3023346  8.6506402
#>   [89,]  2.6142427  3.4283674  3.8306440  3.9689366
#>   [90,]  1.1955880  1.2603701  1.3058608  1.3337666
#>   [91,]  3.7643314  4.1576085  4.3352019  4.4946157
#>   [92,]  8.5772576  9.7543241 10.0917310 10.1737683
#>   [93,]  7.1124560  7.4212597  7.7113470  7.7856847
#>   [94,]  5.8576579  6.1052696  6.2421695  6.2548949
#>   [95,]  5.7120446  7.5859064  8.0776873  8.4924515
#>   [96,]  4.2538999  4.6160569  4.6635474  4.6877149
#>   [97,]  3.6563995  4.9142194  5.8615378  6.2108573
#>   [98,]  5.2740725  7.2518966  7.9943390  8.5407359
#>   [99,]  4.1233251  4.6246360  4.8994293  4.9472869
#>  [100,]  2.5026751  3.1291545  3.1756719  3.1820987
#>  [101,]  2.9999366  3.4496764  3.7555683  3.7785292
#>  [102,]  1.7713078  2.0048148  2.1374932  2.1667171
#>  [103,]  4.2595237  5.1324880  5.1766738  5.1844358
#>  [104,]  4.2151732  5.2942224  5.9881016  6.3951340
#>  [105,]  3.4929395  3.9641118  4.1407370  4.2928693
#>  [106,]  9.0061922 10.5111888 10.9009438 11.1274875
#>  [107,]  8.2701474  9.5142291  9.7709408  9.9713767
#>  [108,]  4.6038982  4.8764444  4.9952825  5.0768324
#>  [109,] 10.0823568 10.7799282 11.3083662 11.5819070
#>  [110,]  3.5740048  4.9269372  4.9931377  5.0116522
#>  [111,]  3.4498957  4.2967883  4.3903862  4.4150502
#>  [112,] 10.2590463 11.6215271 11.9788327 12.0063885
#>  [113,]  3.5313517  3.9205790  4.1591492  4.3367332
#>  [114,]  5.1782508  6.5040842  6.6441110  6.7064281
#>  [115,]  4.8003032  5.9685616  6.3653134  6.6920595
#>  [116,]  3.9045305  4.9397955  5.5832748  5.6206033
#>  [117,]  2.8505697  3.7448672  4.4642562  4.9942859
#>  [118,]  4.8648435  5.9687447  6.9055359  7.3293361
#>  [119,]  2.9247692  3.7313545  4.2944796  4.7639640
#>  [120,]  2.0962479  2.4860607  2.8112384  2.8403608
#>  [121,]  3.8149452  4.2522499  4.3348851  4.3521663
#>  [122,]  1.6889501  1.8517845  1.9502866  2.0056303
#>  [123,]  2.2988227  2.8254447  2.8547540  2.8810228
#>  [124,]  3.5266438  4.0415033  4.5363574  4.7335632
#>  [125,]  4.7218939  5.9224823  6.3840448  6.8253127
#>  [126,]  3.0783504  3.3980512  3.5499893  3.6512869
#>  [127,]  1.2291066  1.3246712  1.4008539  1.4113510
#>  [128,]  3.8987044  4.6000842  4.7785179  4.8411014
#>  [129,]  6.6128078  7.1375502  7.2330612  7.2931056
#>  [130,]  0.5569965  0.6150896  0.6646613  0.7065652
#>  [131,]  1.6584407  2.0441953  2.3641570  2.4295577
#>  [132,] 13.6645739 14.8946073 15.3437094 15.5496412
#>  [133,]  4.5262789  5.3837472  5.6371369  5.7661692
#>  [134,]  4.5053097  4.7042959  4.8501238  4.8734597
#>  [135,]  3.3911655  3.8361123  4.1976309  4.3477182
#>  [136,] 10.9743764 12.5550950 13.4697654 13.6976160
#>  [137,]  3.6959314  3.8688276  3.9266288  3.9347250
#>  [138,]  4.8332973  5.3317448  5.4842232  5.5525715
#>  [139,]  2.9076514  3.6077796  3.9401208  4.0093520
#>  [140,]  2.6738599  3.3670768  3.9302047  4.4177301
#>  [141,]  5.3125036  5.6322311  5.8590612  5.9771312
#>  [142,]  3.0692833  3.3165337  3.4480044  3.5658351
#>  [143,]  6.7640786  8.2430067  9.5406516  9.7561608
#>  [144,]  5.3014322  6.7636544  6.8683122  6.8908644
#>  [145,]  3.6467219  4.6031835  4.8841815  5.0280328
#>  [146,]  6.3596619  7.1842574  7.9457974  8.5671027
#>  [147,]  4.5454371  5.1041103  5.3672168  5.4372641
#>  [148,]  3.8342364  4.5427068  5.1275974  5.2425592
#>  [149,]  1.2263635  1.3759170  1.3807569  1.3813634
#>  [150,]  7.3051230  8.3010473  9.2808298  9.5911998
#>  [151,]  6.2111290  6.7106303  7.1684162  7.3753612
#>  [152,]  2.2392264  2.4382422  2.5045839  2.5606074
#>  [153,]  6.6243250  7.4441416  8.1226954  8.3430562
#>  [154,]  2.0841646  2.2819924  2.3898611  2.4841869
#>  [155,]  5.2748790  5.8090337  6.0240713  6.1969647
#>  [156,]  1.8059317  2.4342363  3.0116630  3.2058554
#>  [157,]  5.0127984  5.2821298  5.3524786  5.3836648
#>  [158,]  2.7276886  3.1537193  3.5447428  3.6432210
#>  [159,]  2.1786363  3.0172139  3.5641041  3.9111910
#>  [160,]  5.3202816  6.1824266  6.5131064  6.6543844
#>  [161,]  4.6647646  6.0707377  7.1738008  7.2510015
#>  [162,]  4.2888884  5.0773314  5.6944504  5.8042090
#>  [163,]  2.1998093  2.5051759  2.6345666  2.7048414
#>  [164,] 13.1205198 16.1810144 18.8997110 19.2600210
#>  [165,]  8.8370158  9.5686661  9.8895135  9.9113700
#>  [166,]  3.2470312  3.6054973  3.9164187  4.0930968
#>  [167,]  1.3456461  1.6875821  1.9687317  2.1210025
#>  [168,]  3.6018626  4.6356523  5.3175525  5.9266095
#>  [169,]  1.4946526  1.9698336  2.4113364  2.7634137
#>  [170,]  2.1832953  2.3965126  2.5347239  2.6126472
#>  [171,]  3.8024158  5.0517998  6.0200328  6.2928092
#>  [172,]  7.0268541  8.3629833  8.6001311  8.6697829
#>  [173,]  3.8422287  4.9577020  5.3596220  5.7570310
#>  [174,]  2.1987630  2.6685385  3.0658626  3.1708834
#>  [175,]  4.2672709  5.4469948  5.8619215  6.2305746
#>  [176,]  5.1617040  5.3872850  5.4310240  5.4384284
#>  [177,]  7.3672640  7.6022086  7.8331237  7.9134865
#>  [178,]  4.7235895  6.4632024  7.6816208  8.6966631
#>  [179,]  4.6336238  4.8110254  4.9114588  4.9407070
#>  [180,]  3.5334323  3.9647796  4.2412864  4.3466842
#>  [181,]  2.4803484  3.0121837  3.4803291  3.9292903
#>  [182,] 10.2902323 13.1605587 14.1752277 14.4673644
#>  [183,]  5.5944452  6.5160601  6.8821008  6.9645561
#>  [184,]  3.0616996  3.7637110  4.3926120  4.5767100
#>  [185,]  3.8764243  4.4077225  4.5565654  4.5583476
#>  [186,]  3.7265453  3.8630631  3.9981384  4.0440404
#>  [187,]  4.6019843  4.7065776  4.7885027  4.8398414
#>  [188,]  2.7486283  3.5235076  3.8105410  4.0346740
#>  [189,]  1.9780531  2.3297957  2.4124377  2.4813923
#>  [190,]  2.8860760  4.1949326  4.6770467  4.7369339
#>  [191,]  4.1211895  5.6610371  6.2529046  6.3155374
#>  [192,]  7.0029213  8.6365257  8.9123475  9.0225010
#>  [193,]  6.0860054  6.4338917  6.5620331  6.6706346
#>  [194,]  4.7372203  5.3955975  5.8293467  6.2171163
#>  [195,]  2.4190109  2.6693257  2.7721197  2.8442317
#>  [196,]  2.0267618  2.7554503  3.2776111  3.5496815
#>  [197,]  2.6596853  2.7885826  2.9088670  3.0181166
#>  [198,]  2.3144783  2.9407723  3.5305511  3.8933851
#>  [199,] 13.0477868 17.3243684 18.3090065 18.6514412
#>  [200,]  1.7202284  1.9605054  1.9818212  1.9821037
#>  [201,]  8.7118716  9.0753412  9.1995487  9.2091028
#>  [202,]  5.2940878  6.1625413  6.4409754  6.6563569
#>  [203,]  3.4098434  3.5746098  3.7219392  3.8390443
#>  [204,]  6.0805473  7.3421455  8.4571683  9.1686665
#>  [205,]  2.0982539  2.5304543  2.8317561  2.8697484
#>  [206,]  3.3295322  4.0361882  4.3872305  4.5135104
#>  [207,]  0.8129788  0.9916763  1.0384379  1.0696410
#>  [208,]  3.0231167  4.0342024  4.4708361  4.8032117
#>  [209,]  2.0786487  2.6475153  3.2008227  3.6927826
#>  [210,]  5.3360775  6.2691947  7.0664804  7.3506574
#>  [211,] 11.1097506 11.5281236 11.6141394 11.6582396
#>  [212,]  2.6675987  2.9830718  3.1747135  3.3018460
#>  [213,]  2.0069820  2.3155018  2.5235735  2.5639159
#>  [214,]  2.3877328  3.0744549  3.6623304  4.2360117
#>  [215,]  2.4382630  2.6848153  2.8855376  3.0583071
#>  [216,]  0.9759056  1.0196628  1.0363416  1.0372007
#>  [217,]  3.7557739  4.9271932  5.5413892  5.8075731
#>  [218,]  5.5090516  5.8339846  6.1554152  6.4448707
#>  [219,]  2.6456797  3.0013462  3.2025632  3.2373953
#>  [220,]  3.9892146  5.3047009  5.9464001  6.1341511
#>  [221,]  4.6131306  5.1647493  5.6368641  5.8150447
#>  [222,]  1.9905653  2.3798504  2.4586877  2.5045455
#>  [223,]  1.7174968  1.8876201  1.9836424  2.0045184
#>  [224,]  5.9590978  6.9706065  7.9569072  8.1007490
#>  [225,]  2.2128483  2.6052025  2.9366838  3.1460516
#>  [226,]  4.6833499  5.8568938  6.5884909  6.6000746
#>  [227,]  5.2848938  5.9443662  6.1581880  6.1988977
#>  [228,]  1.0404550  1.2799626  1.4050382  1.4255446
#>  [229,]  4.6183330  4.9652875  5.2572944  5.3863778
#>  [230,]  5.8834320  7.7529044  8.6391475  9.1639438
#>  [231,]  2.9543496  3.4467965  3.8124363  3.8396944
#>  [232,]  8.1896432  9.6944967 10.0354621 10.0475159
#>  [233,] 10.4188579 11.4980981 11.6217143 11.6280545
#>  [234,]  4.8085807  5.4439893  5.6434127  5.6579953
#>  [235,]  3.3150276  3.5830966  3.6105456  3.6259102
#>  [236,]  9.2104527 10.8515452 11.9512611 12.6671252
#>  [237,]  3.5156027  3.6743845  3.7017851  3.7233951
#>  [238,]  6.5138024  9.0817228  9.7989799 10.0891915
#>  [239,]  3.4520000  5.0100399  5.4653498  5.6244298
#>  [240,]  4.5005708  5.0489013  5.4818197  5.6385527
#>  [241,]  2.2680618  3.0238956  3.2992351  3.3518775
#>  [242,]  1.7048726  2.1820354  2.2375420  2.2852694
#>  [243,]  1.5809113  2.0006248  2.2990426  2.4625279
#>  [244,]  5.4182182  6.7628168  7.8742649  8.4441558
#>  [245,]  2.0777332  3.0102605  3.3730307  3.5081747
#>  [246,]  4.4201973  5.2057529  5.7456129  5.9270090
#>  [247,]  2.2855691  3.1645035  3.9167645  4.4370870
#>  [248,]  4.5476994  6.3467017  7.2033293  7.2324372
#>  [249,]  5.6835033  6.3064467  6.6881736  6.7972980
#>  [250,]  6.3652183  7.1411317  7.7488110  8.2661789
#>  [251,]  1.9152598  2.1588986  2.3942999  2.6207264
#>  [252,]  4.2703656  5.7422822  6.4731422  6.9623072
#>  [253,]  7.7415051  8.8530139  9.2941300  9.3858711
#>  [254,]  2.2359096  2.5345599  2.7498829  2.7630257
#>  [255,]  5.1113779  5.3745593  5.4308750  5.4580757
#>  [256,]  1.7535457  1.7830583  1.8101143  1.8292328
#>  [257,]  5.5185400  5.5682004  5.5761293  5.5821772
#>  [258,] 11.0081155 11.4555455 11.7026961 11.7444183
#>  [259,]  3.2123559  4.2879876  4.6004971  4.6374017
#>  [260,] 13.8812485 15.9363814 17.0414748 17.3007981
#>  [261,]  1.9419000  2.0947757  2.1554063  2.2008067
#>  [262,]  2.4671154  2.7388151  2.7763397  2.7984648
#>  [263,]  3.1706308  3.5026538  3.8180027  4.1098136
#>  [264,]  6.1540126  7.0553669  7.4661690  7.5823163
#>  [265,]  4.3359990  5.8894174  7.0798206  7.8672002
#>  [266,]  1.6740269  2.1251101  2.3815074  2.4005427
#>  [267,]  3.9168103  4.3712763  4.7357344  4.7460908
#>  [268,]  3.0099283  3.4429254  3.4539512  3.4623049
#>  [269,]  6.3930688  7.3125453  7.5657822  7.6035219
#>  [270,]  0.7861381  0.9214960  1.0398934  1.0612489
#>  [271,]  6.9577498  8.6095994 10.2129803 11.2436589
#>  [272,]  4.8226783  5.7306434  6.4806079  6.8892089
#>  [273,]  0.3487075  0.4183005  0.4680019  0.4881020
#>  [274,]  2.8422360  4.2375128  4.9064617  5.4437689
#>  [275,]  0.6023502  0.7411140  0.7809372  0.8015998
#>  [276,]  3.5789526  5.1391562  6.5862342  7.1181471
#>  [277,]  1.1964095  1.5536615  1.8629017  2.1409738
#>  [278,]  4.0742079  5.6299131  6.4659100  6.6771101
#>  [279,]  7.4168055  8.3325412  8.4183376  8.4699442
#>  [280,]  2.6582242  3.1620499  3.6196067  4.0091254
#>  [281,]  1.7032743  2.0904259  2.2521284  2.3867036
#>  [282,]  4.2999960  5.8631193  6.8529221  7.2682291
#>  [283,]  2.4816265  3.0674329  3.3982011  3.4887966
#>  [284,]  3.3348452  4.4343136  5.2844939  6.1306569
#>  [285,]  6.9048975  8.7384866  8.8678982  8.8721664
#>  [286,]  2.6650900  3.4264532  3.7756711  4.1122736
#>  [287,]  5.4653540  6.7483655  6.9877452  7.0344291
#>  [288,]  4.5149920  5.5323698  5.9225968  6.2546242
#>  [289,]  0.3871406  0.5112469  0.6170594  0.6688646
#>  [290,]  4.0507948  4.4994236  4.6249872  4.6954247
#>  [291,] 10.0763906 10.2910704 10.4067165 10.5061419
#>  [292,]  4.8342739  4.8918049  4.9093228  4.9257363
#>  [293,]  6.4042338  7.7962887  8.7155389  9.1881173
#>  [294,]  2.8958779  3.1330623  3.3269012  3.4545754
#>  [295,]  5.1763805  7.1075288  7.3053081  7.3523979
#>  [296,]  2.9201743  3.3169971  3.5005776  3.5982797
#>  [297,]  3.7138074  4.4731806  5.0715462  5.5082136
#>  [298,]  4.2209039  5.9303012  6.2780370  6.3496718
#>  [299,]  5.2231666  6.6699431  7.1765180  7.2678977
#>  [300,]  4.1112357  5.3635225  6.1561959  6.8374889
#>  [301,]  2.2846765  2.3956446  2.4348200  2.4426829
#>  [302,] 15.1073867 17.8601008 18.0417938 18.1767661
#>  [303,]  4.1840164  5.1484108  5.9554884  6.5394083
#>  [304,]  2.9966352  3.4295595  3.5935436  3.7332293
#>  [305,]  7.0007807  7.4705311  7.9272187  8.3584408
#>  [306,]  0.3841119  0.4749587  0.5478183  0.6123117
#>  [307,]  2.1795815  2.2517429  2.2858870  2.2910512
#>  [308,] 12.3647135 14.1171945 14.4034530 14.6260993
#>  [309,]  4.0808613  5.4162580  6.3339967  7.1442225
#>  [310,]  1.6899555  2.2404711  2.6284474  2.6632901
#>  [311,]  3.3623252  3.6121578  3.8132784  3.8535234
#>  [312,]  2.4714423  3.1476087  3.4728485  3.7732511
#>  [313,]  4.6361895  4.8315723  4.9844628  5.1018327
#>  [314,]  4.2719676  5.3834219  5.8510385  6.2495630
#>  [315,]  2.6403117  3.0158474  3.2101507  3.2972398
#>  [316,]  5.1682193  5.5329348  5.8655528  5.9090268
#>  [317,]  5.5476960  5.8053264  5.9547611  6.0187147
#>  [318,]  1.9003345  2.1467257  2.2508645  2.2552826
#>  [319,]  2.4005117  3.1643591  3.3522815  3.4875180
#>  [320,]  0.4674386  0.5549757  0.6329135  0.6943040
#>  [321,]  3.6015614  4.1800841  4.5556594  4.8579497
#>  [322,]  1.5354738  1.8714594  2.0638121  2.0909474
#>  [323,]  5.2639853  5.7456700  6.1405289  6.4374073
#>  [324,]  7.3976666  7.5120936  7.5491964  7.5603136
#>  [325,]  5.6477819  7.2431715  8.8323666 10.0169482
#>  [326,]  4.4233095  4.9947822  5.3846337  5.6732239
#>  [327,]  7.3605492  7.9313381  8.4425207  8.5107382
#>  [328,]  1.6933846  2.3048513  2.8475572  2.9278235
#>  [329,]  3.6309866  3.9870214  4.1985465  4.3096897
#>  [330,]  5.0954082  5.9847121  6.2933204  6.4513346
#>  [331,]  4.6150187  5.2116749  5.3092734  5.3722073
#>  [332,]  8.2148115  9.0690672  9.3401568  9.4153357
#>  [333,]  8.0314555  8.5709083  8.8942634  8.9338303
#>  [334,]  2.0965232  2.6678513  3.0100966  3.2852600
#>  [335,]  1.8755569  2.1695272  2.1738224  2.1765124
#>  [336,]  2.2438983  2.6013492  2.6458720  2.6577865
#>  [337,]  3.3088389  3.8632311  4.3317780  4.5427859
#>  [338,]  1.4519424  1.6314620  1.8089654  1.8123750
#>  [339,]  2.9762907  4.0995807  4.9279533  5.1933288
#>  [340,]  3.3880109  3.6310119  3.7910784  3.8795333
#>  [341,]  4.0876794  4.9138896  5.5268270  5.8014660
#>  [342,]  7.4821397  8.9686186  9.4493347  9.6575460
#>  [343,]  2.4621901  2.9260386  3.2632038  3.4648547
#>  [344,]  6.2506206  7.4559242  7.5613574  7.6096344
#>  [345,]  3.0383352  3.6493313  3.9210163  4.1915358
#>  [346,]  1.4714363  2.0409711  2.4124358  2.4485446
#>  [347,]  1.4641539  1.8274815  2.1736396  2.4459663
#>  [348,]  2.4650282  3.2655375  3.8108516  4.3491548
#>  [349,]  5.6591314  6.6748368  7.5125617  7.5604378
#>  [350,]  8.7413040 10.9019402 11.5497009 12.0280140
#>  [351,]  6.9676666  7.6099239  7.9057101  7.9969714
#>  [352,]  9.5275406 10.8260219 11.2290017 11.4361994
#>  [353,]  2.4473917  2.9733484  3.3037561  3.6182468
#>  [354,]  2.2421439  3.1495792  3.6630240  3.8574012
#>  [355,]  4.3858463  5.4839847  6.3045661  6.5408864
#>  [356,]  2.0373181  2.5527405  2.7508213  2.9292822
#>  [357,]  0.7705406  0.9081110  1.0159430  1.0412529
#>  [358,]  4.5079338  5.4877185  6.2308143  6.3706642
#>  [359,]  7.1672877  7.4338843  7.5030970  7.5415390
#>  [360,]  2.2337672  2.7715900  3.1863911  3.4771108
#>  [361,]  1.8049200  2.0899282  2.1233858  2.1306590
#>  [362,]  3.3178274  4.0300096  4.5041260  4.6204053
#>  [363,]  2.5059024  2.7438364  2.9517345  3.1342814
#>  [364,]  3.4823931  4.6968350  4.9535648  5.0474982
#>  [365,]  5.5175198  7.3924274  8.6680929  9.8267008
#>  [366,]  5.7017944  5.9352967  6.0247051  6.0635879
#>  [367,]  5.7853733  7.0605428  7.1965092  7.2621457
#>  [368,]  7.8571645  9.5578849 11.0123938 11.3840101
#>  [369,]  3.7014343  4.9635982  5.4707137  5.9585020
#>  [370,]  3.9129332  5.1837838  6.0081090  6.6138125
#>  [371,]  3.0239042  3.0467625  3.0474208  3.0474582
#>  [372,]  7.6878995  8.1081408  8.3428163  8.5030186
#>  [373,]  9.2441355 10.7778664 12.1524052 12.8708022
#>  [374,]  2.2665220  2.6056388  2.6705781  2.7147242
#>  [375,]  7.0411075  7.4568483  7.8539741  7.9413959
#>  [376,]  1.7275110  2.0283329  2.1013608  2.1334714
#>  [377,]  2.8855863  3.4324487  3.6218158  3.6818722
#>  [378,]  9.6281569 11.8503096 13.3948251 14.0707491
#>  [379,]  1.2088258  1.5332385  1.5872653  1.6245610
#>  [380,]  4.9457893  5.4496374  5.8614867  6.0953611
#>  [381,]  9.3313801 10.8390759 12.0646498 12.3720349
#>  [382,]  5.4147853  6.5070259  7.3440777  7.6437511
#>  [383,]  4.8368389  6.6704491  8.0846980  8.7949680
#>  [384,]  3.1449545  4.3694190  4.7708070  4.7911387
#>  [385,] 12.0158227 12.9045153 13.2780794 13.3893295
#>  [386,]  5.5301166  7.8648035  8.6472401  8.8151357
#>  [387,]  5.3090119  6.6479823  7.7977035  7.9125515
#>  [388,]  4.0689099  5.0085914  5.8659002  6.1567114
#>  [389,]  7.8209481  8.3772475  8.8145686  8.9625380
#>  [390,]  9.0794146 10.1625767 10.3478821 10.3576188
#>  [391,]  3.6263143  4.4084998  4.9575873  5.0283087
#>  [392,]  3.5513304  4.3259279  4.3685672  4.3766984
#>  [393,]  4.9870165  7.2350855  8.1172860  8.2189923
#>  [394,]  1.1449408  1.4391982  1.5655754  1.6727474
#>  [395,]  1.8538947  2.5411204  2.6105554  2.6562299
#>  [396,]  1.6531760  1.8225579  1.9374317  1.9772749
#>  [397,]  3.8036204  4.2443261  4.4014067  4.4616514
#>  [398,]  7.7454113  7.9402350  7.9845887  8.0110287
#>  [399,]  6.2576872  8.1089603  8.5745367  8.9567719
#>  [400,]  3.9495749  4.9709602  5.4614261  5.8915584
#>  [401,]  2.5714922  2.8781003  3.1721967  3.1915355
#>  [402,]  6.0010022  6.8961337  7.3689520  7.8025959
#>  [403,]  1.1711127  1.3045432  1.3951838  1.4637615
#>  [404,]  3.2791542  3.9859723  4.6467319  5.2126827
#>  [405,]  3.6534457  4.0054945  4.0311365  4.0483917
#>  [406,] 15.3958193 18.8947013 19.6500127 20.3544261
#>  [407,]  1.8290315  2.4035346  2.8339003  2.8421289
#>  [408,]  9.9956202 10.9577003 11.8277936 12.0948066
#>  [409,]  4.3221283  4.7788494  5.1779089  5.1851680
#>  [410,]  9.8893657 13.3382107 13.8898304 14.2881001
#>  [411,]  4.8799995  5.0615228  5.1713033  5.2074698
#>  [412,]  6.3361077  9.1364805 11.6623998 12.5959668
#>  [413,]  2.4082319  2.8234789  3.1020439  3.2843851
#>  [414,]  9.6209263 12.9431948 13.0038139 13.0203960
#>  [415,]  2.0980130  2.1232658  2.1461230  2.1533910
#>  [416,]  2.7072264  3.5828430  3.8546057  3.9863283
#>  [417,]  2.1660844  2.4605352  2.6878924  2.7103308
#>  [418,]  1.6941786  2.0628529  2.1656313  2.2456961
#>  [419,]  6.4548139  7.9031260  8.7709164  9.4444460
#>  [420,]  2.8655859  3.5123331  4.0971702  4.2646313
#>  [421,]  4.1351320  4.3809387  4.5868781  4.7548064
#>  [422,]  5.7119560  6.9862071  7.6940575  7.8274904
#>  [423,]  5.4564571  7.0392498  7.8077358  8.2218764
#>  [424,]  9.1417010 11.1323770 11.5708081 11.8385622
#>  [425,]  6.5872638  6.9518600  7.2242408  7.4603348
#>  [426,]  2.6639242  3.1979772  3.7151570  4.1016846
#>  [427,]  5.1900691  5.4749667  5.7429641  5.8105869
#>  [428,]  2.8583084  3.3223850  3.5587368  3.7110660
#>  [429,]  2.0086027  2.3684392  2.6082000  2.7479083
#>  [430,]  4.1717810  5.0485392  5.4504657  5.8134962
#>  [431,]  4.7300641  5.7751516  6.6876410  7.3338915
#>  [432,]  3.4628041  4.1015480  4.1553392  4.2058848
#>  [433,]  7.3720175  8.6836247  9.7923007 10.4497712
#>  [434,]  4.8616673  6.1915116  6.2736707  6.3305940
#>  [435,]  4.5270294  5.8807148  7.1564149  7.8555294
#>  [436,] 10.7003598 12.1530466 13.0402839 13.4364006
#>  [437,]  1.3036078  1.7459866  2.1353692  2.4786603
#>  [438,]  1.0573452  1.1638024  1.2441274  1.3123321
#>  [439,]  3.7035851  4.0405728  4.2353656  4.3525210
#>  [440,]  4.1079092  4.7665373  5.1225675  5.3020167
#>  [441,]  8.6350432  8.9060577  9.0714063  9.1594255
#>  [442,]  6.7829867  8.1271063  9.0443322  9.0647529
#>  [443,]  3.9751868  5.5726010  6.6887320  6.9103824
#>  [444,]  1.4014274  1.6878757  1.8189557  1.8198162
#>  [445,]  1.4086317  1.6636323  1.9013908  2.0975567
#>  [446,]  9.1240454  9.9214487 10.0774081 10.1754367
#>  [447,]  5.9194042  6.4167576  6.8575596  6.9437941
#>  [448,]  6.5877761  8.3269293  8.7698337  8.8444008
#>  [449,]  8.1584953  9.2372704  9.4812339  9.6128501
#>  [450,]  2.2559187  2.3748236  2.4526432  2.4946741
#>  [451,]  2.3893829  2.7206593  2.8979574  2.9531099
#>  [452,]  3.0580559  3.5238666  3.7378986  3.8526653
#>  [453,]  3.6772588  4.1495351  4.4294498  4.5849737
#>  [454,]  6.1715926  7.0577961  7.4743059  7.5353253
#>  [455,]  3.8369112  4.5096257  4.9488378  5.2601623
#>  [456,] 10.0416468 10.4595301 10.8494142 10.8702331
#>  [457,]  6.5256759  7.1575592  7.5256246  7.5533408
#>  [458,]  3.2501974  4.2676823  4.3503694  4.4028649
#>  [459,]  2.3500304  3.3903856  3.9353941  4.3199837
#>  [460,]  3.7917092  4.3388927  4.7851929  4.8830882
#>  [461,]  4.8616351  5.3330609  5.4383528  5.4714411
#>  [462,]  3.0971680  3.6146961  3.8356398  4.0117491
#>  [463,]  2.6156008  2.8205286  2.9213569  2.9642546
#>  [464,]  1.8583413  2.4365122  2.4466957  2.4551523
#>  [465,]  6.1912025  6.8195694  7.4335540  8.0189487
#>  [466,]  4.1112047  4.2829853  4.3185330  4.3515755
#>  [467,]  6.5589696  7.9434340  8.1509002  8.2673399
#>  [468,]  1.9205802  2.4100429  2.6648146  2.7552763
#>  [469,]  6.4977702  7.0730432  7.4300291  7.6000870
#>  [470,]  0.8884967  1.2080239  1.2834272  1.3231501
#>  [471,]  5.4976544  6.5238122  7.4663497  7.5192227
#>  [472,]  1.5900691  2.0706866  2.5087304  2.5845594
#>  [473,]  1.5502586  1.9372896  2.0570761  2.1659555
#>  [474,]  4.6372167  6.3188558  7.7270156  8.7260013
#>  [475,]  5.2746821  5.7855955  6.0615921  6.3037574
#>  [476,]  7.3262325  7.9987894  8.2181220  8.2915843
#>  [477,]  1.4726751  1.9776759  2.4224265  2.7464252
#>  [478,]  4.5937403  4.7718105  4.8969965  4.9179372
#>  [479,]  5.4172458  5.9693057  6.2969830  6.3766340
#>  [480,] 10.8398815 12.1715818 13.2122110 13.2988625
#>  [481,]  3.1954986  3.8258906  4.2859011  4.6950566
#>  [482,]  3.9312324  5.4558597  6.0816384  6.3287315
#>  [483,]  5.7437030  6.1991504  6.6330256  6.7369879
#>  [484,]  8.2560354 11.1958105 13.0532344 14.0817490
#>  [485,]  1.1222192  1.4499835  1.6718730  1.8332416
#>  [486,]  2.6758520  3.3551843  3.7148486  3.8858275
#>  [487,]  3.4403277  4.0171963  4.4047496  4.4946862
#>  [488,]  1.9354077  2.4681606  2.9997504  3.1335740
#>  [489,]  6.0387568  6.5151232  6.5647590  6.5799360
#>  [490,]  0.7443977  0.9649651  1.1778246  1.1885450
#>  [491,]  7.5958974  9.6778472  9.8335988  9.8374888
#>  [492,]  8.2896012  9.0109694  9.2744853  9.3475290
#>  [493,]  7.4479710  9.3407169  9.5293865  9.6767008
#>  [494,]  2.2054824  3.0186809  3.0608020  3.0846203
#>  [495,]  3.0505670  3.9848752  4.7664721  4.8407228
#>  [496,]  1.0079450  1.2277911  1.3395001  1.4060000
#>  [497,]  2.4983142  3.1540708  3.7279647  3.7451691
#>  [498,]  5.0853075  6.0505513  6.7292496  7.1027252
#>  [499,]  2.7248083  3.0497261  3.1146323  3.1654774
#>  [500,]  0.3996562  0.4795665  0.5418678  0.5848371
#>  [501,]  1.7181294  2.1750678  2.2878198  2.2988530
#>  [502,]  1.9558692  2.6586902  3.1353888  3.1631425
#>  [503,]  8.8393064  9.9625602 10.4503792 10.8573779
#>  [504,]  7.0011731  8.1499902  8.4911840  8.6435791
#>  [505,]  3.8051665  4.0958042  4.3588569  4.5580981
#>  [506,]  4.7207100  6.1517512  6.9904338  7.0916305
#>  [507,]  5.6770853  6.4580468  7.1390186  7.5454088
#>  [508,]  1.6179379  2.3003374  2.5260848  2.6843491
#>  [509,]  1.8476385  2.2256285  2.4356548  2.5923546
#>  [510,]  2.8274266  3.1215361  3.3953707  3.5666378
#>  [511,]  1.3383942  1.5147019  1.6161346  1.6955566
#>  [512,]  1.8351251  2.2606221  2.3709996  2.4058073
#>  [513,]  0.2845229  0.3420622  0.3961837  0.4034967
#>  [514,]  2.8560235  3.9520667  4.3481196  4.4699788
#>  [515,]  2.5821129  3.1259883  3.6639136  3.8866529
#>  [516,]  5.4847557  5.5659149  5.5936612  5.6147322
#>  [517,]  1.8752315  2.1853018  2.3235548  2.4584180
#>  [518,]  1.7864005  2.0300975  2.1167608  2.1838834
#>  [519,] 11.9825455 12.5616239 12.8938260 13.0319305
#>  [520,]  6.1533668  7.4692177  8.4953562  8.8913631
#>  [521,]  4.1247247  5.8839425  6.4592336  6.4791729
#>  [522,]  8.4222230  8.9214370  9.3996026  9.7239375
#>  [523,]  1.7292736  1.8453989  1.9587103  2.0182093
#>  [524,]  6.0964589  7.1281547  7.8032103  8.1539994
#>  [525,]  4.2483521  4.4945648  4.5738281  4.5796621
#>  [526,]  9.1654458 11.3207489 11.3906098 11.3914824
#>  [527,]  2.1316872  2.8034453  3.2480873  3.5067751
#>  [528,]  5.7408158  6.2244801  6.4956256  6.6675465
#>  [529,]  6.7936219  6.9151822  7.0156192  7.0378588
#>  [530,]  3.0207811  3.3234653  3.4915864  3.6029510
#>  [531,]  1.3966097  1.6025986  1.7894051  1.9100525
#>  [532,]  1.3548740  1.5088839  1.5275703  1.5439077
#>  [533,]  3.9238842  4.7243003  5.0336963  5.0994805
#>  [534,]  1.9313383  2.3479893  2.7363962  3.0711247
#>  [535,]  8.4584027  9.4225834 10.1327265 10.6595369
#>  [536,]  2.7510681  3.3962054  3.9955056  4.3129008
#>  [537,]  4.4188530  5.8270889  6.5546520  6.6404492
#>  [538,]  3.6251688  4.3730110  4.5986751  4.6349721
#>  [539,]  4.8566660  5.1766071  5.4775055  5.5142949
#>  [540,]  3.9345828  4.5578362  5.0812514  5.1262562
#>  [541,]  5.6186508  6.6674427  7.2044586  7.3174150
#>  [542,]  2.5699933  3.0059385  3.2571599  3.4618320
#>  [543,]  1.1018891  1.5331073  1.7767677  1.9109669
#>  [544,]  6.9640368  9.0276888  9.7184557  9.9350788
#>  [545,]  5.5189579  6.1616429  6.4670078  6.6357792
#>  [546,]  5.4132637  7.7844894  8.0929836  8.1350861
#>  [547,]  6.3693133  6.8610829  7.0461518  7.1218501
#>  [548,] 11.5343256 13.8096177 14.1529871 14.4779484
#>  [549,]  4.4195097  4.9052448  5.0582376  5.1892003
#>  [550,]  6.6646223  7.7553558  8.4486243  9.0318511
#>  [551,]  2.5150728  3.0848355  3.6121498  3.9656758
#>  [552,] 14.1746241 16.1861748 16.7159517 16.9784775
#>  [553,] 11.3167670 13.0029524 14.0685847 14.8530666
#>  [554,]  3.1680031  4.1184602  5.0042243  5.2148503
#>  [555,]  2.1072513  2.5294100  2.6797633  2.7175166
#>  [556,]  3.3179814  3.5041015  3.6841129  3.8442262
#>  [557,]  6.3025558  8.0660899  8.8534696  9.1908067
#>  [558,]  5.0615703  5.6802474  6.0968014  6.5011178
#>  [559,] 11.6513817 11.9837171 12.0895592 12.1597367
#>  [560,]  3.1410489  4.1182338  4.7067770  5.0236004
#>  [561,]  3.7883821  3.8776807  3.9068330  3.9326646
#>  [562,]  3.9705676  5.2906386  6.0128395  6.1997226
#>  [563,]  3.4098636  3.6810041  3.8324401  3.9109272
#>  [564,]  4.9750429  5.3257928  5.4173716  5.4986902
#>  [565,]  4.3851960  4.6891679  4.8633836  4.8769849
#>  [566,]  5.7277816  6.4389718  7.0309287  7.4030586
#>  [567,]  3.0979341  4.4826139  5.7418890  6.9284589
#>  [568,]  3.3357446  3.7051628  3.9242564  3.9504018
#>  [569,]  7.5637693  9.7538535 10.7287458 11.4503440
#>  [570,]  6.6851373  8.1876768  8.5761338  8.7920571
#>  [571,]  1.8916351  2.2013710  2.4287069  2.5493827
#>  [572,]  6.1361767  6.9735995  7.1961464  7.2274338
#>  [573,]  5.2756260  7.2595051  8.9745774  9.3869867
#>  [574,]  2.0815419  2.2822770  2.4672600  2.5908069
#>  [575,]  1.8530924  2.2714933  2.6535155  2.6977853
#>  [576,]  6.3726858  8.4981408  8.6229086  8.6392998
#>  [577,]  4.1891922  5.1748910  6.0368837  6.1838990
#>  [578,]  6.0467663  6.6725473  7.1545606  7.5063250
#>  [579,]  6.3412941  6.4099918  6.4498320  6.4678180
#>  [580,]  3.6782296  3.9414791  4.1690156  4.2884298
#>  [581,]  5.6259958  6.2159625  6.6362853  6.6387590
#>  [582,]  5.3430586  7.4463947  8.0252846  8.1594172
#>  [583,]  2.9297490  3.5088196  3.7136979  3.8268742
#>  [584,]  1.7397952  1.8069022  1.8692765  1.9219242
#>  [585,]  8.4682038  9.8134833  9.9282028  9.9633687
#>  [586,]  1.4971668  1.6420040  1.7750530  1.8775291
#>  [587,]  5.7591703  7.2709432  7.7484901  8.0799965
#>  [588,]  2.5856985  3.3768145  4.0876403  4.7269760
#>  [589,]  2.4686127  3.0776744  3.2187791  3.2376784
#>  [590,]  1.6403508  1.9184349  2.1722523  2.1766272
#>  [591,]  1.2733889  1.3517155  1.3868507  1.3995165
#>  [592,]  5.0021185  6.7974395  7.8371776  8.4703289
#>  [593,] 12.3885418 17.0866214 18.7597178 19.5076983
#>  [594,]  3.2176490  3.9136982  4.5788663  4.6927669
#>  [595,]  1.4631522  1.6037778  1.7078417  1.7698910
#>  [596,]  3.5838297  4.8187761  4.9708834  5.0372189
#>  [597,]  5.6569067  7.7707810  7.8445424  7.9159928
#>  [598,]  4.2453207  4.4329982  4.6039847  4.6322831
#>  [599,]  8.0298389  8.8427307  9.3476979  9.7986539
#>  [600,]  7.1489830 10.0421319 10.8230795 11.3747362
#>  [601,]  3.0604145  3.2127580  3.2817149  3.3092188
#>  [602,]  1.1105900  1.3002029  1.4493966  1.5641377
#>  [603,]  1.9677307  2.2244068  2.2910624  2.3025454
#>  [604,]  1.0222332  1.1674977  1.2503503  1.3028607
#>  [605,]  4.3621465  4.7713198  4.9253313  4.9730851
#>  [606,] 10.0984998 11.4908408 12.8360822 13.8364599
#>  [607,]  1.9589357  2.1592311  2.2715702  2.3337726
#>  [608,]  6.3376650  6.3613788  6.3747101  6.3802028
#>  [609,]  2.3798259  3.5116497  4.1027451  4.5001245
#>  [610,] 10.9934528 12.1174950 12.7680866 13.0323200
#>  [611,]  4.3504536  4.7966356  4.9074869  4.9666332
#>  [612,]  3.2228763  3.2508051  3.2699975  3.2822310
#>  [613,]  5.5353070  5.9024201  6.1497729  6.1612041
#>  [614,]  5.1818433  6.8074405  8.2787044  8.6646694
#>  [615,]  6.6216698  8.3852644  9.9846808 10.7153762
#>  [616,]  2.5427157  2.9218668  3.0211148  3.0735007
#>  [617,]  3.5298702  4.3626224  5.1028838  5.3071743
#>  [618,]  3.4909501  3.8792036  4.0127055  4.1309639
#>  [619,]  5.4526601  6.0650032  6.3166124  6.3556792
#>  [620,]  1.7074299  2.4927362  2.8682741  2.8723325
#>  [621,]  7.3497997  8.5089939  8.6936177  8.8641372
#>  [622,]  3.2151799  3.8066592  3.8439996  3.8660281
#>  [623,]  2.6908693  3.2807097  3.3013483  3.3054577
#>  [624,]  3.0693323  3.2564020  3.4094580  3.4942966
#>  [625,]  2.7467288  2.7974498  2.8366334  2.8478899
#>  [626,]  1.3771189  1.9235257  2.1614663  2.3549045
#>  [627,]  0.9540555  1.2977043  1.5330404  1.7000028
#>  [628,]  7.2507310  8.1637162  8.5906669  8.6636472
#>  [629,]  0.8387180  1.2086004  1.4994655  1.6333029
#>  [630,]  6.5122452  7.6821266  8.1723993  8.1940616
#>  [631,]  3.1676026  3.2361076  3.2822679  3.3265079
#>  [632,]  6.0361014  8.4061631  9.8151007  9.8384947
#>  [633,]  2.3543366  2.7864071  3.1099245  3.3521851
#>  [634,]  3.8232105  4.4584868  5.0051852  5.0825801
#>  [635,]  2.7447851  3.5496900  4.2960401  4.7837359
#>  [636,]  3.4039326  3.6883493  3.7695944  3.7986929
#>  [637,]  6.8870879  8.8594696  9.0950157  9.2055831
#>  [638,]  3.1549237  3.5489255  3.8072496  3.9013873
#>  [639,]  7.2139957  8.3067875  9.1454709  9.6288870
#>  [640,]  4.1563430  4.8737176  5.4098801  5.7846661
#>  [641,]  0.8185600  0.9936403  1.1539850  1.2837447
#>  [642,]  3.7072894  4.3365062  4.6036752  4.8469993
#>  [643,]  1.9185646  2.1645395  2.3952047  2.4787736
#>  [644,] 10.3525478 11.7097300 13.0372325 14.0983492
#>  [645,]  2.6817746  3.2567221  3.7522275  3.8024312
#>  [646,]  2.4244765  3.0251147  3.1823319  3.2317472
#>  [647,]  2.6767831  3.2963356  3.7845927  3.8230255
#>  [648,]  4.1975397  4.3762330  4.5002828  4.5027505
#>  [649,]  7.2753674  8.3478569  9.0061468  9.0161476
#>  [650,]  5.2066783  5.7535030  6.2743266  6.4366846
#>  [651,]  1.8439417  2.1396471  2.3164843  2.3283172
#>  [652,]  5.9001540  7.1781196  7.5617459  7.6020542
#>  [653,]  6.8278344  8.4636661  9.9657850 10.9325218
#>  [654,]  5.7047792  6.6503827  7.2397202  7.3054759
#>  [655,]  1.8805693  2.0333050  2.0491398  2.0618028
#>  [656,]  6.7716230  8.7638891  9.4414471  9.7996033
#>  [657,]  3.8487804  4.8432988  5.0242067  5.0836459
#>  [658,]  4.7333790  5.3707862  5.5532293  5.6184660
#>  [659,]  5.1293815  5.3572175  5.4796217  5.5557787
#>  [660,]  3.8323636  4.6021439  4.7597581  4.7894210
#>  [661,]  4.5544552  5.1226511  5.6142393  5.9948103
#>  [662,]  0.7546719  1.0116002  1.2253710  1.3239133
#>  [663,]  0.9563251  1.1708627  1.3529790  1.5172610
#>  [664,]  3.0804688  3.6561240  4.0927189  4.3830086
#>  [665,]  3.5159321  3.8618341  3.9028285  3.9301412
#>  [666,]  2.6565345  2.8871777  3.0934553  3.2406266
#>  [667,]  5.7898967  7.2485654  8.0211090  8.3442275
#>  [668,]  4.7748419  6.2665260  7.5080106  8.4621722
#>  [669,]  3.6551813  4.6349013  5.0408823  5.1201832
#>  [670,]  4.0066254  4.2912353  4.5628336  4.7305388
#>  [671,]  5.3037817  7.0787552  7.5615572  7.8657487
#>  [672,]  3.7851317  4.5338018  5.1174705  5.2271565
#>  [673,]  2.9108219  3.6936067  4.4314498  4.5331442
#>  [674,]  2.2389223  2.7485773  2.8332425  2.8871714
#>  [675,]  3.9263599  5.2949849  6.0713210  6.7888762
#>  [676,]  1.5298099  1.8192092  1.9201296  1.9267979
#>  [677,]  2.9534787  3.2429057  3.3788014  3.4273811
#>  [678,]  0.7509388  0.8814557  0.9422739  0.9584735
#>  [679,]  1.6589102  1.7853733  1.9033110  1.9315244
#>  [680,]  5.1995557  7.2819080  7.9252036  7.9774049
#>  [681,]  2.5721542  2.8439078  2.9042144  2.9195204
#>  [682,]  2.1924575  2.5410336  2.7256005  2.8971270
#>  [683,]  9.4458050 12.1240419 12.5700239 12.9494166
#>  [684,]  1.9955435  2.8019028  2.9054479  2.9931360
#>  [685,]  1.1130568  1.4171783  1.6772902  1.8288523
#>  [686,]  0.4226047  0.4996686  0.5389983  0.5468963
#>  [687,]  6.2713538  7.6039520  7.8614716  7.8952200
#>  [688,]  6.9985489  7.2819257  7.3042978  7.3229461
#>  [689,]  6.2230299  6.6528114  6.6669047  6.6693064
#>  [690,]  4.1677174  5.1976178  5.3109069  5.3495103
#>  [691,]  5.5031182  5.9994367  6.3885716  6.4802350
#>  [692,]  6.4365637  7.6320896  7.8417916  7.9936425
#>  [693,]  1.6498946  2.3074525  2.8345997  2.9835288
#>  [694,]  0.6315983  0.7848372  0.9150626  0.9258634
#>  [695,]  3.1725502  3.7683884  4.2666637  4.5125778
#>  [696,]  5.8564511  7.1413080  8.4200204  8.9682538
#>  [697,]  1.2063586  1.4293485  1.4613900  1.4862027
#>  [698,]  2.9961095  3.3553742  3.5586549  3.7127553
#>  [699,]  3.9269464  4.6643010  5.4012106  5.9530907
#>  [700,]  1.9333297  2.5954590  3.1779873  3.6134614
#>  [701,]  4.2192521  4.6578541  5.0010258  5.0484021
#>  [702,]  9.1207477 13.2275321 14.7226966 14.9801679
#>  [703,]  3.5737402  4.8657193  5.2817589  5.6215427
#>  [704,]  7.3078365  8.6426863  9.8160190 10.6179909
#>  [705,]  4.4487478  6.5592696  8.1756151  8.4374682
#>  [706,]  2.8275091  3.4673924  3.8631971  3.8957559
#>  [707,]  3.5333025  4.1619621  4.7388712  4.7488719
#>  [708,]  1.1752312  1.5438393  1.9080539  1.9350779
#>  [709,]  1.6709538  1.9887515  2.1701617  2.2884114
#>  [710,]  4.7400267  5.4395125  6.0256339  6.2409794
#>  [711,]  7.3892741  8.2501866  9.0765753  9.7292804
#>  [712,]  4.0118048  4.3378631  4.5746433  4.7658526
#>  [713,]  9.2434353 10.4712760 10.6471548 10.8146824
#>  [714,]  2.6976868  3.2668405  3.2807814  3.2908115
#>  [715,]  1.7152385  2.1875809  2.4032804  2.5297555
#>  [716,]  2.2695754  2.9274994  3.3607128  3.4880770
#>  [717,]  2.9691313  3.3154098  3.3586702  3.3691816
#>  [718,]  2.6026545  2.8383399  2.9544495  2.9954641
#>  [719,]  5.2075371  5.6023339  5.9640035  6.0385223
#>  [720,]  1.3975069  1.6389478  1.8496548  2.0492241
#>  [721,]  4.0939766  5.7520935  6.4211816  6.7889432
#>  [722,]  2.4631596  3.0644039  3.4369682  3.4506549
#>  [723,]  8.0598852  8.4901675  8.7703519  8.9136608
#>  [724,]  6.9071031  7.6159210  8.1861989  8.7092420
#>  [725,]  6.9744917  7.8537748  8.2248586  8.2481280
#>  [726,]  6.2528824  6.7485340  6.8856473  6.9546163
#>  [727,]  6.3453321  7.6694240  8.7257070  8.8482472
#>  [728,]  4.1700812  5.1763555  6.1515372  6.4365034
#>  [729,]  7.3359361  9.5102512 11.2791363 12.2596127
#>  [730,]  3.3787332  4.0567662  4.6854260  4.9188512
#>  [731,]  5.6535005  6.2094402  6.7042553  7.0437736
#>  [732,]  4.9449457  5.0201117  5.0893385  5.1498458
#>  [733,]  2.7191517  3.4060265  3.8496739  4.1638119
#>  [734,]  2.5948261  3.2792461  3.9196878  4.0440860
#>  [735,]  2.0116247  2.2207738  2.3777693  2.4834710
#>  [736,]  7.5227072  8.9768931  9.3480210  9.7044468
#>  [737,]  8.0634928  9.4259804  9.9843013 10.2107477
#>  [738,]  5.8076064  6.6666439  7.1776915  7.3176551
#>  [739,]  4.8772210  6.3488542  7.1783178  7.4665568
#>  [740,]  5.3251887  6.4903092  7.4706155  7.5591336
#>  [741,]  4.4665633  5.6266387  6.6721688  6.8659280
#>  [742,]  3.1653241  3.9909136  4.6054223  4.7781495
#>  [743,] 11.8068393 15.8818335 17.7036575 18.3330765
#>  [744,]  7.6994664  9.3678193  9.6628556  9.7580186
#>  [745,]  1.1694198  1.6245783  2.0603961  2.3470050
#>  [746,]  8.0829603  9.0411650  9.9586188 10.6743556
#>  [747,]  6.4470867  7.4096490  7.8842617  8.0090405
#>  [748,]  2.4465432  3.1967015  3.4677441  3.4777282
#>  [749,]  2.0511883  2.3759554  2.6053295  2.7875892
#>  [750,]  3.9865599  5.4029549  6.3204238  6.7320535
#>  [751,]  8.3648093  8.8608794  9.2404799  9.3923551
#>  [752,]  2.4773497  2.8919341  3.2915343  3.3883931
#>  [753,]  3.1689945  3.5877247  3.7621173  3.8240285
#>  [754,]  3.9503523  4.4819990  4.7334161  4.8494836
#>  [755,]  7.0312074  7.8868292  8.6981832  9.0571852
#>  [756,]  2.9019526  3.4121691  3.6419748  3.8081107
#>  [757,]  5.6018936  6.0662102  6.4686853  6.4719000
#>  [758,]  4.1618945  4.7656125  4.8262072  4.8784919
#>  [759,]  7.0594069  8.0875628  8.6936592  8.8345801
#>  [760,]  3.3053165  4.0733583  4.3042578  4.5066799
#>  [761,]  8.7631089 10.5211127 12.1359390 12.9590999
#>  [762,]  9.7858942 10.6721660 11.4258224 11.8875406
#>  [763,]  6.4473821  7.4062193  7.4728225  7.5290353
#>  [764,]  4.5975147  5.5984185  6.5594109  7.2594953
#>  [765,]  3.5030977  4.5287953  5.1979484  5.2302553
#>  [766,]  1.1665392  1.4491640  1.5530460  1.6341839
#>  [767,]  3.8080178  4.4028502  4.6231326  4.8386894
#>  [768,]  2.7410223  3.5424860  4.1883848  4.3088579
#>  [769,]  3.8307637  4.1301519  4.4246917  4.5931919
#>  [770,]  5.2793627  6.2826852  6.3418712  6.3994274
#>  [771,]  3.8077624  4.4012623  4.6173288  4.7005808
#>  [772,]  0.7636988  0.7944112  0.8106748  0.8223303
#>  [773,]  6.9032485  7.9762842  8.8142735  9.0803105
#>  [774,]  2.9452390  3.6832284  3.7732923  3.8359388
#>  [775,]  4.4019475  4.6712875  4.9059805  4.9109852
#>  [776,]  3.2032635  4.0407366  4.5351062  4.5736725
#>  [777,]  3.7118043  4.0192104  4.0962118  4.1301113
#>  [778,]  1.4783502  2.1400336  2.5336461  2.9244314
#>  [779,]  4.7019708  5.6736808  6.4793846  6.9062007
#>  [780,]  3.1979125  4.1738966  4.8884670  5.2599763
#>  [781,]  3.3249108  3.5677164  3.7350334  3.7398596
#>  [782,]  4.7173209  5.8662926  6.3075098  6.3566383
#>  [783,]  5.4744112  7.5760863  8.7000331  9.1928837
#>  [784,]  3.8926718  4.8187981  5.3307614  5.4649274
#>  [785,]  7.3140753  7.8438144  8.3146092  8.7511021
#>  [786,]  1.2809322  1.8335672  2.0200691  2.0472352
#>  [787,]  4.8377824  5.3606518  5.4950484  5.4974058
#>  [788,]  4.7306699  5.6156901  5.8517318  6.0740742
#>  [789,]  1.6986411  1.9663429  2.1810736  2.2951704
#>  [790,]  1.7795154  2.4097692  2.5224586  2.6074797
#>  [791,]  5.5187718  6.1283840  6.5623436  6.6437425
#>  [792,]  1.2676020  1.6653701  1.9618685  2.1670575
#>  [793,]  4.6471924  5.5704609  6.3219946  6.9179839
#>  [794,]  6.6037249  7.3491891  7.5920639  7.7437691
#>  [795,]  4.7332548  5.6619092  5.7688158  5.7929896
#>  [796,]  4.6510114  5.1206023  5.1267270  5.1322502
#>  [797,]  3.7330287  4.6763791  5.3771047  6.0197254
#>  [798,]  3.7701093  4.5664520  4.7323127  4.8898692
#>  [799,]  7.5319264  7.8457782  8.0772238  8.3050303
#>  [800,]  3.8457418  5.3807344  6.2413691  6.3315773
#>  [801,]  4.5048032  5.2383102  5.8621602  6.0037388
#>  [802,]  2.0977390  2.8535078  2.9903353  3.0846644
#>  [803,]  4.0531504  5.1720728  6.0076046  6.1373238
#>  [804,]  4.4194517  4.8127283  5.1485418  5.3427427
#>  [805,]  4.5255364  5.0610935  5.4132709  5.7075714
#>  [806,]  3.1617512  3.5760011  3.9363893  4.0704932
#>  [807,]  6.1264012  7.3220522  7.8928406  8.1418579
#>  [808,]  4.4208019  5.9081281  6.4882227  6.8635386
#>  [809,]  1.6507687  2.0607395  2.2809827  2.4933051
#>  [810,]  2.0873797  2.6045208  2.8832318  3.1331553
#>  [811,]  3.5833534  4.3120478  4.7965760  4.9184992
#>  [812,]  1.7396346  2.2797097  2.3569994  2.3655667
#>  [813,]  0.3485696  0.4514357  0.5430363  0.5586836
#>  [814,]  1.3031439  1.6023050  1.7530960  1.8783622
#>  [815,]  2.7341509  3.2679131  3.3035640  3.3194803
#>  [816,]  2.6194440  2.8600568  2.8987262  2.9086019
#>  [817,]  1.4870810  1.5510846  1.5786409  1.5925539
#>  [818,]  3.3865481  4.1138160  4.2567208  4.3604792
#>  [819,]  3.5955980  3.8641491  4.1162773  4.1583747
#>  [820,]  5.6610182  6.4153121  6.8146084  7.1073300
#>  [821,]  3.2088007  3.2954171  3.3328549  3.3670014
#>  [822,]  2.8859551  3.1950194  3.3328729  3.4450948
#>  [823,]  6.1061096  8.5037366  8.9547832  9.0081728
#>  [824,]  1.6957520  2.2336080  2.6954331  2.8823675
#>  [825,]  1.6853391  2.2609799  2.4652647  2.5388855
#>  [826,]  6.6659175  7.5170926  7.7656276  7.9041094
#>  [827,]  4.6604001  4.7624945  4.8483516  4.9180533
#>  [828,]  1.0433647  1.1257815  1.1841305  1.2234108
#>  [829,]  5.3107891  5.7750863  6.0401096  6.0847530
#>  [830,]  7.8505977  9.4004708 10.5138035 11.2003140
#>  [831,]  1.3885266  1.5364194  1.6632668  1.7821097
#>  [832,]  5.6332933  6.6972832  6.8456376  6.8864337
#>  [833,]  2.8868877  2.9649692  2.9929532  3.0183570
#>  [834,]  3.3679269  4.3608701  4.5545318  4.6144027
#>  [835,]  5.6246798  6.2175935  6.5866164  6.8028509
#>  [836,]  3.1948656  3.3091932  3.3706595  3.3835055
#>  [837,]  0.4970605  0.6712150  0.7349965  0.7902609
#>  [838,]  4.4143588  4.9778865  5.0340532  5.0783691
#>  [839,]  1.0060910  1.3796951  1.6031593  1.6266672
#>  [840,]  8.0261340  9.6679115 10.1493167 10.5066027
#>  [841,]  4.1378128  5.3096236  5.4872043  5.6419172
#>  [842,] 10.3656191 11.0656375 11.2232871 11.3260458
#>  [843,]  1.2048573  1.4098052  1.4868673  1.5547140
#>  [844,]  2.4711952  2.8501841  3.1722910  3.4204032
#>  [845,]  2.7255370  2.9740857  3.1416195  3.1870008
#>  [846,]  3.6177761  4.1076200  4.3838596  4.5902323
#>  [847,]  7.0372961  9.9014174 12.6018448 12.8907234
#>  [848,]  4.6721502  5.4852435  6.2465877  6.5217407
#>  [849,]  4.6956018  5.3452962  5.7766378  5.8216253
#>  [850,]  9.1777638  9.6137167 10.0116307 10.3059496
#>  [851,]  8.0840069  8.5653087  8.7832917  8.8673741
#>  [852,]  1.9533878  2.5609311  2.9958217  3.4176051
#>  [853,]  7.9744332  9.1097989 10.0644852 10.1757649
#>  [854,]  6.0583312  6.5929590  6.7511937  6.8464221
#>  [855,]  1.0572447  1.1864162  1.2656173  1.2658613
#>  [856,]  6.4997205  7.2142692  7.6496376  8.0486357
#>  [857,]  3.7037586  4.1546910  4.6035258  4.8847781
#>  [858,]  6.9126559  7.5325075  8.1482016  8.5925498
#>  [859,]  2.1078205  2.5968560  2.6756851  2.7516314
#>  [860,]  2.6058147  3.3810448  4.0439434  4.5644845
#>  [861,]  6.1006458  7.0886325  7.4112436  7.5923248
#>  [862,]  5.8299706  7.2079206  8.1423168  9.0132166
#>  [863,] 10.3513051 11.4628453 12.5461317 13.3248353
#>  [864,]  0.8762429  1.1309804  1.2825455  1.3304644
#>  [865,]  5.1553498  5.9443537  6.5648827  6.8342669
#>  [866,]  6.4883473  7.0228118  7.2489230  7.4495785
#>  [867,]  2.0247230  2.5179589  2.9761132  2.9994295
#>  [868,]  4.4082409  5.4284870  6.1286724  6.7695332
#>  [869,]  2.8940557  3.5792601  3.7294517  3.8132324
#>  [870,]  3.7747659  3.9155393  4.0532831  4.0923732
#>  [871,]  5.7770179  6.2592189  6.6766272  6.8463155
#>  [872,]  2.9929150  3.9377936  4.2332008  4.4477302
#>  [873,]  4.7612498  6.5207285  8.2488333  9.3757424
#>  [874,]  3.5170719  3.9466686  4.1043536  4.2604136
#>  [875,]  3.8357542  5.3618713  6.7610673  7.7986677
#>  [876,]  2.5015222  3.3567436  3.6771591  3.8754627
#>  [877,]  7.7555549  8.7736873  9.6831854  9.9641714
#>  [878,]  6.5462556  7.6268801  8.4930034  9.0330897
#>  [879,]  1.0655744  1.4296516  1.5955833  1.6643316
#>  [880,]  8.6385287 10.3163117 11.9082402 12.0586806
#>  [881,]  3.5625083  4.9571837  5.6330464  6.0875087
#>  [882,]  4.2532284  4.6268186  4.8879288  5.0431314
#>  [883,]  3.6494767  4.4943988  5.2313203  5.5413583
#>  [884,]  9.5676852 12.0905148 12.1861489 12.2691278
#>  [885,]  2.0459552  2.4431018  2.8079937  2.8425897
#>  [886,]  9.3844377 10.8593247 12.0849516 13.0382846
#>  [887,]  4.1502938  5.5028187  6.5662003  6.6400159
#>  [888,]  2.5526752  3.0891580  3.4992076  3.9015678
#>  [889,]  4.4578497  4.9599261  5.2138600  5.2830721
#>  [890,]  2.6829374  3.1244792  3.2776180  3.3314064
#>  [891,]  1.9328427  2.4164389  2.4454284  2.4661401
#>  [892,]  5.6316734  6.5280992  7.3384612  7.8751431
#>  [893,]  2.3159866  2.9163700  3.2317643  3.5266685
#>  [894,] 14.8525153 19.5480509 20.8667365 21.1974021
#>  [895,]  5.8300862  6.6846737  7.1140699  7.5264761
#>  [896,]  4.6737802  5.2967731  5.8288697  6.0277434
#>  [897,]  3.0790755  3.7509522  4.0739684  4.1188681
#>  [898,]  4.5075264  5.2338509  5.8712697  6.1405426
#>  [899,]  4.4969051  5.1614537  5.8194849  5.9091885
#>  [900,]  6.1302797  7.8098529  9.3035375  9.6538788
#>  [901,]  3.6991571  4.0446306  4.1222521  4.1834630
#>  [902,]  5.7428660  7.0567052  8.3170151  9.0806094
#>  [903,]  4.8316214  5.2161924  5.5705730  5.8728029
#>  [904,]  4.9486959  5.9839680  6.9770857  6.9878713
#>  [905,]  3.1890483  4.1249769  5.0041960  5.0073814
#>  [906,]  1.3368168  1.9050452  2.1131348  2.1394299
#>  [907,]  7.4602374  8.5530322  8.9951738  9.4006243
#>  [908,] 11.0103492 12.6359211 14.2409859 15.5533504
#>  [909,]  5.2324648  5.8261438  6.1259091  6.2111457
#>  [910,]  6.7720682  9.9744161 13.0518833 15.7685516
#>  [911,]  2.7015518  3.9179352  4.0042405  4.0755390
#>  [912,]  2.7152098  2.9323333  2.9851627  3.0176963
#>  [913,]  6.6684349  7.3307812  7.5865431  7.7189455
#>  [914,]  2.0010477  2.4404630  2.5038098  2.5631089
#>  [915,]  1.8791663  2.0132275  2.1147491  2.1503603
#>  [916,]  7.6808446  8.0075931  8.1737058  8.2579135
#>  [917,]  2.5077333  2.7094804  2.8943330  3.0101947
#>  [918,]  5.9767989  7.8168834  8.4480829  8.5426700
#>  [919,]  4.8550284  5.3514227  5.6329094  5.8189252
#>  [920,]  4.4147335  5.1207318  5.6142023  5.7999172
#>  [921,]  2.3689071  2.6942194  2.8016304  2.8892248
#>  [922,]  4.4739638  4.9306801  5.2884222  5.5476615
#>  [923,]  6.5177956  8.8537942 10.7862828 10.8129584
#>  [924,]  4.8970748  6.9271186  7.5839975  8.1345121
#>  [925,]  4.2668199  5.5747313  6.5612981  6.5900125
#>  [926,] 10.7801655 13.9248121 14.9366090 15.1033722
#>  [927,]  2.0360417  2.5591000  3.0472040  3.3941661
#>  [928,]  6.2044348  7.6672789  8.1292431  8.3046281
#>  [929,]  4.8842646  5.6510723  6.2535403  6.4346866
#>  [930,]  3.7481817  4.5171693  4.7030693  4.8600779
#>  [931,]  1.4971839  2.0006442  2.1204487  2.1249015
#>  [932,]  2.4480711  2.6892962  2.7271798  2.7549117
#>  [933,]  6.3890614  7.1364008  7.5858018  7.9345342
#>  [934,]  2.9686043  3.4347712  3.5181222  3.5280768
#>  [935,]  4.5068048  5.4189945  5.9143297  6.3657526
#>  [936,]  3.3970014  3.7837644  4.0230360  4.0682623
#>  [937,]  5.2238731  6.1577816  6.8217849  7.2010418
#>  [938,]  2.9857077  3.3283680  3.4044009  3.4544211
#>  [939,]  4.7728950  5.8359461  6.4137194  6.6889730
#>  [940,]  2.1568536  2.9044245  3.5780998  3.6628873
#>  [941,]  7.4451551  8.9534025 10.1805875 11.1777208
#>  [942,]  3.3370489  3.8752250  4.3713816  4.5741109
#>  [943,]  7.1422590  7.7690812  8.0681645  8.2739581
#>  [944,]  7.1732700  8.4387402  9.0242306  9.5703576
#>  [945,]  1.0802619  1.3080189  1.5014372  1.5996236
#>  [946,]  4.6100179  6.0569738  6.6067586  6.7098908
#>  [947,]  6.9084199  8.4151567  9.5173737  9.8542781
#>  [948,]  2.8459673  3.0696583  3.1425198  3.1453247
#>  [949,]  4.3095919  5.7921213  6.4700352  6.4931365
#>  [950,]  2.4266553  3.0363502  3.2313523  3.4219333
#>  [951,]  2.2865919  2.8674402  3.2896478  3.3444727
#>  [952,]  3.3236385  3.8697595  4.3700331  4.7204323
#>  [953,]  2.3461305  3.0936857  3.2541666  3.2948110
#>  [954,]  3.6778599  3.7847130  3.8153809  3.8157668
#>  [955,]  7.7592731  8.3777802  8.4923708  8.5570112
#>  [956,]  6.8500015  9.3255773  9.4973412  9.6164970
#>  [957,]  1.4459056  1.8314855  2.0190137  2.0560123
#>  [958,]  1.4428157  1.9703020  2.4048403  2.4111147
#>  [959,]  3.4255640  3.5520791  3.6608249  3.6984233
#>  [960,]  5.2667484  6.6048104  6.8265722  7.0402472
#>  [961,]  1.9661765  2.5907695  2.8991787  3.0196271
#>  [962,]  3.7795944  4.3159581  4.7248426  5.1065537
#>  [963,]  5.6311130  6.1435171  6.6244216  6.7108480
#>  [964,]  8.4794516  9.4706181  9.5247139  9.5758715
#>  [965,]  3.1124978  3.3539694  3.4890226  3.5034019
#>  [966,]  2.3098750  3.1630055  3.8148390  3.9870436
#>  [967,]  3.3802457  4.0547703  4.1129197  4.1676413
#>  [968,] 17.6639471 18.0158437 18.0665373 18.1060771
#>  [969,]  1.7649785  2.1000222  2.1578434  2.2020843
#>  [970,]  5.0404673  6.6416767  7.6048032  8.0912918
#>  [971,]  7.6075746  9.6413169  9.7927860  9.8673808
#>  [972,]  0.5852333  0.7077405  0.7421614  0.7548073
#>  [973,]  5.6485784  7.0307894  8.1244997  9.0847421
#>  [974,]  3.2618296  3.8071406  4.0786391  4.2481990
#>  [975,]  4.0771195  4.9986968  5.5306415  5.7047386
#>  [976,]  5.0503267  5.7161979  5.8854986  5.9267369
#>  [977,]  2.8717389  3.4378639  3.8788510  4.2435928
#>  [978,]  4.2329528  5.0480418  5.1965065  5.1990198
#>  [979,]  1.7580405  2.2238184  2.4989904  2.5662157
#>  [980,]  4.8859987  5.4040820  5.6091427  5.7344794
#>  [981,]  3.5570949  3.7711227  3.8696795  3.8812803
#>  [982,]  4.2250903  4.5496841  4.7036781  4.7398181
#>  [983,]  2.4886398  2.8893609  3.2188825  3.3560597
#>  [984,]  8.6965403 10.8490804 11.1619604 11.2095392
#>  [985,]  6.9743675  8.1780380  8.8015358  9.2809039
#>  [986,]  1.8082926  2.2988431  2.6622119  2.9834404
#>  [987,]  2.2830516  2.4384258  2.5394957  2.5418311
#>  [988,]  7.9918163  8.8851769  9.3999365  9.5000691
#>  [989,]  3.7758841  4.0885473  4.3242384  4.3462550
#>  [990,]  4.4672979  4.9557126  5.4056564  5.8487380
#>  [991,]  4.4481123  4.7507357  4.8009011  4.8116992
#>  [992,]  5.9901945  6.7715535  7.3035463  7.8259682
#>  [993,]  6.9553855  9.5609768 11.9881398 12.4606324
#>  [994,]  5.6683513  6.7025231  7.6195238  7.7537863
#>  [995,]  4.8623982  5.9248015  6.7515246  7.0886559
#>  [996,]  3.7406782  4.3291250  4.6729658  4.9891856
#>  [997,]  5.5267843  6.4670604  7.3446755  7.9720127
#>  [998,]  3.6178809  4.0168742  4.3986288  4.4155200
#>  [999,]  2.9914233  3.6011177  4.1849499  4.4511286
#> [1000,] 14.1979611 14.3920695 14.4854792 14.5020005