generate_null_distribution_T3

This function generate the null distribution of T_eta, should be run before calculating TWT. Please see Setp 2 in our manuscript for more information.

generate_null_distribution_T3(m, n, cov_mat, cutoff_value)

Arguments

m

Dimension.

n

Number of the generated sample size.

cov_mat

Covariance matrix for generating random samples.

cutoff_value

Truncated value, should be in 0 to 1 and must contain 1.

Value

Null distribution matrix with number of rows equal n and number of columns equal length of cutoff value minus 1.

Examples

ThreeWayTest::generate_null_distribution_T3(m=6,n=1000,
cov_mat=diag(nrow = 6, ncol = 6), cutoff_value=c(0.2,0.4,0.6,0.8,1))
#>                [,1]       [,2]        [,3]        [,4]
#>    [1,]  1.31207845  1.8280632  1.92734099  1.99875197
#>    [2,]  3.60910825  3.9263654  4.24295958  4.52604795
#>    [3,]  1.81770884  2.6313180  3.06583057  3.13404812
#>    [4,] 13.07704530 14.7576598 15.69479088 15.93237747
#>    [5,]  7.54075011  8.9674623 10.02229539 10.50951158
#>    [6,]  2.94422793  3.0743726  3.16615307  3.19334397
#>    [7,]  2.53800177  3.1023624  3.55476853  3.82832370
#>    [8,]  3.62652738  4.0742509  4.32428399  4.37219329
#>    [9,]  3.20793584  3.3548989  3.49381947  3.56771843
#>   [10,]  1.47035820  1.5464516  1.59368651  1.62104457
#>   [11,]  1.13204136  1.5955134  1.80133927  1.99036875
#>   [12,]  4.07481688  4.8954973  5.55671363  5.67742638
#>   [13,] 11.64812394 12.7245870 13.06404097 13.29062731
#>   [14,]  9.35165075 10.0853414 10.34543151 10.40502541
#>   [15,]  2.89792941  3.9448483  4.20998694  4.35017292
#>   [16,]  2.90821629  3.2493281  3.52186037  3.75358105
#>   [17,]  5.88399550  7.9321118  8.18283151  8.23934387
#>   [18,]  7.19566394  9.6469086 10.34936627 10.84839799
#>   [19,]  3.17984313  3.2721020  3.34495884  3.38775090
#>   [20,]  6.75660003  7.2625041  7.47156384  7.53702198
#>   [21,]  5.16253652  6.2970272  6.72281298  6.94163303
#>   [22,]  3.31823337  3.7218387  3.93537474  4.13881317
#>   [23,]  4.36903102  4.4997609  4.51361933  4.51686928
#>   [24,]  8.32923493 11.3576003 14.38390844 16.54585132
#>   [25,]  4.68651650  6.6114556  7.74906230  7.98985149
#>   [26,]  2.25751022  2.7249365  3.12979726  3.48022646
#>   [27,]  3.05671257  4.3204952  5.19151248  5.74455348
#>   [28,]  2.94625580  3.3650419  3.42769411  3.48197394
#>   [29,]  1.75638492  2.2632984  2.59021519  2.77445985
#>   [30,]  5.28245895  6.2951567  6.69197567  6.97876731
#>   [31,]  3.58115074  4.8504175  5.83199942  6.03327609
#>   [32,]  4.81518783  5.7791364  6.41695083  6.96443556
#>   [33,]  1.31869834  1.7203187  1.82001081  1.83115697
#>   [34,]  2.27351042  2.8347349  3.22670565  3.39927057
#>   [35,]  5.13944036  6.6683411  7.61989722  8.12240588
#>   [36,]  9.65518486 11.8570833 12.64094758 12.67167637
#>   [37,]  7.37959293  7.8495664  8.24059463  8.53971152
#>   [38,]  4.46894301  5.9097196  6.18977193  6.22714184
#>   [39,]  4.56377173  6.3495343  7.81688544  8.20273969
#>   [40,]  6.43469033  7.6757343  8.67547865  9.43507539
#>   [41,]  3.22048640  4.4312020  4.73484870  4.98360583
#>   [42,] 10.53425867 12.8095822 13.02268722 13.21369346
#>   [43,]  3.26235555  3.7633620  3.82032228  3.82365388
#>   [44,]  3.47710841  3.6087146  3.69412534  3.69826129
#>   [45,]  8.57498023  8.9489781  9.30466857  9.41322327
#>   [46,]  2.37029137  2.6445075  2.88976020  2.98722816
#>   [47,]  8.18254648  9.9241222 11.01532536 11.07151954
#>   [48,]  3.90129274  5.0233433  5.55039802  5.94751027
#>   [49,]  1.77696518  2.3664991  2.62398058  2.83212746
#>   [50,]  2.75261953  3.0800466  3.24929947  3.38247958
#>   [51,]  5.52895538  6.8966224  7.69649945  7.99360822
#>   [52,]  4.25167323  4.9116583  5.00285012  5.03810933
#>   [53,]  7.38613635  8.7123167  9.33913271  9.79044984
#>   [54,]  5.43940393  6.2770667  6.77886849  6.97908830
#>   [55,]  3.32247545  4.7160300  5.19163558  5.22430344
#>   [56,] 11.81200476 12.4951547 12.74143217 12.88127655
#>   [57,]  2.18008617  2.6538806  2.76093501  2.80317490
#>   [58,]  3.36504392  4.0940565  4.53802823  4.68404810
#>   [59,]  4.36868216  5.5200201  5.63406785  5.68196468
#>   [60,]  3.67865990  4.1902928  4.59971947  4.74069383
#>   [61,]  5.75727452  6.3127398  6.80471148  7.26187897
#>   [62,]  8.38208737  8.8678687  8.93948380  8.94533200
#>   [63,]  2.59360458  3.0796088  3.39644418  3.51151742
#>   [64,]  3.74255858  4.8645007  5.67084080  6.15464599
#>   [65,]  3.85654539  4.4987876  4.93744931  5.31029771
#>   [66,]  3.59072664  4.8352883  5.61029807  6.10539358
#>   [67,]  2.16670494  2.9326241  3.10640405  3.13288890
#>   [68,]  3.67016457  5.0963700  6.46496909  7.05176845
#>   [69,]  7.88701959  8.6417344  9.07063855  9.09847466
#>   [70,]  4.20563107  5.4257051  6.07541344  6.62960734
#>   [71,]  1.66628947  2.3218737  2.79345507  2.83589317
#>   [72,]  3.48537746  3.6172637  3.73463488  3.78230772
#>   [73,]  7.36653279  7.8746501  8.34088346  8.64500912
#>   [74,]  1.93310100  2.2131347  2.27603416  2.33429934
#>   [75,]  6.04915599  7.4529062  8.29992370  8.99175838
#>   [76,]  0.92761870  1.0486844  1.15470712  1.24158200
#>   [77,]  2.69458550  3.0303936  3.10810525  3.12181058
#>   [78,]  7.11766864  8.6788179  9.35605986  9.63673619
#>   [79,]  7.06932007  9.0545370 10.29494749 10.96144089
#>   [80,]  4.18965576  4.6211839  4.70095760  4.70835250
#>   [81,]  2.73718099  3.3221898  3.63767997  3.90256465
#>   [82,]  5.42446676  5.9231589  6.13554545  6.30772039
#>   [83,]  3.39358225  3.7585244  3.94503925  3.98662446
#>   [84,]  1.99516896  2.7260575  3.39574755  3.59209628
#>   [85,] 10.56385701 12.3635627 12.69837535 12.79415479
#>   [86,]  6.22116457  6.9228528  7.13160584  7.30068511
#>   [87,]  2.04708416  2.1806337  2.27979514  2.33411218
#>   [88,]  2.54283088  2.8265101  2.87727081  2.91601486
#>   [89,]  4.70383506  5.1203514  5.42387469  5.56906099
#>   [90,]  3.39696385  3.6819455  3.84605961  3.90819987
#>   [91,]  2.42431527  2.9953102  3.22571205  3.36988752
#>   [92,]  4.22595576  4.9265192  5.51384622  5.55725449
#>   [93,]  9.84769186 12.8543913 13.83628599 13.87862109
#>   [94,]  8.56435433  9.0413643  9.34815804  9.37797053
#>   [95,]  2.78606595  3.5832480  3.77748553  3.78938046
#>   [96,]  3.94832127  4.4243597  4.51447599  4.52070235
#>   [97,]  4.85295327  5.4910281  5.63855134  5.65825966
#>   [98,]  7.66004184  9.9600623 11.82581792 13.58106269
#>   [99,]  6.77416431  7.1765087  7.35970470  7.45463347
#>  [100,]  2.75784458  3.7873022  4.42787400  4.80547058
#>  [101,]  3.42016882  4.4769691  5.45156216  5.62214404
#>  [102,]  3.60341533  4.4333548  5.04797320  5.63129189
#>  [103,]  3.10917375  3.7418397  4.34449778  4.56152311
#>  [104,]  3.33705447  3.9606573  4.37395997  4.39141804
#>  [105,]  3.55635198  4.1998337  4.52599041  4.57530723
#>  [106,]  4.77008494  6.4595193  6.76893112  6.77250299
#>  [107,]  3.14675652  3.6507283  3.89309426  4.07980915
#>  [108,]  2.27450371  2.4912639  2.66196246  2.71323958
#>  [109,]  7.63181874  8.6647119  9.60758966  9.67855639
#>  [110,]  2.60801323  3.2233575  3.59542612  3.63261855
#>  [111,]  1.64642437  1.8002576  1.84785904  1.85635356
#>  [112,]  7.16970813  8.5898360  8.91843086  9.02987895
#>  [113,] 15.76652977 16.6261006 16.88468762 17.07265235
#>  [114,]  2.81829562  3.2935411  3.70266601  4.04449192
#>  [115,]  2.62970102  3.3036295  3.66164399  3.98205632
#>  [116,]  1.86054518  2.1211301  2.31059543  2.35388200
#>  [117,]  3.02943885  3.7910921  4.35955769  4.81759219
#>  [118,]  1.55712805  1.8622117  2.06208998  2.20832724
#>  [119,]  4.93862029  5.5956449  5.87804054  5.90440459
#>  [120,]  4.34533369  5.3701764  6.22558983  6.54444517
#>  [121,]  0.59534587  0.6683968  0.73467073  0.78598602
#>  [122,]  5.01747866  6.5917483  6.69258144  6.77115386
#>  [123,]  2.03015431  2.2495465  2.40602722  2.40916632
#>  [124,]  2.66635917  2.6671602  2.66779537  2.66825016
#>  [125,]  3.89556150  5.1638834  5.81720311  6.01663304
#>  [126,]  3.73650372  4.7925394  5.41976102  5.54412339
#>  [127,]  2.88117631  3.1688909  3.41571609  3.48824655
#>  [128,]  3.91568218  4.4218161  4.76450029  5.05080675
#>  [129,]  4.38841344  5.3797389  5.63725152  5.80070192
#>  [130,]  6.30216507  6.7767788  7.07056366  7.23905593
#>  [131,]  7.77964875  8.4561331  8.81525656  8.97915200
#>  [132,]  0.85535993  1.0825069  1.19034335  1.27560573
#>  [133,]  2.01920633  2.4473576  2.53030262  2.56745188
#>  [134,]  4.73788270  5.3739158  5.63619540  5.74096567
#>  [135,]  3.91545116  4.1822850  4.37123449  4.40755325
#>  [136,]  3.29533295  4.1339309  4.69388042  4.76917518
#>  [137,]  2.50970964  2.8642402  3.18514775  3.47646761
#>  [138,]  3.64153999  4.0997708  4.52437734  4.78377926
#>  [139,]  5.06372926  5.5835682  5.68906575  5.74778867
#>  [140,] 10.00059206 11.9708464 12.20550576 12.29924982
#>  [141,]  1.79986519  2.3678714  2.71543793  2.74854465
#>  [142,]  4.53097795  5.4133883  6.07170198  6.33586622
#>  [143,]  2.08039773  2.1949971  2.29084490  2.37648040
#>  [144,]  2.68664920  3.6073240  4.03513844  4.05584489
#>  [145,] 10.05339424 13.8597962 14.39239453 14.41189902
#>  [146,]  1.35129408  1.5040250  1.62184957  1.70499604
#>  [147,]  1.78400335  2.1055669  2.21836691  2.24607591
#>  [148,]  2.83847950  3.5080953  3.93146506  4.08066881
#>  [149,]  8.24222844 10.4887205 12.44009397 14.13582091
#>  [150,]  3.52746478  4.2277695  4.73125080  5.21715312
#>  [151,]  8.53403919  9.6186014  9.71311158  9.77385468
#>  [152,]  3.83107944  5.5345290  5.75298341  5.91717160
#>  [153,]  5.31613242  7.1014994  8.53612126  8.72486473
#>  [154,]  3.10763268  3.2054492  3.27807692  3.32333633
#>  [155,]  0.21322303  0.2977606  0.34172380  0.36471341
#>  [156,] 13.78668700 15.3780040 16.23563358 16.40741683
#>  [157,]  1.72210744  1.8465492  1.89296333  1.90832060
#>  [158,] 10.85964003 12.0947714 12.23431120 12.27919587
#>  [159,]  8.00885505  8.4963114  8.88415178  9.26964589
#>  [160,] 10.34761912 10.5134695 10.64669910 10.69883174
#>  [161,]  3.04916583  3.2547295  3.32424373  3.37182541
#>  [162,]  3.30841465  4.8001966  5.00619551  5.06122435
#>  [163,]  2.46116630  3.0189085  3.10087532  3.12628950
#>  [164,]  6.98304205  7.7243493  8.19374573  8.25270572
#>  [165,]  3.68198974  4.0692288  4.38598767  4.59477742
#>  [166,]  7.21941721  8.6922894  9.47873920  9.89493574
#>  [167,]  5.42769918  5.8629257  6.07256072  6.09769958
#>  [168,]  4.73363250  5.1290517  5.31771457  5.43167409
#>  [169,]  2.24012867  2.5053634  2.68828587  2.83614840
#>  [170,]  8.82535462 10.1625533 10.32567511 10.36411833
#>  [171,]  7.76838092 11.1672649 12.25783753 13.00570362
#>  [172,]  3.99211844  5.3378817  5.42066792  5.42534147
#>  [173,]  2.83953550  3.5696289  4.23724269  4.38909073
#>  [174,]  2.50173496  2.5836692  2.61570503  2.63684363
#>  [175,]  4.54298753  5.3968568  5.44853566  5.48661252
#>  [176,]  2.57754113  3.2225791  3.33185193  3.41410854
#>  [177,]  1.96207382  2.4214269  2.85221057  2.95827847
#>  [178,]  5.15000012  6.1891078  7.05334868  7.38086242
#>  [179,]  3.00009643  4.1878576  5.06348234  5.31863062
#>  [180,]  2.74895702  2.9510791  3.13048985  3.23544695
#>  [181,]  0.47265647  0.5407046  0.60732610  0.61423450
#>  [182,]  5.52153031  7.6117235  9.05495960  9.78928155
#>  [183,]  6.82560373  7.8813696  8.27558187  8.29989882
#>  [184,]  7.51328323  8.0162035  8.39978068  8.77986517
#>  [185,]  6.31015176  6.6389280  6.65556043  6.66849330
#>  [186,]  3.32064654  4.5441487  4.88222936  4.98631808
#>  [187,]  2.56033684  3.2349377  3.90921429  4.25894936
#>  [188,]  3.83626877  4.7410925  4.74810309  4.74817389
#>  [189,]  6.25503409  6.6552805  7.01794337  7.28594125
#>  [190,]  5.63921444  6.4319401  6.90050932  7.23300575
#>  [191,] 12.00690653 12.7243202 13.42478120 13.72656411
#>  [192,]  6.05037007  6.7688666  7.47819001  7.98484083
#>  [193,]  0.72583490  0.9980742  1.21087424  1.25142115
#>  [194,] 12.00866267 14.4234909 16.49768540 17.76135635
#>  [195,]  6.93581858  8.0206629  8.59500300  8.72876952
#>  [196,]  3.75614489  4.3555498  4.66670885  4.71304597
#>  [197,]  3.13513881  3.3426860  3.51535572  3.64131885
#>  [198,]  1.88811289  2.4207520  2.74562480  2.88406706
#>  [199,]  9.72073734 13.0607275 14.50213642 15.12150834
#>  [200,]  2.77505546  2.9022264  2.98100392  3.03274050
#>  [201,]  5.25697063  6.2703132  6.66833653  6.87049294
#>  [202,]  2.89525803  3.4438593  3.47798296  3.48469139
#>  [203,]  9.37480133 11.0625103 12.18362880 12.44767300
#>  [204,]  7.71447960  8.6802505  9.10880473  9.39655891
#>  [205,]  4.20150336  5.6689847  6.41406702  6.44995022
#>  [206,]  1.11838942  1.2369087  1.34371376  1.37341726
#>  [207,]  1.00759798  1.2819845  1.48645052  1.56105341
#>  [208,]  3.45848251  4.8624796  5.25521525  5.58325434
#>  [209,]  4.42160928  4.9052636  5.04029147  5.17115139
#>  [210,]  5.71779583  7.1309087  7.94333182  8.16438724
#>  [211,] 10.55498680 11.4818715 11.57363274 11.58067650
#>  [212,]  5.54943155  6.1219594  6.47337848  6.47511114
#>  [213,]  4.44271467  5.4008827  5.47731387  5.48464223
#>  [214,]  4.33051363  4.4248939  4.49092794  4.54720435
#>  [215,]  3.25920572  3.3878399  3.41590666  3.42096267
#>  [216,]  3.81907564  4.6353022  5.03951818  5.04818434
#>  [217,] 12.57384652 14.4152128 16.15380932 16.55942746
#>  [218,]  5.62213246  6.4865618  7.17826970  7.19947787
#>  [219,]  2.61534034  3.5262657  4.13828488  4.43650773
#>  [220,]  4.69701476  5.6249490  5.88973693  6.13041236
#>  [221,]  1.29933808  1.7614731  1.87558781  1.91785288
#>  [222,]  3.62538596  4.0150072  4.38435614  4.46272714
#>  [223,]  3.32067617  3.8581490  4.24043863  4.40001937
#>  [224,]  4.69065521  6.0878241  7.44963425  8.40691732
#>  [225,]  9.53310696 10.7739238 11.21785589 11.51798032
#>  [226,]  6.70993330  8.8497949  9.97166128 10.81256387
#>  [227,]  3.83793228  5.0271223  5.63499277  5.92470797
#>  [228,]  2.49213476  3.4786317  4.45350174  5.13772938
#>  [229,]  4.37386689  4.8208709  5.02933136  5.07562116
#>  [230,]  2.48303193  3.0914578  3.64989481  3.83568661
#>  [231,]  8.44197835  9.7164001 10.43882514 10.62238299
#>  [232,]  1.60603172  2.1756196  2.71707310  3.04548349
#>  [233,]  9.09949569 10.0411270 10.14108801 10.14415443
#>  [234,] 11.37137729 12.5273756 12.82396305 12.84072126
#>  [235,]  3.26511651  3.5211865  3.73239467  3.78895076
#>  [236,]  4.98654249  6.3251407  6.45470582  6.46611427
#>  [237,] 12.69552204 16.2553925 16.72082554 16.84637858
#>  [238,]  3.37513156  4.4138300  4.86475769  5.08676939
#>  [239,]  2.81745831  3.7601606  4.32443921  4.53308761
#>  [240,]  1.01805457  1.1910305  1.22334533  1.24872843
#>  [241,]  4.66936637  5.5618782  5.88132008  5.93572859
#>  [242,]  2.02508472  2.0663971  2.10350365  2.13488368
#>  [243,]  3.69594480  3.9966931  4.17302239  4.17736117
#>  [244,]  4.26709730  4.5874660  4.79858831  4.96744793
#>  [245,]  2.20650682  2.4696250  2.71016541  2.88849838
#>  [246,]  1.91424510  2.1279734  2.31334345  2.45316645
#>  [247,]  5.26624248  6.1432410  6.59066682  6.75959925
#>  [248,]  3.68354631  4.3257832  4.75263376  5.06059658
#>  [249,]  1.85319220  2.1851979  2.23606792  2.26451291
#>  [250,]  2.70815818  3.5702777  3.99316951  4.02274607
#>  [251,]  3.50406423  4.8824489  5.87301774  6.03121607
#>  [252,]  4.53559275  5.1922864  5.50635604  5.78739981
#>  [253,]  8.08609001 10.5870513 11.31735466 11.96555492
#>  [254,]  3.56648000  3.6917167  3.81168779  3.83885334
#>  [255,]  5.64599528  5.8269253  5.95766204  6.02128956
#>  [256,]  5.92063542  6.4391721  6.55738748  6.61060462
#>  [257,] 10.21627065 11.4140573 11.84383387 12.14027102
#>  [258,]  1.14604772  1.2317308  1.24938830  1.25705405
#>  [259,]  5.17646940  5.4239774  5.50779452  5.57057683
#>  [260,]  3.94091254  4.0708429  4.16438119  4.25373000
#>  [261,]  6.19278858  8.0293346  9.36297420  9.54780820
#>  [262,]  2.66923261  3.0811370  3.14568088  3.18371063
#>  [263,]  3.09221330  3.5743917  3.88192710  3.91325315
#>  [264,]  5.32980569  6.3626249  7.09824702  7.20485880
#>  [265,]  6.60892269  7.8870494  8.73044201  9.48418715
#>  [266,] 10.12088400 11.8207430 12.34107629 12.76136837
#>  [267,]  1.13068023  1.2753514  1.39440019  1.50380685
#>  [268,]  1.23238121  1.4677243  1.64889275  1.66307804
#>  [269,] 11.30237611 13.6109418 14.12544509 14.35448825
#>  [270,] 10.01058001 10.9218437 11.30731448 11.31371787
#>  [271,]  0.97879296  1.2520315  1.44789227  1.48074422
#>  [272,]  6.42720593  6.4985851  6.51043062  6.52093876
#>  [273,]  7.56748121  7.9491335  8.32947803  8.70704299
#>  [274,]  6.90086612  8.6010529  9.11350131  9.38619073
#>  [275,]  6.59597104  7.4729271  8.19601435  8.59155454
#>  [276,]  6.00995566  6.7449308  7.04670327  7.15525000
#>  [277,]  4.72781386  5.4709199  6.17866611  6.31271151
#>  [278,]  6.14704907  7.9516159  9.17537178 10.25729027
#>  [279,]  5.06296496  5.9150555  6.55418839  6.59532544
#>  [280,]  5.23854008  5.4013681  5.43888028  5.46706428
#>  [281,]  3.50538391  4.1382776  4.63186329  4.68064950
#>  [282,]  1.21126367  1.5065740  1.67801212  1.80444029
#>  [283,]  3.93470225  4.7716818  5.54429613  5.65153181
#>  [284,]  2.33176183  2.6612417  2.69844955  2.70035477
#>  [285,]  8.55791494  8.9815601  9.27496387  9.55087674
#>  [286,]  6.14969491  8.6371359  9.14752372  9.48520040
#>  [287,]  4.26002112  4.8233827  5.12876938  5.18433384
#>  [288,]  3.75357633  4.8438203  5.79134427  6.73079118
#>  [289,]  4.34698347  4.6672908  4.93330122  5.00313128
#>  [290,]  4.93767654  5.9724995  6.65837141  6.71942313
#>  [291,]  3.70947663  4.3681354  4.44914798  4.52726723
#>  [292,]  5.74865768  7.9176967  8.07193932  8.12060754
#>  [293,]  1.65524573  1.9323350  2.08798120  2.09919691
#>  [294,]  4.87152068  5.1924435  5.43581963  5.62926285
#>  [295,]  1.34986808  1.5270537  1.61128778  1.68379158
#>  [296,]  1.58263812  2.1950883  2.69320643  3.00738344
#>  [297,]  5.62690198  6.4996035  6.73945161  6.80658689
#>  [298,]  3.76355262  4.7502220  5.57568264  5.71925838
#>  [299,]  0.54928609  0.5929713  0.60025690  0.60693700
#>  [300,]  1.56749925  1.9466943  2.18121335  2.35378395
#>  [301,]  4.39787074  4.5128910  4.62449368  4.69112293
#>  [302,]  0.99447841  1.4726341  1.93298134  2.18993898
#>  [303,]  0.56183935  0.7630481  0.89412318  0.96414595
#>  [304,] 11.14553722 14.3297081 14.61930947 14.65206089
#>  [305,]  3.71164814  3.9461388  4.02053207  4.08114873
#>  [306,]  0.85262804  0.9793905  1.03770705  1.06101725
#>  [307,]  2.32050489  2.4347783  2.52599770  2.56266660
#>  [308,]  1.17795438  1.3024810  1.40771445  1.51007743
#>  [309,]  1.90936953  2.3746015  2.61156831  2.79541772
#>  [310,]  0.99208901  1.0383936  1.06426094  1.07227907
#>  [311,]  6.98379582  9.7182157 10.58654259 10.82832342
#>  [312,]  2.08124656  2.1061780  2.11498985  2.11957455
#>  [313,]  3.33866830  4.4084969  5.22872235  5.67944585
#>  [314,]  1.46810961  1.8479881  2.10292751  2.13693698
#>  [315,]  4.02697817  4.9603099  5.40463883  5.44711084
#>  [316,]  2.22855484  3.2851155  3.58266442  3.63781246
#>  [317,]  5.50062947  6.1560038  6.51103088  6.70482996
#>  [318,]  1.04340001  1.4566573  1.86436186  2.17233160
#>  [319,]  5.28894429  6.4776308  6.59218899  6.61637785
#>  [320,]  4.94635095  6.3014619  7.25990687  7.32370194
#>  [321,]  3.61860280  4.7238485  5.25727330  5.43121759
#>  [322,]  2.90560799  3.9160078  4.04547621  4.05724915
#>  [323,]  9.97536454 11.2222717 12.15881212 12.17640329
#>  [324,]  1.82687153  2.1673183  2.49024600  2.50957557
#>  [325,]  2.88119099  3.8288914  4.20293335  4.42355563
#>  [326,]  4.75653698  5.1232230  5.46916482  5.47001884
#>  [327,]  6.88146154  7.0006078  7.07860000  7.10262851
#>  [328,]  2.46660921  3.5041950  3.98142295  4.24589689
#>  [329,]  4.01956360  5.1968471  5.52924477  5.65253069
#>  [330,]  5.37963480  7.2882905  8.23900549  8.82882820
#>  [331,]  2.94140873  3.6934831  4.27290509  4.38618994
#>  [332,]  5.95972524  7.9868569  8.36619130  8.40995768
#>  [333,]  2.22785809  2.6561002  2.90272168  3.03249757
#>  [334,]  1.36797501  1.8143626  2.20922197  2.33559559
#>  [335,]  2.85848028  3.4014429  3.66234570  3.86682489
#>  [336,]  3.42432468  3.8921137  4.23346373  4.40181635
#>  [337,]  4.63915483  6.2510895  7.37794629  8.01087920
#>  [338,]  1.79673173  2.1041543  2.36729597  2.60861381
#>  [339,]  3.07374716  3.6126561  4.01895675  4.10344154
#>  [340,]  2.77012690  3.1626549  3.55303674  3.85274011
#>  [341,]  4.73134412  5.0059935  5.09513103  5.15973873
#>  [342,]  1.50711776  2.0310753  2.41104818  2.47994405
#>  [343,]  4.11609780  5.2735394  6.11721120  6.17500316
#>  [344,]  3.74079178  3.8382496  3.91428250  3.91810349
#>  [345,]  7.08378461  9.2889240  9.51972522  9.63311031
#>  [346,]  5.24515590  6.6460981  7.97150242  8.98473352
#>  [347,]  0.77780408  1.0143990  1.12550765  1.14330479
#>  [348,]  8.16317265  9.1551964 10.04055381 10.35439343
#>  [349,]  7.95034834  9.4877485 10.93313997 11.84276555
#>  [350,]  6.59342562  7.3409358  7.65101279  7.85044729
#>  [351,]  9.67643671  9.8142393  9.89309043  9.96182946
#>  [352,]  2.58744314  2.8985492  3.13003453  3.25905991
#>  [353,]  1.97032430  2.0752135  2.14842595  2.18957901
#>  [354,]  2.38841550  2.7022110  2.72311735  2.74281010
#>  [355,]  3.89644972  4.7724431  4.86219543  4.87509857
#>  [356,]  4.45879206  5.1752759  5.32411447  5.38019504
#>  [357,]  1.34914579  1.8994352  2.03735440  2.13181267
#>  [358,]  5.36591465  5.6472096  5.66770862  5.68783908
#>  [359,]  3.57740502  4.0056829  4.41286308  4.46011914
#>  [360,]  4.35973817  4.6786825  4.83507711  4.88192284
#>  [361,] 10.12352618 12.4613486 12.79814623 12.93815839
#>  [362,] 11.16896958 12.6148687 13.64727169 13.72268073
#>  [363,]  4.69664726  6.0129089  6.41340009  6.65747921
#>  [364,]  3.47731389  4.4749131  5.01302156  5.36767456
#>  [365,]  1.79540150  2.5215906  3.14788166  3.30962549
#>  [366,]  9.78361541 10.1514194 10.37689747 10.39373298
#>  [367,]  1.46536099  1.6421195  1.78000094  1.81546229
#>  [368,]  1.90492098  2.1225186  2.18815754  2.19047432
#>  [369,]  5.43116969  7.0022236  8.44650742  8.81979835
#>  [370,]  2.26361770  2.9296833  3.10216257  3.11243107
#>  [371,]  6.71903594  7.5923939  8.16730565  8.56939803
#>  [372,]  4.96265156  5.8389584  5.92166422  5.97878258
#>  [373,]  8.32251463  9.1425933  9.88366118 10.61319584
#>  [374,]  2.62836541  3.5084952  4.20825121  4.82361455
#>  [375,]  1.33686692  1.5564046  1.57827072  1.59821412
#>  [376,]  2.06963714  2.3908139  2.58878777  2.59246275
#>  [377,]  1.29654088  1.5326775  1.66329233  1.76915060
#>  [378,]  6.86597398  8.9915898 11.01616843 12.18532691
#>  [379,]  2.92922932  3.5278673  3.65571886  3.78141989
#>  [380,]  1.44001969  1.5247367  1.55477831  1.57660430
#>  [381,]  1.66888230  1.9529404  2.16615957  2.33535118
#>  [382,]  9.50137193 10.6948192 11.30631234 11.53607701
#>  [383,]  7.22416568  8.3902831  9.43669852 10.10061059
#>  [384,]  3.59688166  4.4300723  4.47479758  4.47734391
#>  [385,]  6.52448433  7.1073055  7.42437003  7.68978774
#>  [386,]  3.60560541  4.7551518  4.91026830  4.97037428
#>  [387,]  4.51250310  5.1050296  5.56731187  5.59246993
#>  [388,]  8.48820764 10.6699109 11.82820435 12.12668338
#>  [389,]  6.65195716  7.6009471  7.74884593  7.84985825
#>  [390,]  2.64648988  3.3867007  3.92560331  4.24520487
#>  [391,]  9.47897662 11.3198870 11.81436067 11.96504675
#>  [392,]  3.34740177  4.1496144  4.89763477  5.44422221
#>  [393,]  4.55741688  5.5289325  5.95029462  6.29411852
#>  [394,]  1.18912562  1.5105048  1.67711488  1.80670465
#>  [395,]  2.66960006  3.2903956  3.85684572  4.13321911
#>  [396,]  6.21385349  6.8583187  7.40228676  7.78189882
#>  [397,]  2.68460540  3.3152406  3.37582392  3.41149228
#>  [398,]  3.27103046  3.5139964  3.72230888  3.91150689
#>  [399,]  6.54754988  8.5431215 10.25121170 10.39388458
#>  [400,]  6.30299421  6.9704819  7.11183945  7.23901949
#>  [401,]  1.33004295  1.7328102  2.01505830  2.17040567
#>  [402,] 12.39480499 13.7231187 14.91664899 15.09065350
#>  [403,]  4.26987397  5.4121718  5.61354312  5.77103237
#>  [404,]  1.34647161  1.5279767  1.66424346  1.71195456
#>  [405,]  2.97625276  3.7469282  3.83468344  3.84643034
#>  [406,]  2.85514361  3.1011489  3.33695069  3.39948843
#>  [407,]  1.64937804  2.0717617  2.20617163  2.28470306
#>  [408,]  2.99243263  3.1904957  3.35484984  3.39588974
#>  [409,]  5.74361797  8.3109143  9.69093302  9.88337369
#>  [410,]  3.87460121  3.9749923  4.03038801  4.04968746
#>  [411,]  1.28153727  1.3909562  1.41399491  1.43010582
#>  [412,]  3.65870198  4.5618011  5.37796671  5.55180254
#>  [413,]  3.74820970  4.5647106  4.83557838  4.90693713
#>  [414,]  0.88068564  1.2599855  1.46915188  1.48484899
#>  [415,]  4.49296925  5.9951135  7.08693510  7.21158944
#>  [416,]  3.82151438  4.3850322  4.94022220  5.05283678
#>  [417,]  2.34489846  2.8567972  3.03878438  3.09410432
#>  [418,]  1.64027091  1.8955412  2.06803191  2.18814612
#>  [419,]  2.47885587  2.8841855  3.17738552  3.38256106
#>  [420,]  4.02424912  4.6698639  5.26438663  5.40584671
#>  [421,]  0.81348517  0.9845339  1.05181943  1.06598807
#>  [422,]  3.24932409  3.5724651  3.81958209  3.99359786
#>  [423,]  6.73244224  7.6377261  8.11786178  8.39258534
#>  [424,]  4.79447886  5.1530123  5.49392703  5.69434180
#>  [425,]  0.97078843  1.0572162  1.08912565  1.10199946
#>  [426,]  7.96111004  9.2191469  9.77065815 10.12407690
#>  [427,]  1.71494102  1.8896373  2.03244631  2.10209272
#>  [428,]  5.39952628  6.0871557  6.71153086  6.99137652
#>  [429,]  2.62510997  3.6656587  4.16772564  4.27511089
#>  [430,]  3.06194544  3.5299712  3.82828030  4.04048165
#>  [431,]  4.01778494  4.3717057  4.57273084  4.68977324
#>  [432,]  1.42342770  1.7472718  1.90458818  2.04359269
#>  [433,]  2.10485188  2.7063886  3.28203493  3.60338197
#>  [434,]  6.10136105  6.9567484  7.07817059  7.11613517
#>  [435,]  4.65554409  5.5863818  6.50522154  6.89039214
#>  [436,]  7.61606977  8.2974603  8.84149344  9.18506316
#>  [437,]  3.53302944  4.2570275  4.85623177  5.36361400
#>  [438,]  3.27141657  4.4066268  5.07343445  5.20670855
#>  [439,]  5.16202324  6.9316235  7.62557946  7.62612662
#>  [440,]  5.68005379  6.2465077  6.34235472  6.42948131
#>  [441,]  2.15217061  2.5065517  2.76055269  2.99313024
#>  [442,]  4.24953048  4.5753985  4.69038013  4.78381560
#>  [443,]  7.61411694 10.2454156 10.85041612 10.89037425
#>  [444,]  2.78215927  3.0132688  3.16007738  3.26504643
#>  [445,]  3.00321109  3.5282922  3.85792507  3.97003550
#>  [446,]  6.71216348  7.1055098  7.46843557  7.81577047
#>  [447,]  1.28299896  1.5093570  1.58927125  1.64846765
#>  [448,]  2.44221091  2.9495951  3.33360973  3.57073296
#>  [449,]  2.42820003  3.1804482  3.74797688  3.82375340
#>  [450,]  5.71918842  6.3812124  6.63344557  6.81149902
#>  [451,]  8.55342610 12.0165734 12.98760583 13.43082377
#>  [452,]  2.13545649  2.6999584  3.11894055  3.27345514
#>  [453,]  2.81671102  3.6356633  3.75034135  3.79165284
#>  [454,]  6.17355977  6.9872204  7.41826061  7.47255777
#>  [455,]  5.22456356  5.9365552  6.33952338  6.35129309
#>  [456,]  3.87604593  4.5838819  5.21229586  5.60950035
#>  [457,]  7.73106389  9.0803729  9.60852067 10.04273102
#>  [458,]  1.88428267  2.4594517  2.86692151  3.13545500
#>  [459,]  4.66770133  5.0008331  5.29050694  5.53852013
#>  [460,]  4.06952112  4.7478697  5.23094634  5.51509388
#>  [461,]  6.29248164  6.4371207  6.57098304  6.64200622
#>  [462,]  3.81311088  4.0750591  4.16329802  4.21397051
#>  [463,]  9.01596734  9.3344711  9.54484302  9.63930427
#>  [464,]  4.58311963  5.5855367  5.92984550  6.07723891
#>  [465,]  9.29844460 12.2878495 13.59454026 14.61512453
#>  [466,]  4.44477472  4.9092584  5.32484706  5.39151260
#>  [467,]  7.70191022 10.2332563 12.43381500 12.77702307
#>  [468,]  4.32165950  4.8707728  5.04289232  5.13918686
#>  [469,] 11.98806661 12.1624696 12.32723670 12.42438931
#>  [470,]  2.61012242  3.5332801  4.07419918  4.12979186
#>  [471,] 10.60850066 13.9889894 15.87772143 16.27810910
#>  [472,]  1.01640668  1.1600021  1.21881833  1.25594219
#>  [473,]  2.69832535  2.9699707  3.12567717  3.25415211
#>  [474,]  2.52040524  2.7236087  2.90693311  2.97972711
#>  [475,]  6.93164221  7.9955832  9.05166699  9.21021436
#>  [476,]  3.48062471  4.4466916  5.40872329  5.75921126
#>  [477,]  1.57180615  2.3060834  2.93164441  3.29908369
#>  [478,]  5.83164410  6.8736430  7.45204760  7.73629603
#>  [479,]  5.44541359  6.6782487  6.97460858  7.17309432
#>  [480,]  3.79137025  4.4455875  4.72286832  4.72695766
#>  [481,] 14.46552149 16.1744077 16.60389224 16.73745072
#>  [482,]  2.96055153  3.7755249  4.56492421  5.06531874
#>  [483,]  8.94629378  9.8542230 10.05857486 10.14898016
#>  [484,]  2.51165991  2.7363209  2.89322280  2.96150260
#>  [485,]  2.95332487  3.3469287  3.50349965  3.54471346
#>  [486,]  1.61459675  2.0403656  2.16968715  2.19366686
#>  [487,] 10.52644629 11.8746639 12.44226201 12.45500713
#>  [488,]  9.28754385  9.6304587  9.80413045  9.83080380
#>  [489,]  2.82111349  3.8798755  4.72122005  5.14415330
#>  [490,]  6.44631189  7.3499180  7.57446147  7.70364334
#>  [491,]  4.48719163  4.7671597  4.85844944  4.89594944
#>  [492,]  6.06257613  7.0385912  7.83846989  7.94204270
#>  [493,]  1.10072003  1.5347256  1.70358180  1.78715888
#>  [494,]  7.62255697  8.1819373  8.62527349  8.71331922
#>  [495,]  3.62606649  4.4761757  5.03133658  5.16764441
#>  [496,]  1.49531423  1.5184565  1.52164153  1.52463783
#>  [497,]  5.77009132  6.0856070  6.09303693  6.09479881
#>  [498,]  6.72241009  7.6364396  7.73506505  7.82705776
#>  [499,]  5.65516106  6.8827808  7.85775350  8.22430888
#>  [500,]  6.06420450  7.1180522  7.96990314  8.58341257
#>  [501,]  8.64408375  9.7766777  9.99614255 10.01331448
#>  [502,]  1.63251615  1.7130094  1.77326397  1.82362086
#>  [503,]  4.13158415  4.7073349  4.75667106  4.80412769
#>  [504,]  0.86427614  1.2714294  1.44285380  1.58725383
#>  [505,]  9.52861780 11.2524512 11.45366911 11.58371879
#>  [506,]  1.77192176  2.3155106  2.68471809  2.76687527
#>  [507,]  7.39495346  9.1914971 10.46069839 10.90776873
#>  [508,]  4.78316042  6.2425743  7.23619030  7.39447693
#>  [509,]  1.59369345  2.0910136  2.22924762  2.32835356
#>  [510,]  2.87260837  3.2083116  3.38967784  3.49995120
#>  [511,]  2.51500461  3.5277692  3.96041050  3.99458354
#>  [512,]  4.88552007  5.9594211  6.56325605  6.65268833
#>  [513,]  1.32614918  1.6620535  1.68694831  1.68878392
#>  [514,]  2.70606822  3.5166205  4.25387894  4.78224527
#>  [515,]  4.64413594  5.1131161  5.26784033  5.31236829
#>  [516,]  2.54321252  2.9657614  2.98345333  2.99650192
#>  [517,]  7.19688312  8.4813185  9.15454294  9.80946565
#>  [518,]  7.20149966  8.6462163  8.83995676  8.93717898
#>  [519,] 14.55901570 14.9722136 15.05187857 15.12058894
#>  [520,]  5.71725878  7.0882915  7.80341411  7.98682825
#>  [521,] 10.88622959 12.3449194 13.19675078 13.33266492
#>  [522,]  5.06403546  5.1789495  5.24171106  5.29340849
#>  [523,]  2.26202764  2.5822644  2.88146087  2.88239187
#>  [524,]  4.56368065  5.6902397  5.82535979  5.92243755
#>  [525,]  5.88131895  6.5511239  7.14712569  7.37340261
#>  [526,]  6.87743936  7.6891141  8.41655191  8.60341817
#>  [527,]  4.48881824  5.2764314  5.80525348  6.14743296
#>  [528,]  3.62272333  3.8923878  4.02847169  4.15229559
#>  [529,]  3.15809436  4.6349542  5.52457693  5.75320907
#>  [530,]  6.01905919  6.2694068  6.42459250  6.46252715
#>  [531,]  6.06877848  7.1413264  7.37714267  7.40169743
#>  [532,]  1.74572685  1.7860996  1.81332908  1.81978826
#>  [533,]  3.76599213  4.2552811  4.54585485  4.61492323
#>  [534,]  5.55118449  7.9713709  9.47005327  9.78071212
#>  [535,]  8.78355618 11.6087929 12.87626742 14.01413578
#>  [536,]  4.17396147  4.7940165  5.15060274  5.32959565
#>  [537,]  2.31737985  2.7813753  3.22654603  3.58365015
#>  [538,]  2.25199591  2.7059099  2.98617699  3.19840022
#>  [539,]  2.99524057  3.5566974  4.04666823  4.22659603
#>  [540,]  8.76325191  9.1460869  9.51847789  9.54990498
#>  [541,]  2.33297631  2.6533302  2.80352529  2.87095062
#>  [542,]  5.69678610  6.5237655  6.69689324  6.74739978
#>  [543,]  5.67189232  6.8444833  7.22479929  7.60283626
#>  [544,]  3.29050900  4.0247139  4.25849065  4.31117875
#>  [545,]  2.55459478  2.7494198  2.92596752  2.93895362
#>  [546,]  3.38067044  3.7853147  3.92752789  4.01533050
#>  [547,]  1.87590272  2.6079237  3.19000436  3.39939019
#>  [548,] 10.25535231 11.1972384 11.68380858 11.76322568
#>  [549,]  4.59496425  5.3415825  5.94135817  6.36439474
#>  [550,]  3.71071693  4.5449469  4.94427255  5.30362409
#>  [551,]  2.81595454  3.5399491  4.13299372  4.64615643
#>  [552,]  3.08214752  3.8706051  4.22328417  4.30925805
#>  [553,]  1.95262869  2.8007998  3.33802921  3.58866289
#>  [554,]  6.05753873  7.0895536  7.78870943  8.04512681
#>  [555,]  3.33982246  4.2792821  4.67495090  4.92527918
#>  [556,]  3.03380096  3.5373873  3.80679548  3.81240480
#>  [557,]  3.45115019  3.6818873  3.80612112  3.86392496
#>  [558,]  1.59639494  1.9967777  2.38980833  2.69173534
#>  [559,]  5.51713582  6.2580188  6.97877520  7.11784928
#>  [560,]  8.10087229  8.9839609  9.33619199  9.60863106
#>  [561,]  3.41824937  3.6208811  3.68723374  3.73706374
#>  [562,]  1.74791953  1.9574997  2.05639352  2.10079462
#>  [563,]  2.64575614  3.4667298  4.04390748  4.22394721
#>  [564,]  6.32060517  8.2132112  8.88301579  9.52014217
#>  [565,]  5.18516672  5.8809709  6.37767863  6.85596086
#>  [566,]  1.83492373  2.0313611  2.22062413  2.29840787
#>  [567,]  3.97543782  4.7962267  5.32827941  5.72882508
#>  [568,]  1.74644628  1.8400928  1.92851967  1.93840598
#>  [569,]  3.90889199  4.0995452  4.21407679  4.24210973
#>  [570,]  2.00527158  2.1127304  2.16887130  2.17607198
#>  [571,]  1.77141006  2.1630714  2.53456806  2.63801854
#>  [572,]  1.65382379  2.0296655  2.08705772  2.13323828
#>  [573,]  1.80790256  1.9201575  1.99765503  2.05301824
#>  [574,]  2.50943550  2.7191732  2.84169345  2.90530007
#>  [575,]  4.41356777  5.4786291  6.21028749  6.85551210
#>  [576,]  7.25222510  7.8094138  8.03670631  8.24837491
#>  [577,]  5.52537526  6.4484904  6.88689570  7.10736605
#>  [578,]  4.61655772  5.2759444  5.30066220  5.30753448
#>  [579,]  3.08453789  3.7465580  4.15233571  4.34475878
#>  [580,]  4.11289728  5.0174206  5.78246756  5.85727962
#>  [581,]  2.06137837  2.8293172  3.15725117  3.29727277
#>  [582,]  1.11685743  1.2497064  1.27331593  1.28924186
#>  [583,]  6.90773584  7.6483531  8.23882484  8.57916384
#>  [584,]  0.74578088  0.8108986  0.86981333  0.90199552
#>  [585,]  3.70325064  4.8730054  5.23769653  5.25852214
#>  [586,]  3.70070684  4.6503947  5.14750576  5.51933994
#>  [587,]  3.18443021  3.7722516  4.05958039  4.15410469
#>  [588,]  4.61850247  4.9879880  5.26358624  5.47736096
#>  [589,]  6.09820263  6.5228885  6.53296136  6.54293194
#>  [590,]  4.56261981  5.7994519  6.78930934  7.16073522
#>  [591,]  5.01266721  5.1967313  5.32849142  5.38111188
#>  [592,]  2.26849347  2.6655838  2.99863352  3.10229464
#>  [593,]  8.36498238  8.7011846  8.80350373  8.80871790
#>  [594,]  4.59986060  6.2296810  6.60766678  6.70929896
#>  [595,]  7.01464693  9.0900808 10.16651533 10.28974659
#>  [596,]  3.34603836  3.5547891  3.75061469  3.81235916
#>  [597,]  4.16376594  4.4474973  4.57173062  4.66810599
#>  [598,]  0.85821717  1.0813664  1.27969754  1.32973406
#>  [599,]  2.30436228  3.3090489  3.76009659  3.79277278
#>  [600,]  6.56892691  7.1405509  7.21775117  7.22854890
#>  [601,]  4.38894214  5.4233618  6.11637696  6.71441752
#>  [602,]  3.55564869  4.0558121  4.42990275  4.44129622
#>  [603,]  6.81912478  7.2751207  7.56189311  7.70902669
#>  [604,]  4.99746404  5.3469906  5.64946733  5.65399590
#>  [605,]  6.36665323  6.8063017  7.21404513  7.36601282
#>  [606,]  1.47169346  1.9934409  2.38416009  2.39559445
#>  [607,]  2.34891541  2.6012955  2.76302843  2.83223823
#>  [608,]  4.76984786  6.0330129  6.82509073  7.56751242
#>  [609,]  2.96248589  3.6854763  4.10294862  4.18752941
#>  [610,]  2.87733858  3.1157606  3.23964044  3.32050833
#>  [611,]  3.34234975  3.8749514  4.25963850  4.34170067
#>  [612,]  4.81212380  6.3944171  7.14715152  7.40501002
#>  [613,]  3.82698225  4.0398092  4.14561175  4.14694177
#>  [614,]  6.88499325  7.5273948  7.79273048  7.88439248
#>  [615,]  6.53177587  7.9463683  8.22694654  8.47618760
#>  [616,]  9.61001497 12.1068385 13.64147697 15.05176049
#>  [617,]  3.49658236  3.8345802  4.15579799  4.21972028
#>  [618,]  3.67613946  5.0402447  5.63025788  5.75737510
#>  [619,]  2.66668931  3.0668397  3.23649608  3.28595729
#>  [620,]  3.21491756  3.5331158  3.78495656  3.89016058
#>  [621,] 10.21814778 11.0015473 11.53170943 12.03626436
#>  [622,]  3.12826700  4.4524708  5.68444968  6.06998399
#>  [623,]  9.83849049 10.1245924 10.28001554 10.43161851
#>  [624,]  7.20351577  9.2007971  9.99604185 10.15659748
#>  [625,]  6.32197509  8.3649135  9.55462948 10.52705777
#>  [626,]  8.07437035  8.9485414  9.19900090  9.41705291
#>  [627,]  2.30836254  2.6831324  2.92266856  3.15056786
#>  [628,]  4.19903093  4.8215833  4.97954513  5.04843718
#>  [629,]  7.20880898  7.4256552  7.56537274  7.69062217
#>  [630,]  3.15636443  4.1103483  4.63000784  4.65153542
#>  [631,]  3.25934968  4.1947749  4.74302584  4.84494663
#>  [632,]  0.88329821  1.1519852  1.25435190  1.31413584
#>  [633,]  3.01672916  3.6301532  3.96856395  4.14225713
#>  [634,]  3.80600771  4.2060613  4.44955820  4.52269477
#>  [635,]  1.66513516  2.2106387  2.28204555  2.30243888
#>  [636,]  3.13320321  4.0651649  4.33429606  4.36986285
#>  [637,]  5.56703749  7.0513635  7.62335638  7.68064793
#>  [638,]  3.37227217  3.6443304  3.73760447  3.78979059
#>  [639,]  3.10647534  3.1850785  3.19587757  3.20531885
#>  [640,]  5.89945916  6.6130178  6.96513913  7.27865233
#>  [641,]  9.88911391 10.9842297 11.43658057 11.55627238
#>  [642,]  3.12613893  3.6284874  4.04296912  4.25415854
#>  [643,]  3.59893453  4.4643436  5.10373108  5.17119362
#>  [644,]  3.07388113  3.2332491  3.29520359  3.31627356
#>  [645,]  4.35690028  4.7875183  4.83110069  4.84306102
#>  [646,]  3.16373956  3.2636386  3.35592602  3.36960135
#>  [647,]  4.04616771  5.4104401  5.71761568  6.00972871
#>  [648,]  5.12978547  5.7660195  6.01496271  6.19963637
#>  [649,]  3.80184960  4.3085423  4.37169045  4.40355962
#>  [650,]  1.98004499  2.0206846  2.05939970  2.06089261
#>  [651,]  0.06501469  0.0719623  0.07800526  0.08316023
#>  [652,]  1.70181277  2.1802867  2.28291597  2.29911243
#>  [653,]  3.14089921  3.8018742  3.98062013  4.02285265
#>  [654,]  3.88101336  4.1032680  4.24193412  4.27112520
#>  [655,]  4.25259562  4.3960351  4.50567013  4.53723342
#>  [656,]  3.83495541  4.4582420  5.02852501  5.37893849
#>  [657,] 10.09612513 10.2385222 10.32882783 10.34326136
#>  [658,]  9.39553003  9.4826050  9.49392329  9.49470474
#>  [659,]  4.00232700  5.2292169  6.00525169  6.31190430
#>  [660,]  0.19950245  0.2401876  0.27088744  0.29726339
#>  [661,]  8.23269230 10.0145474 11.38220590 12.58935050
#>  [662,]  5.87106745  8.1655359  8.23809116  8.29006440
#>  [663,]  1.53786867  1.8838905  1.98621082  2.07858982
#>  [664,]  3.73969398  4.5707167  4.91323537  4.96519973
#>  [665,]  5.12757592  5.8431357  6.12044606  6.30089293
#>  [666,]  1.84017809  2.0010815  2.13258294  2.18414960
#>  [667,]  2.99858401  3.3962422  3.66816469  3.74362160
#>  [668,]  2.68112959  2.9970996  3.28283163  3.55399869
#>  [669,]  8.42217869  9.1157825  9.30336015  9.45832221
#>  [670,]  3.51217469  3.8572658  4.18255028  4.44953808
#>  [671,]  3.59724909  4.5409918  4.91658393  5.19941429
#>  [672,]  2.86880303  3.4467669  3.49833545  3.51264254
#>  [673,]  2.92403224  4.0689538  5.16247038  5.76609602
#>  [674,]  2.53845181  3.0187251  3.43105695  3.51396017
#>  [675,]  4.44215142  4.9984400  5.29090409  5.51094755
#>  [676,]  2.91063216  3.3106008  3.35797388  3.39860135
#>  [677,]  6.12576839  8.4890093  9.85074812 10.64455980
#>  [678,]  4.98300393  5.4032546  5.63891334  5.70815779
#>  [679,]  6.10030972  8.1867323  9.36133943  9.92947752
#>  [680,]  5.32544844  7.3851999  8.78093515  8.87807549
#>  [681,]  8.98549705  9.5810092  9.70512656  9.81080830
#>  [682,]  3.09279007  3.5418265  3.71726725  3.79642607
#>  [683,]  7.57812441  9.0830837  9.72646658 10.10205275
#>  [684,]  0.70868801  0.9950033  1.21570829  1.32067646
#>  [685,]  6.20141423  7.8540180  7.85652471  7.85863066
#>  [686,]  0.68626581  0.7399052  0.74929079  0.75758254
#>  [687,]  8.35667933  8.7765246  8.98985931  9.18370547
#>  [688,]  2.85911847  3.7527280  4.25525497  4.57835181
#>  [689,]  7.36579104 10.5278816 11.87169915 11.90212904
#>  [690,]  4.49067415  5.3892605  5.54329376  5.56884252
#>  [691,]  4.15290200  4.7589646  4.83442068  4.89887960
#>  [692,]  3.48071459  4.1046718  4.40195717  4.56177980
#>  [693,]  6.63112922  8.3885331  9.43444284  9.77695408
#>  [694,] 10.00855013 10.9288166 11.63970314 11.65691890
#>  [695,]  2.39369039  3.0573272  3.49464852  3.80854514
#>  [696,]  2.83221933  2.9127874  2.94073178  2.94627145
#>  [697,]  4.64972432  4.8254338  4.88500642  4.92701367
#>  [698,]  6.89576986  8.6021331  8.95825547  9.09853460
#>  [699,]  1.89273736  1.9305907  1.94445892  1.95593872
#>  [700,]  3.77538281  3.9241137  4.01507125  4.09893642
#>  [701,]  6.03939017  6.3218056  6.34519051  6.36806687
#>  [702,]  1.11577481  1.4832366  1.69797717  1.78096581
#>  [703,]  5.35938058  5.7283965  5.83676679  5.84890070
#>  [704,]  4.32779810  4.5487866  4.60023872  4.61541725
#>  [705,]  3.84905130  4.9782475  5.08212050  5.16049929
#>  [706,]  2.67943004  3.5386419  4.00049690  4.44111531
#>  [707,]  3.47176646  3.8054168  4.06163507  4.08698434
#>  [708,]  0.45912024  0.6478200  0.76665892  0.78911666
#>  [709,]  2.89085788  3.5832909  3.73571243  3.83181998
#>  [710,]  6.27732662  7.6441105  8.07863148  8.39737465
#>  [711,]  4.08096168  5.1801113  6.21668885  6.88331532
#>  [712,]  2.28402050  3.0290728  3.15762537  3.23411286
#>  [713,]  5.64069846  6.5370529  6.97378828  6.98704594
#>  [714,]  6.41859982  7.0067187  7.20758463  7.33262665
#>  [715,]  2.56212681  3.1103692  3.56546533  3.57785588
#>  [716,]  5.34748436  7.7787908  9.63767200 10.81002723
#>  [717,]  5.27032818  6.1359759  6.73861431  6.81592489
#>  [718,]  4.80287828  4.9911005  5.09751525  5.18531615
#>  [719,]  2.12104204  2.8931663  3.38794123  3.43577840
#>  [720,]  1.12243771  1.1714878  1.21723036  1.21733988
#>  [721,]  4.32480501  5.7243288  6.44879441  6.64285045
#>  [722,] 10.83176113 11.6102087 11.62007799 11.62088268
#>  [723,]  3.38376860  3.9926934  4.29410054  4.33887749
#>  [724,]  6.10480820  8.8316491 10.39281201 11.26715677
#>  [725,]  7.06189967  8.5619425 10.00915391 10.38206818
#>  [726,]  1.88205668  2.5904236  2.81066175  2.90115822
#>  [727,]  3.41113991  4.3166490  4.95748461  5.37488600
#>  [728,]  3.85105241  4.9060417  4.99343461  5.07596223
#>  [729,]  3.87481757  4.0107387  4.11783289  4.22456600
#>  [730,]  1.81594581  2.3104075  2.55803553  2.66226900
#>  [731,]  2.03590086  2.7137185  3.23073500  3.31129984
#>  [732,]  7.97486338  8.9465565  9.77180029 10.02631360
#>  [733,]  3.80245306  5.3385953  5.74909041  5.95267053
#>  [734,]  4.91204208  5.5428935  5.82486925  5.86622094
#>  [735,] 13.45497366 14.7852053 15.61270998 16.28654164
#>  [736,]  3.38627279  3.6674185  3.83404364  3.93583104
#>  [737,]  7.58821471  9.7444641 11.84564387 12.05439375
#>  [738,]  7.20254797  7.5986485  7.75644682  7.89565930
#>  [739,]  1.00127976  1.1482980  1.24482278  1.29480468
#>  [740,]  2.87587208  3.8517962  4.13595666  4.27279677
#>  [741,]  2.63939257  3.8909734  4.56603453  5.20478342
#>  [742,]  2.01065459  2.2715115  2.48293773  2.52933981
#>  [743,]  7.00103393  7.2570862  7.28020750  7.29762030
#>  [744,]  4.37808109  5.1423229  5.39173165  5.54957112
#>  [745,] 10.30260159 11.1391452 11.78146172 12.31859818
#>  [746,]  4.71383709  6.7442871  7.15787704  7.31311934
#>  [747,]  7.14556975  8.5709309  9.06985229  9.17216043
#>  [748,]  2.11976644  2.3375422  2.45004730  2.48445524
#>  [749,]  5.97523130  6.3559056  6.72568800  6.99227293
#>  [750,]  6.53203863  8.0614490  8.55451384  8.66656896
#>  [751,]  6.08666444  6.9590689  7.78282106  8.51591263
#>  [752,]  4.06876033  5.5094007  6.61295972  7.18958605
#>  [753,]  1.73707475  2.3227685  2.87721590  3.00325888
#>  [754,]  4.44023590  4.8562932  4.99823474  5.02032045
#>  [755,] 11.53668870 13.5893097 14.52097736 15.01685527
#>  [756,]  9.55197974 10.2387505 10.71197415 10.85129723
#>  [757,]  1.25025911  1.4200425  1.50226805  1.51777431
#>  [758,]  8.19970158 10.7606900 10.80380762 10.83138270
#>  [759,]  3.75769725  4.2141736  4.36396042  4.42194904
#>  [760,]  5.13112581  6.7434592  6.79178822  6.82066338
#>  [761,]  6.79909847  9.6241966  9.73001897  9.81681463
#>  [762,]  8.77610966  9.5073114  9.92754083 10.32168172
#>  [763,]  3.69351757  4.9476944  5.40867532  5.78029019
#>  [764,]  5.38445518  6.1512838  6.20734662  6.23438019
#>  [765,]  2.57870530  3.1233938  3.49080212  3.77416311
#>  [766,]  6.65755333  8.6087081 10.20665627 10.62272846
#>  [767,]  3.73908070  4.3714604  4.98152074  4.99778773
#>  [768,]  2.95511475  3.1578680  3.30039487  3.43361879
#>  [769,]  2.74830902  3.3287996  3.82114686  4.12767826
#>  [770,]  7.20671944 10.3695137 10.80598012 11.21266633
#>  [771,]  3.89494536  5.0538964  5.72137680  5.95078385
#>  [772,]  1.53993857  1.7135810  1.73828908  1.74744211
#>  [773,]  4.94479949  5.6306794  5.82391099  5.87235297
#>  [774,]  2.06455245  2.6628054  3.25893785  3.52678284
#>  [775,]  4.24759349  4.4750480  4.66280894  4.80070599
#>  [776,]  6.06887282  6.1332802  6.19078098  6.19119326
#>  [777,]  6.06063904  7.6322534  8.60704122  9.11833339
#>  [778,]  4.78641679  6.8787657  7.03633557  7.07352194
#>  [779,]  3.30988209  3.6932809  3.81197958  3.90308926
#>  [780,]  1.73916506  2.0830896  2.19455214  2.24027209
#>  [781,] 10.89917014 12.5239494 14.03148426 14.36720174
#>  [782,]  2.44822094  2.7492802  3.04642854  3.08796984
#>  [783,]  4.82633216  5.0194061  5.15727362  5.19251620
#>  [784,]  4.99720592  5.4648206  5.52140997  5.57462916
#>  [785,]  2.36347100  2.8635597  3.18700913  3.30526224
#>  [786,]  3.26318479  4.0548314  4.64010810  4.90893637
#>  [787,]  7.36906582  7.5576405  7.70628609  7.84254852
#>  [788,]  1.29408862  1.3588904  1.39401874  1.40928068
#>  [789,]  1.30345366  1.4453859  1.50165609  1.51152399
#>  [790,]  5.67123158  7.1166390  7.42790902  7.52409399
#>  [791,]  2.83800311  3.1540473  3.32412056  3.42932174
#>  [792,]  1.51498235  1.5803049  1.63053795  1.64345414
#>  [793,]  8.66509587 10.0877858 10.70368984 11.11496845
#>  [794,]  5.63355590  6.3107695  6.50921880  6.66404476
#>  [795,]  4.62585149  5.1417341  5.60084347  5.61197403
#>  [796,]  3.04291389  4.3174706  4.93259090  5.38449070
#>  [797,]  4.96192811  6.1812950  6.73668467  6.84639951
#>  [798,]  1.10632896  1.3723880  1.58855006  1.60789064
#>  [799,]  6.28579581  6.5477211  6.65706221  6.76163753
#>  [800,]  4.64675905  5.0984748  5.46606949  5.51277012
#>  [801,]  3.45912480  3.9515696  4.32714969  4.55372713
#>  [802,]  1.57673777  1.9450563  2.13551338  2.28023165
#>  [803,]  8.68936060  9.5060465 10.22479933 10.36639094
#>  [804,]  3.07039991  3.3159306  3.34664044  3.36094418
#>  [805,]  2.90063563  3.4982903  4.07275287  4.38828621
#>  [806,]  4.51803799  5.3434501  5.66382750  5.74676394
#>  [807,]  3.83267047  4.4657811  4.53414830  4.54182828
#>  [808,]  8.25356382 11.4304751 11.54397219 11.57334515
#>  [809,]  9.97167625 10.2726466 10.53094794 10.56174329
#>  [810,]  5.03589415  6.2288693  6.35368117  6.38906691
#>  [811,]  2.80056823  2.9926233  3.07398022  3.13216786
#>  [812,]  2.07221439  2.8403465  3.44910415  3.60379967
#>  [813,]  3.98274810  4.8881120  5.37715877  5.72009744
#>  [814,]  9.13513188  9.7832127  9.86522254  9.88594778
#>  [815,]  6.63833117  7.2735342  7.63429754  7.93751824
#>  [816,]  3.98532546  4.7821012  5.16281062  5.40534738
#>  [817,]  6.21121475  7.3942700  7.66159639  7.80456756
#>  [818,]  4.38892461  4.6662057  4.88311446  4.96433467
#>  [819,]  1.84263133  2.6339020  2.82247832  2.88426063
#>  [820,]  2.41493436  2.4754049  2.51674851  2.53435113
#>  [821,]  5.37179010  5.6418463  5.70571582  5.74398467
#>  [822,]  0.70198008  0.9133126  1.00802454  1.07119865
#>  [823,]  8.56280509  9.0056834  9.19589476  9.22431496
#>  [824,]  5.29964973  6.9627764  7.28833228  7.36333221
#>  [825,]  3.90892107  5.6680891  6.69238385  6.75351683
#>  [826,]  3.74652167  3.9027364  3.97748148  4.00560494
#>  [827,]  8.51095278 10.0322549 10.48211590 10.69041249
#>  [828,]  4.56221399  5.1716116  5.74203959  6.06037541
#>  [829,]  4.58176205  4.6035328  4.61695519  4.62223798
#>  [830,]  3.78884741  5.3459368  5.88565825  6.36096974
#>  [831,]  3.73133664  4.3026443  4.44796447  4.53464375
#>  [832,]  3.31305444  3.8179975  3.99092078  4.07284277
#>  [833,]  2.14054460  2.5617096  2.76564584  2.89603705
#>  [834,]  5.79024158  7.7624653  8.80919379  9.57247666
#>  [835,]  3.08059897  3.8039109  4.17136746  4.51436736
#>  [836,]  3.54785824  4.5499107  5.25807997  5.45429289
#>  [837,]  5.59907627  7.9105960  9.16510269 10.31692384
#>  [838,]  1.94478875  2.3524720  2.66907005  2.71554800
#>  [839,]  3.27106421  3.8848740  4.40218497  4.41488224
#>  [840,]  7.40373425  7.6736849  7.88012010  8.06861962
#>  [841,]  4.22250712  4.6643625  4.96396274  4.97396587
#>  [842,]  7.46308158  8.8291793  9.95827766 10.55838269
#>  [843,]  6.93926977  8.8286691 10.50516526 11.61509513
#>  [844,]  3.05638209  3.4588761  3.59874181  3.59921983
#>  [845,]  3.37129191  3.6100439  3.61816888  3.62239216
#>  [846,]  1.69906892  2.4859529  2.98641533  3.04900456
#>  [847,]  2.18848701  2.9143507  3.51674943  4.03415379
#>  [848,]  3.43154507  4.0850655  4.56530412  4.64716195
#>  [849,]  3.72749567  4.5872186  5.38455104  5.72697467
#>  [850,]  5.12222511  5.2092182  5.26026904  5.26457741
#>  [851,]  3.11620814  3.9623142  4.31551423  4.32311441
#>  [852,]  2.89112726  3.7232028  4.18197686  4.27938513
#>  [853,]  3.38253655  3.7818062  3.87921951  3.90804576
#>  [854,]  8.75278832  8.8969249  8.92022868  8.92982838
#>  [855,] 13.27042184 14.9474779 15.69603383 15.97906258
#>  [856,]  6.14160703  6.7197982  6.87127482  6.89811766
#>  [857,]  2.31664479  3.3992671  3.77537327  4.03498154
#>  [858,]  7.76182742  8.1734841  8.42152924  8.52565114
#>  [859,]  4.53077211  4.9787453  5.23996883  5.43296082
#>  [860,]  3.63473409  3.9791409  4.03574354  4.04940998
#>  [861,]  1.18918037  1.3397637  1.39705183  1.43612222
#>  [862,]  2.89378594  3.6016262  4.20942246  4.43212900
#>  [863,]  4.75839238  5.5855582  5.63024141  5.66027822
#>  [864,]  5.42099722  6.6148858  6.81879911  6.93965856
#>  [865,]  2.79975044  3.5386354  3.80610916  3.90715680
#>  [866,]  2.74589358  2.8990395  3.03865488  3.12724930
#>  [867,]  8.29235924  9.3403288  9.83893764  9.92539670
#>  [868,] 12.13817085 14.0056815 14.89445441 15.49698897
#>  [869,]  0.41734027  0.4916312  0.55132886  0.60622465
#>  [870,]  5.83756696  6.5301534  6.89110937  7.06686026
#>  [871,]  4.09635627  4.2717954  4.39034446  4.50208722
#>  [872,]  2.37783029  2.9444238  3.48235734  3.57890284
#>  [873,]  7.85696134  9.0749087 10.18931575 10.24620216
#>  [874,]  3.78997790  4.3777883  4.51773796  4.56194912
#>  [875,]  6.48602320  8.2705022  8.86013908  9.34195865
#>  [876,]  0.60818290  0.6930244  0.76416418  0.81246900
#>  [877,]  3.72694564  4.2456464  4.51025734  4.55585920
#>  [878,]  2.17976887  2.8882165  3.55371240  3.89623139
#>  [879,]  3.31277785  4.0345961  4.66329672  4.79743612
#>  [880,]  5.69563592  5.8115087  5.88222237  5.89048722
#>  [881,]  2.25201841  2.8149513  3.15290556  3.45107899
#>  [882,]  6.51208760  7.4328354  7.68417553  7.71612839
#>  [883,]  1.27640543  1.7633701  2.18213534  2.47299271
#>  [884,]  4.15451985  4.5315427  4.80720993  4.98848105
#>  [885,]  1.97246164  2.4193657  2.71234467  2.86349153
#>  [886,]  2.67873986  2.8521792  2.97294102  3.02158386
#>  [887,]  5.01123532  5.2239016  5.41634857  5.47354552
#>  [888,]  4.79258718  5.4854322  5.73606876  5.75342762
#>  [889,]  0.85822007  1.1144417  1.27240684  1.31246381
#>  [890,]  1.47808753  2.0436769  2.58204415  2.82685604
#>  [891,]  2.34806267  3.0431651  3.42738877  3.57444787
#>  [892,]  7.91898591  8.6678138  8.69711782  8.72411337
#>  [893,]  3.31400794  3.7673619  3.87582070  3.92815959
#>  [894,]  2.61384928  3.1948350  3.67560109  3.86600957
#>  [895,]  3.90949863  5.1104645  6.23323356  6.80311742
#>  [896,]  4.94320119  5.7341898  6.40733622  6.98019160
#>  [897,]  4.79862667  4.8434116  4.86172291  4.86467688
#>  [898,]  1.15671273  1.3955845  1.42944507  1.44867589
#>  [899,]  8.26961475 10.5080187 12.61779029 13.39850589
#>  [900,]  4.58319820  5.3349080  5.35312413  5.35405478
#>  [901,] 11.91344357 13.4356819 14.06452004 14.17771723
#>  [902,]  3.55428514  5.1624670  6.25387928  7.29927712
#>  [903,]  3.64328123  4.0614000  4.44985393  4.83170996
#>  [904,] 14.09871631 15.2385455 16.29862918 16.90071308
#>  [905,]  3.77612103  5.3176466  6.43214701  7.11902988
#>  [906,]  4.97427878  5.4012756  5.70374346  5.93916317
#>  [907,] 11.29668359 13.0245859 13.85913499 14.67677193
#>  [908,]  0.84552654  0.9198233  0.98530721  0.99732222
#>  [909,]  3.04084884  3.3745688  3.63623298  3.73668381
#>  [910,]  3.89566282  4.0892561  4.28231090  4.36738696
#>  [911,] 14.64790974 16.0819030 16.52690895 16.80480555
#>  [912,]  0.71656008  1.0158809  1.27374810  1.46913397
#>  [913,]  3.98185506  4.0395665  4.06809292  4.08864051
#>  [914,]  3.39589282  3.9887496  4.33457395  4.50213635
#>  [915,]  5.35526768  6.4648473  6.79478918  6.97724077
#>  [916,]  2.60632905  3.2992388  3.36849472  3.38971697
#>  [917,]  1.19120133  1.6149005  1.99448513  2.27598568
#>  [918,]  8.10005579  9.6163448  9.64412386  9.64446852
#>  [919,]  9.34345988 12.0646001 14.72249998 15.66652039
#>  [920,]  1.61826749  1.8525395  1.99026282  2.03235808
#>  [921,]  5.88586925  7.3508302  8.51151778  9.28528747
#>  [922,]  4.16464468  4.5488473  4.74931656  4.75416464
#>  [923,]  5.95126503  8.3718826 10.46050973 10.47749623
#>  [924,]  7.45133595  7.8610152  8.09764885  8.12051295
#>  [925,]  5.14979576  6.6172694  6.82856233  6.93868865
#>  [926,]  1.18711585  1.4037763  1.52300169  1.53779864
#>  [927,]  1.35517142  1.5170104  1.67164068  1.73167438
#>  [928,]  8.73624542  9.2967455  9.81499983 10.18277063
#>  [929,]  1.31113005  1.3446953  1.35545868  1.36368112
#>  [930,]  4.68589491  5.6472336  6.04063559  6.13627072
#>  [931,]  1.03498439  1.2317502  1.42798204  1.51074035
#>  [932,]  3.36963663  3.9384509  4.39385198  4.79048962
#>  [933,]  6.28422176  7.9812273  8.78703444  9.45944099
#>  [934,]  2.77251246  3.4045638  3.94025810  4.11522619
#>  [935,]  3.11721329  3.5281654  3.86455645  4.11405674
#>  [936,]  2.00863935  2.2994875  2.50514159  2.68883275
#>  [937,]  0.88738321  1.0369332  1.09198495  1.11037628
#>  [938,]  4.08078399  5.5280762  6.67512947  6.74386108
#>  [939,]  3.83785517  4.1675204  4.34892525  4.44231310
#>  [940,]  2.68818045  3.0537775  3.15506451  3.20673002
#>  [941,]  1.53930203  2.2164129  2.58789013  2.67880633
#>  [942,]  1.60161359  1.9408403  2.06575528  2.07977836
#>  [943,]  1.69892539  1.9369167  2.17279210  2.38467820
#>  [944,]  3.67636375  4.9862211  5.86055528  6.11971145
#>  [945,]  4.82836462  5.3902063  5.47860403  5.54506733
#>  [946,]  6.01991129  7.6014051  8.96328927  9.55357028
#>  [947,]  4.51883896  5.5148717  6.32976249  6.92407360
#>  [948,]  3.79554706  4.5673419  5.24471717  5.70151161
#>  [949,]  4.76092789  6.4433799  8.08092774  8.45987582
#>  [950,]  3.37276340  3.9154765  4.03089113  4.07559442
#>  [951,]  9.09500513  9.9192911 10.42713833 10.57130687
#>  [952,]  7.78168401 10.5466941 11.74448166 11.75039850
#>  [953,]  4.15033990  4.6917315  5.01954961  5.04060133
#>  [954,]  3.01416267  4.1054832  5.16463128  6.08362230
#>  [955,]  5.60122211  6.6235110  7.52007169  8.23377421
#>  [956,]  2.70211977  3.4707313  3.67439047  3.87286696
#>  [957,]  1.37761379  1.9185469  2.24727193  2.30705931
#>  [958,]  3.49802357  4.5819372  4.71765144  4.85029845
#>  [959,] 12.02171408 15.7743085 18.60954145 18.82128258
#>  [960,]  2.63218708  3.1271731  3.42947605  3.66973338
#>  [961,]  6.45951439  7.9129882  8.09179276  8.20342797
#>  [962,]  1.28070054  1.7161703  2.04195343  2.09888967
#>  [963,]  2.73430707  3.3116120  3.88295294  4.43559606
#>  [964,]  1.02896978  1.2102157  1.35097197  1.43885673
#>  [965,]  6.86985094  8.5100532  8.69572535  8.80264331
#>  [966,]  3.53787942  4.3384958  4.94907128  5.12215866
#>  [967,]  2.71301437  3.2388398  3.74517148  3.88813413
#>  [968,] 10.28793544 11.5231955 12.67599217 13.78076182
#>  [969,]  3.44655213  4.4961023  4.70854327  4.73720615
#>  [970,]  3.26421882  3.6901069  3.99765936  4.01986888
#>  [971,]  1.67170693  1.9444666  1.98851791  2.02002763
#>  [972,]  6.15427903  7.5235770  8.62934864  9.00550506
#>  [973,]  1.83250589  2.2571553  2.65420510  3.00359935
#>  [974,]  2.95896750  3.2428294  3.37353250  3.48022743
#>  [975,]  4.63640064  4.8170194  4.98122008  5.06790497
#>  [976,]  8.29248715  8.7969873  9.20614437  9.47188529
#>  [977,]  1.62182767  1.8284984  1.96415083  2.00617174
#>  [978,]  1.38678001  1.7841117  2.15469173  2.26396783
#>  [979,]  4.81502365  4.9640267  5.00046085  5.00052273
#>  [980,]  3.50982148  4.5635732  5.33404950  5.78773127
#>  [981,]  3.03382480  4.4665777  4.69886080  4.87794659
#>  [982,]  4.67852708  5.5951241  6.14553934  6.41232767
#>  [983,]  8.69239271 12.1152801 15.05552035 16.31697099
#>  [984,]  4.84543959  5.7022086  6.12602092  6.32916600
#>  [985,]  6.10004855  7.0137880  7.74875533  7.78057366
#>  [986,]  3.02879840  3.3107258  3.38060929  3.43398971
#>  [987,]  6.03992128  6.6821643  7.12990781  7.40756156
#>  [988,]  4.45646825  5.2908422  5.88576566  6.14781599
#>  [989,]  9.49066595 10.5817471 11.16841147 11.69025986
#>  [990,]  9.91892696 10.5036417 11.06476029 11.33885967
#>  [991,]  2.74417114  3.3181074  3.64961656  3.95660436
#>  [992,]  2.08893301  2.5640302  2.95643029  3.11980753
#>  [993,]  3.72938606  3.8644702  3.96458576  3.99693060
#>  [994,]  8.57401898 11.5871373 13.63192343 13.65851149
#>  [995,]  1.57499497  2.0928970  2.17893406  2.23166424
#>  [996,]  1.74486696  2.4519517  3.04366492  3.51646439
#>  [997,]  6.72443069  7.7726786  8.34971391  8.88251149
#>  [998,]  2.35796716  3.0347826  3.50935055  3.65256004
#>  [999,]  4.39262758  5.1629804  5.49184720  5.69975677
#> [1000,]  4.58689329  5.2778733  5.69516227  5.86497664